### 2023-10-13 #5942 @editorialbot
[PRE REVIEW]:  A C++ template library for polynomials algebra over discrete euclidean domains 
**Submitting author:** @JeWaVe (<a href="http://orcid.org/0000-0002-6291-3345">regis portalez</a>)
**Repository:** https://github.com/aerobus-open-source/aerobus
**Branch with paper.md** (empty if default branch):
**Version:** v1.1
