### 2023-10-16 cbb5bfb
add chebyshev second kind + improve benchmarks
