+++ src/lib.h
@@ -16,7 +16,7 @@
 #define INLINED __forceinline 
 #else 
 #define ALIGNED(x) __attribute__((aligned(x)))
-#define INLINED __attribute__((always_inline))  
+#define INLINED __attribute__((always_inline)) inline
 #endif
 
 // aligned allocation
@@ -1947,7 +1947,7 @@ namespace aerobus {
 // known polynomials
 namespace aerobus {
 	namespace internal {
-		template<int deg>
+		template<int kind, int deg>
 		struct chebyshev_helper {
 			using type = typename pq64::template sub_t<
 				typename pq64::template mul_t<
@@ -1955,23 +1955,37 @@ namespace aerobus {
 						pq64::inject_constant_t<2>,
 						typename pq64::X
 					>,
-					typename chebyshev_helper<deg-1>::type
+					typename chebyshev_helper<kind, deg-1>::type
 				>,
-				typename chebyshev_helper<deg-2>::type
+				typename chebyshev_helper<kind, deg-2>::type
 			>;
 		};
 
 		template<>
-		struct chebyshev_helper<0> {
+		struct chebyshev_helper<1, 0> {
 			using type = typename pq64::one;
 		};
 
 		template<>
-		struct chebyshev_helper<1> {
+		struct chebyshev_helper<1, 1> {
 			using type = typename pq64::X;
+		};
+
+		template<>
+		struct chebyshev_helper<2, 0> {
+			using type = typename pq64::one;
+		};
+
+		template<>
+		struct chebyshev_helper<2, 1> {
+			using type = typename pq64::template mul_t<
+							typename pq64::inject_constant_t<2>, 
+							typename pq64::X>;
 		};
 	}
 
 	template<size_t deg>
-	using chebyshev = typename internal::chebyshev_helper<deg>::type;
+	using chebyshev_T = typename internal::chebyshev_helper<1, deg>::type;
+	template<size_t deg>
+	using chebyshev_U = typename internal::chebyshev_helper<2, deg>::type;
 }
