+++ src/benchmarks.cpp
@@ -4,14 +4,20 @@
 #include <cmath>
 #include "lib.h"
 #include "vml_math.h"
+#include <immintrin.h>
 
-const size_t N = 134217728 / 4;
+const size_t N = 134217728 / 16;
 
+#ifndef DEGREE 
+#define DEGREE 13
+#endif
+
+using AERO_SIN = aerobus::sin<aerobus::i64, DEGREE>;
 INLINED
 double sin_12(const double x) {
-	using V = aerobus::sin<aerobus::i64, 13>;
-	return V::eval(V::eval(V::eval(V::eval(V::eval(V::eval(
-		V::eval(V::eval(V::eval(V::eval(V::eval(V::eval(x))))))))))));
+
+	return AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(
+		AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(x))))))))))));
 }
 
 template<size_t N>
@@ -53,40 +59,52 @@ void vsin_slow(double * in, double * out) {
 
 template<size_t N>
 INLINED
-void vml_vsin_12(double * in, double * out) {
+void fast_math_vsin_12(double * in, double * out) {
 	static_assert(N % 32 == 0);
 	in = (double*) __builtin_assume_aligned(in, 64);
 	out = (double*) __builtin_assume_aligned(out, 64);
-	vml_vsin(in, out, (unsigned) N);
-	vml_vsin(out, out, (unsigned) N);
-	vml_vsin(out, out, (unsigned) N);
-	vml_vsin(out, out, (unsigned) N);
-	vml_vsin(out, out, (unsigned) N);
-	vml_vsin(out, out, (unsigned) N);
-	vml_vsin(out, out, (unsigned) N);
-	vml_vsin(out, out, (unsigned) N);
-	vml_vsin(out, out, (unsigned) N);
-	vml_vsin(out, out, (unsigned) N);
-	vml_vsin(out, out, (unsigned) N);
-	vml_vsin(out, out, (unsigned) N);
+	#pragma omp parallel for
+	for (unsigned int i = 0; i < N; i += 8)
+	{
+		_mm512_store_pd((void*)(out + i), 
+			_ZGVeN8v_sin(_ZGVeN8v_sin(_ZGVeN8v_sin(_ZGVeN8v_sin(
+			_ZGVeN8v_sin(_ZGVeN8v_sin(_ZGVeN8v_sin(_ZGVeN8v_sin(
+			_ZGVeN8v_sin(_ZGVeN8v_sin(_ZGVeN8v_sin(_ZGVeN8v_sin(
+				_mm512_load_pd((void*)(in + i)))))))))))))));
+	}
+}
+
+double rand(double min, double max)
+{
+  double range = (max - min);
+  double div = RAND_MAX / range;
+  return min + (rand() / div);
 }
 
 int main() {
+	printf("RUN WITH TAYLOR EXPANSION OF DEGREE %d\n", DEGREE);
+	printf("USED POLYNOMIAL AS SIN APPROX : %s\n", AERO_SIN::to_string().c_str());
 	double* in = static_cast<double*>(aerobus::aligned_malloc<double>(N, 64));
-	double* out = static_cast<double*>(aerobus::aligned_malloc<double>(N, 64));
+	double* out_aero = static_cast<double*>(aerobus::aligned_malloc<double>(N, 64));
+	double* out_std = static_cast<double*>(aerobus::aligned_malloc<double>(N, 64));
+	double* out_vml = static_cast<double*>(aerobus::aligned_malloc<double>(N, 64));
[…]

[…]
