### 2023-10-16 e0f2c3c
remove vml and replace by fastmath direct call
