### 2023-10-16 24b0cc3
better benchmarks + chebychev in paper
