+++ src/benchmarks.cpp
@@ -5,7 +5,7 @@
 #include "lib.h"
 #include <immintrin.h>
 
-const size_t N = 134217728 / 16;
+const size_t N = 134217728/2;
 
 
 extern "C" __m512d _ZGVeN8v_sin(__m512d x);
@@ -14,15 +14,15 @@ extern "C" __m512d _ZGVeN8v_sin(__m512d x);
 #define DEGREE 13
 #endif
 
-using AERO_SIN = aerobus::sin<aerobus::i64, DEGREE>;
+template<int deg>
 INLINED
 double sin_12(const double x) {
-
+	using AERO_SIN = aerobus::sin<aerobus::i64, deg>;
 	return AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(
 		AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(x))))))))))));
 }
 
-template<size_t N>
+template<size_t N, int deg>
 INLINED
 void vsin_12(double * in, double * out) {
 	static_assert(N % 32 == 0);
@@ -32,7 +32,7 @@ void vsin_12(double * in, double * out) {
 	for (int i = 0; i < N; i += 8) {
 		#pragma simd
 		for(int j = i; j < i+8; ++j) {
-			out[j] = sin_12(in[j]);
+			out[j] = sin_12<deg>(in[j]);
 		}
 	}
 }
@@ -82,10 +82,42 @@ double rand(double min, double max)
   double div = RAND_MAX / range;
   return min + (rand() / div);
 }
+template<int deg>
+void run_aero(double* in, double* out_aero) {
+	// warmup
+	vsin_12<N, deg>(in, out_aero);
+	vsin_12<N, deg>(in, out_aero);
+	vsin_12<N, deg>(in, out_aero);
+	double best = 1.0E9;
+	for (int i = 0; i < 10; ++i) {
+		auto start = std::chrono::steady_clock::now();
+		vsin_12<N, deg>(in, out_aero);
+		auto stop = std::chrono::steady_clock::now();
+		std::chrono::duration<double> time = stop - start;
+		if (time.count() < best) {
+			best = time.count();
+		}
+	}
+
+	printf("[aerobus deg %d] %.3e Gsin/s\n", deg, (12.0E-9 * (double) N) / best);
+}
+
+void verify(double* out_aero, double* out_std) {
+	double err_aero = 0.0;
+	double max_err = 0.0;
+	for(int i = 0; i < N; ++i) {
+		double err = fabs(out_aero[i] - out_std[i]);
+		err_aero += err;
+		if (err > max_err) {
+			max_err = err;
+		}
+	}
+	err_aero /= (double) N;
+	printf("average error (vs std) : %.2e\n", err_aero);
+	printf("max error (vs std) : %.2e\n", max_err);
+}
 
 int main() {
-	printf("RUN WITH TAYLOR EXPANSION OF DEGREE %d\n", DEGREE);
-	printf("USED POLYNOMIAL AS SIN APPROX : %s\n", AERO_SIN::to_string().c_str());
 	double* in = static_cast<double*>(aerobus::aligned_malloc<double>(N, 64));
 	double* out_aero = static_cast<double*>(aerobus::aligned_malloc<double>(N, 64));
 	double* out_std = static_cast<double*>(aerobus::aligned_malloc<double>(N, 64));
@@ -96,25 +128,6 @@ int main() {
 	#pragma omp parallel for
 	for(int i = 0; i < N; ++i) {
[…]
