### 2023-10-17 379b6a6
typos in paper.md
