+++ paper.bib
@@ -18,8 +18,6 @@
 	url="https://doi.org/10.1007/978-3-319-06486-4_7"
 }
 
-
-
 @misc{vml,
   author = {S. Kang, G. Shen and F. Du},
   title = {vml: Vectorized Math Library},
@@ -27,4 +25,10 @@
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/intel/VML}
+}
+
+@misc{OEIS,
+  title = {On-Line Encyclopedia of Integer Sequences},
+  howpublished = {\url{https://oeis.org/}},
+  note = {Founded: 1964}
 }
