### 2023-10-17 #5942 @JeWaVe
I pushed some modifications in both paper and bibligraphy.
