### 2023-10-23 #5942 @gkthiruvathukal
@drvinceknight Are you willing to help review this JOSS submission?
