### 2023-10-23 #5942 @gkthiruvathukal
@JeWaVe As we await the reviewers, I notice that this C++ project is a bit lacking when it comes to the build process. The Makefile doesn't appear to be portable.  You should consider using a Makefile generator (e.g. autoconf) or a more modern build system. I personally use CMake in my projects.
