### 2023-10-23 #5942 @JeWaVe
Ok. Im currently AFK but I'll find some time to fix that later this week.
Thank you

Envoyé depuis Proton Mail mobile

-------- Message d'origine --------
Le 23 oct. 2023 à 04:36, George K. Thiruvathukal a écrit :

> ***@***.***(https://github.com/JeWaVe) As we await the reviewers, I notice that this C++ project is a bit lacking when it comes to the build process. The Makefile doesn't appear to be portable. You should consider using a Makefile generator (e.g. autoconf) or a more modern build system. I personally use CMake in my projects.
>
