### 2023-10-29 82589a9
cmake and some documentation
