+++ src/lib.h
@@ -40,6 +40,7 @@ namespace aerobus {
 // concepts
 namespace aerobus
 {
+	/// Concept to express R is a Ring
 	template <typename R>
 	concept IsRing = requires {
 		typename R::one;
@@ -49,6 +50,7 @@ namespace aerobus
 		typename R::template mul_t<typename R::one, typename R::one>;
 	};
 	
+	/// Concept to express R is an euclidean domain
 	template <typename R>
 	concept IsEuclideanDomain = IsRing<R> && requires {
 		typename R::template div_t<typename R::one, typename R::one>;
@@ -59,6 +61,7 @@ namespace aerobus
 		R::is_euclidean_domain == true;
 	};
 
+	/// Concept to express R is a field
 	template<typename R>
 	concept IsField = IsEuclideanDomain<R> && requires {
 		R::is_field == true;
@@ -123,11 +126,8 @@ namespace aerobus {
 			(i >= 5 && i * i > n)>> : std::true_type {};
 	}
 
-	/**
-	 * check if integer is prime or not
-	 * 
-	 * @tparam n integer to check
-	*/
+	/// @brief checks if n is prime
+	/// @tparam n 
 	template<int32_t n>
 	struct is_prime {
 		/// @brief true iff n is prime
@@ -193,6 +193,8 @@ namespace aerobus {
 		};
 	}
 
+	/// @brief computes the greatest common divisor or A and B
+	/// @tparam T Ring type (must be euclidean domain)
 	template<typename T, typename A, typename B>
 	using gcd_t = typename internal::gcd<T>::template type<A, B>;
 }
@@ -200,9 +202,7 @@ namespace aerobus {
 // type_list
 namespace aerobus
 {
-	/**
-	 * Empty pure template struct to handle type lists
-	*/
+	/// @brief Empty pure template struct to handle type list
     template <typename... Ts>
     struct type_list;
 
@@ -332,35 +332,50 @@ namespace aerobus
 
 // i32
 namespace aerobus {
-	/**
-	 * 32 bits signed integers, seen as a algebraic ring with related operations
-	*/
+	///@brief 32 bits signed integers, seen as a algebraic ring with related operations
     struct i32 {
 		using inner_type = int32_t;
+		/// @brief values in i32
+		/// @tparam x an actual integer
 		template<int32_t x>
 		struct val {
 			static constexpr int32_t v = x;
 
+			/// @brief cast x into valueType
+			/// @tparam valueType double for example
 			template<typename valueType>
 			static constexpr valueType get() { return static_cast<valueType>(x); }
 
+			/// @brief is value zero 
 			using is_zero_t = std::bool_constant<x == 0>;
+
+			/// @brief string representation of value
 			static std::string to_string() {
 				return std::to_string(x);
 			}
 
+			/// @brief cast x into valueRing
+			/// @tparam valueRing double for example
 			template<typename valueRing>
 			static constexpr valueRing eval(const valueRing& v) {
 				return static_cast<valueRing>(x);
 			}
 		};
 
+		/// @brief constant zero
 		using zero = val<0>;
+		/// @brief constant one
 		using one = val<1>;
+		/// @brief integers are not a field
 		static constexpr bool is_field = false;
+		/// @brief integers are an euclidean domain
 		static constexpr bool is_euclidean_domain = true;
+		/// @brief inject a native constant
+		/// @tparam x 
+		/// @example i32::template inject_constant_2<2> -> i32::template val<2>
 		template<auto x>
 		using inject_constant_t = val<static_cast<int32_t>(x)>;
+
 		template<typename v>
 		using inject_ring_t = v;
 
@@ -411,33 +426,43 @@ namespace aerobus {
 		};
 
 	public:
+		/// @brief addition operator
[…]

[…]
