### 2023-10-29 #5942 @JeWaVe
Hi @gkthiruvathukal I just added a cmakelists.txt and updated the readme.
In addition, I added some documentation (not complete at all) and updated the readme (main branch)
