### 2024-01-09 ec38dc3
make cmake works on windows with VS2022 add UT for polynomial derivation (including finite field)
