+++ src/main.cpp
@@ -5,6 +5,7 @@
 #include <cstdlib>
 
 #include "lib.h"
+#include "../imports/conwaypolynomials.h"
 
 
 // conformance : https://godbolt.org/z/5chqjYEEs
@@ -212,15 +213,34 @@ int test_poly_derive() {
 			return 1;
 		}
 	} 
+	using Z2Z = zpz<2>;
 	{
 		// in Z/2Z
-		// 1 + x + 3x²
-		using P1 = polynomial<zpz<2>>::template val<zpz<2>::template val<3>, zpz<2>::template val<1>, zpz<2>::template val<1>>;
-		using PP = polynomial<zpz<2>>::template derive_t<P1>;
+		// 1 + x + 3x² -> 1
+		using P1 = polynomial<Z2Z>::template val<Z2Z::template val<3>, Z2Z::template val<1>, Z2Z::template val<1>>;
+		using PP = polynomial<Z2Z>::template derive_t<P1>;
 		if (PP::degree != 0) {
 			return 1;
 		}
 		if (PP::coeff_at_t<0>::v != 1) {
+			return 1;
+		}
+	}
+	{
+		// in Z/2Z
+		// x^3 + x^2 + x + 1 -> 1 + x^2 (because 3 == 1)
+		using P1 = polynomial<Z2Z>::template val<Z2Z::template val<1>, Z2Z::template val<1>, Z2Z::template val<1>, Z2Z::template val<1>>;
+		using PP = polynomial<Z2Z>::template derive_t<P1>;
+		if (PP::degree != 2) {
+			return 1;
+		}
+		if (PP::coeff_at_t<0>::v != 1) {
+			return 1;
+		}
+		if (PP::coeff_at_t<1>::v != 0) {
+			return 1;
+		}
+		if (PP::coeff_at_t<2>::v != 1) {
 			return 1;
 		}
 	}
@@ -589,6 +609,45 @@ int test_add_q32() {
 	return 0;
 }
 
+int test_is_zero_q32() {
+	using a = q32::zero;
+	if (!a::is_zero_t::value) {
+		return 1;
+	}
+	using b = q32::one;
+	if (b::is_zero_t::value) {
+		return 1;
+	}
+
+	return 0;
+}
+
+int test_gt_q32() {
+	{
+		using a = q32::inject_constant_t<1>;
+		using b = q32::zero;
+		if (!q32::gt_t<a, b>::value) {
+			return 1;
+		}
+	} 
+	{
+		using a = q32::zero;
+		using b = q32::inject_constant_t<2>;
+		if (q32::gt_t<a, b>::value) {
+			return 1;
+		}
+	}
+	{
+		using a = q32::zero;
+		using b = q32::zero;
+		if (q32::gt_t<a, b>::value) {
+			return 1;
+		}
+	}
+
+	return 0;
+}
+
 int test_sub_q32() {
 	{
 		using a = q32::val<i32::val<1>, i32::val<2>>;
@@ -770,6 +829,60 @@ int test_fraction_field_of_fraction_field () {
 	return 0;
 }
 
+int test_quotient_ring_is_z2z() {
+	using QQ = Quotient<i32, i32::template val<2>>;
+	using two = QQ::inject_constant_t<2>;
+	using three = QQ::inject_constant_t<3>;
+	using one = QQ::inject_constant_t<1>;
+	using zero = QQ::zero;
+	if (!QQ::eq_t<two, zero>::value) {
+		return 1;
+	}
+	if (!QQ::eq_t<one, three>::value) {
+		return 1;
+	}
+
+	return 0;
+}
+
+int test_instanciate_F4() {
+	using F2 = zpz<2>;
[…]
