+++ CMakeLists.txt
@@ -35,12 +35,12 @@ check_cxx_source_runs("
 set(CMAKE_CXX_STANDARD 20)
 
 if(MSVC) 
-  set(CMAKE_CXX_FLAGS "${OpenMP_CXX_FLAGS} /O2 /we4756 /D_USE_MATH_DEFINES /Wall /wd4711 /wd4710 /wd4127 /wd4577 /wd4100")
+  set(CMAKE_CXX_FLAGS "${OpenMP_CXX_FLAGS} /O2 /we4309 /we4756 /D_USE_MATH_DEFINES /Wall /wd4711 /wd4710 /wd4127 /wd4577 /wd4100")
 else()
   set(CMAKE_CXX_FLAGS "${OpenMP_CXX_FLAGS} -O3 -ftemplate-depth=30000 -ftemplate-backtrace-limit=0 -Werror=overflow -Wall")
 endif()
 
-add_executable(aerobus_test src/main.cpp src/lib.h)
+add_executable(aerobus_test src/main.cpp src/lib.h imports/conwaypolynomials.h)
 if(HAVE_AVX512_EXTENSIONS)
     message("AVX512 detected, adding benchmarks")
     if(MSVC)
@@ -48,7 +48,7 @@ if(HAVE_AVX512_EXTENSIONS)
     else()
       set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -mavx512f -mavx512vl")
     endif()
-    add_executable(aerobus_benchmarks src/benchmarks.cpp src/lib.h)
+    add_executable(aerobus_benchmarks src/benchmarks.cpp src/lib.h imports/conwaypolynomials.)
 else()
     message("AVX512 not detected, benchmarks need manual configuration")
 endif()
