### 2024-01-10 #5942 @JeWaVe
Hi everyone,

I just added a major update with implementation of galois fields (GF(p, n) for all p < 1000 and n < 20).
Everything is in the main branch, I hope I didn't break anything here.
