### 2024-01-11 d4a33a3
add bernouilli_v and others update goldbolt compiler conformance link
