### 2024-01-16 #6233 @editorialbot
[REVIEW]:  A C++ template library for polynomials algebra over discrete euclidean domains 
**Submitting author:** @JeWaVe (<a href="http://orcid.org/0000-0002-6291-3345">regis portalez</a>)
**Repository:** https://github.com/aerobus-open-source/aerobus
**Branch with paper.md** (empty if default branch):
**Version:** 1.5.1
**Archive:** 10.6084/m9.figshare.28143233.v1
