### 2024-01-18 #6233 @pitsianis
@JeWaVe, please make an extra effort to fix orthographic and syntactic mistakes in your paper.
There is no excuse now that ChatGPT, Grammarly, and other tools are widely available.
Please take a look at line 63.
> If 𝔸 is euclidean, we can build it’s field of fractions: the smallest field containg 𝔸.

Please
 * capitalize proper names like Euclid, Taylor
 * do not use contractions like "it's"; it is too informal (for publications), and in this instance, it is also wrong
 * use a spell checker and grammar helper to catch typos like "containg"
 * consistently use mathematical expressions between dollar signs and program items between backquotes.
 * use correct language annotations to highlight the assembly code blocks

Also, the list of citations needs to be more significant for such a rich field.
