+++ src/lib.h
@@ -107,7 +107,7 @@ namespace aerobus {
 		template<template<typename...> typename TT, typename T>
 		inline constexpr bool is_instantiation_of_v = is_instantiation_of<TT, T>::value;
 
-		template <int64_t i, typename T, typename... Ts>
+		template <size_t i, typename T, typename... Ts>
 		struct type_at
 		{
 			static_assert(i < sizeof...(Ts) + 1, "index out of range");
@@ -120,6 +120,17 @@ namespace aerobus {
 
 		template <size_t i, typename... Ts>
 		using type_at_t = typename type_at<i, Ts...>::type;
+
+		template<size_t i, auto x, auto... xs>
+		struct value_at {
+			static_assert(i < sizeof...(xs) + 1, "index out of range");
+			static constexpr auto value = value_at<i-1, xs...>::value;
+		};
+
+		template<auto x, auto... xs> 
+		struct value_at<0, x, xs...> {
+			static constexpr auto value = x;
+		};
 
 
 		template<int32_t n, int32_t i, typename E = void>
@@ -1358,6 +1369,161 @@ namespace aerobus {
 		/// @example polynomial<i32>::template inject_ring_t<i32::template val<2>>
 		template<typename v>
 		using inject_ring_t = val<v>;
+	};
+}
+
+// big integers
+namespace aerobus {
+	struct bigint {
+		enum sign {
+			positive,
+			negative
+		};
+
+		template<sign s, uint32_t an, uint32_t... as>
+		struct val {
+			template<size_t index, typename E = void>
+			struct digit_at {};
+
+			template<size_t index>
+			struct digit_at<index, std::enable_if_t<(index <= sizeof...(as))>> {
+				static constexpr uint32_t value = internal::value_at<(sizeof...(as) - index), an, as...>::value;
+			};
+			
+			template<size_t index>
+			struct digit_at<index, std::enable_if_t<(index > sizeof...(as))>> {
+				static constexpr uint32_t value = 0;
+			};
+
+			static constexpr bool is_positive = s != sign::negative;
+
+			using strip = val<s, as...>;
+			static constexpr uint32_t aN = an;
+			static constexpr size_t digits = sizeof...(as) + 1;
+
+			static std::string to_string() {
+				return std::to_string(aN) + "B^" + std::to_string(digits-1) + " + " + strip::to_string();
+			}
+		};
+
+		template<typename I>
+		struct is_zero {
+			static constexpr bool value = I::digits == 1 && I::aN == 0;
+		};
+
+		template<typename I>
+		static constexpr bool is_zero_v = is_zero<I>::value;
+
+		template<sign s, uint32_t a0>
+		struct val<s, a0> {
+			using strip = val<s, a0>;
+			static constexpr bool is_positive = s != sign::negative;
+			static constexpr uint32_t aN = a0;
+			static constexpr size_t digits = 1;
+			template<size_t index, typename E = void>
+			struct digit_at {};
+			template<size_t index>
+			struct digit_at<index, std::enable_if_t<index == 0>> {
+				static constexpr uint32_t value = a0;
+			};
+			
+			template<size_t index>
+			struct digit_at<index, std::enable_if_t<index != 0>> {
+				static constexpr uint32_t value = 0;
+			};
+
+			static std::string to_string() {
+				return std::to_string(a0);
+			}
+		};
+
+		using zero = val<sign::positive, 0>;
+		using one = val<sign::positive, 1>;
+
+	private:
+		template<uint32_t x, uint32_t y, uint8_t carry_in = 0>
+		struct add_digit_helper {
+		private:
+			static constexpr uint64_t raw = ((uint64_t) x + (uint64_t) y + (uint64_t) carry_in);
+		public:
[…]
