+++ src/Makefile
@@ -6,6 +6,11 @@ build: # to be sure both compile properly
 	clang++ -I. -ftemplate-depth=30000 -std=c++20 -ftemplate-backtrace-limit=0 -O3 main.cpp -o main.exe 
 	g++ -I. -fconcepts-diagnostics-depth=3 -ftemplate-depth=30000 -std=c++20 -ftemplate-backtrace-limit=0 -O3 main.cpp -o main.exe 
 
+debug:
+	g++ -I. -std=c++20 -g -ggdb -O0 main.cpp -o main.exe
+
+test: clean build run
+
 run: 
 	./main.exe
 
@@ -18,5 +23,5 @@ doc:
 	cd ../documentation/latex && make
 
 
-all: clean build run
+all: clean build run benchmarks doc
 
