### 2024-01-18 ff5dff5
fix typos and grammar add paragraph for quotient rings and galois/conway add biblio
