+++ paper.bib
@@ -32,3 +32,28 @@
   howpublished = {\url{https://oeis.org/}},
   note = {Founded: 1964}
 }
+
+@book{lang2012algebra,
+  title={Algebra},
+  author={Lang, Serge},
+  volume={211},
+  year={2012},
+  publisher={Springer Science \& Business Media}
+}
+
+@book{bourbaki2013algebra,
+  title={Algebra II: Chapters 4-7},
+  author={Bourbaki, Nicolas},
+  year={2013},
+  publisher={Springer Science \& Business Media}
+}
+
+@article{evariste1846memoire,
+  title={M{\'e}moire sur les conditions de r{\'e}solubilit{\'e} des {\'e}quations par radicaux},
+  author={{\'E}variste, Galois},
+  journal={Journal de math{\'e}matiques pures et appliqu{\'e}es, Ser},
+  volume={1},
+  number={111846},
+  pages={417--433},
+  year={1846}
+}
