+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Aerobus: a C++ template library for polynomials algebra over discrete euclidean domains'
+title: 'Aerobus: a C++ template library for polynomials algebra over discrete Euclidean domains'
 tags:
   - Polynomials
   - Mathematics
@@ -23,42 +23,44 @@ bibliography: paper.bib
 C++ comes with high compile-time computations capability, also known as metaprogramming with templates.
 Templates are a language-in-the-language which is Turing-complete, meaning we can run every computation at compile time instead of runtime, as long as input data is known at compile time. 
 
-Using these capabilities, vastly extended with latest versions of the standard, we implemented a library 
-for discrete euclidean domain, such as $\mathbb{Z}$. We also provide a way to generate the fraction field of such rings (e.g. $\mathbb{Q}$). 
+Using these capabilities, vastly extended with the latest versions of the standard, we implemented a library for discrete Euclidean domains, such as $\mathbb{Z}$. We also provide a way to generate the fraction field of such rings (e.g. $\mathbb{Q}$). 
 
-We also implemented polynomials over such discrete ring and fields (e.g. $\mathbb{Q}[X]$). Since polynomials are also a ring, the above implementation gives us rational fractions as field of fractions of polynomials.
+We also implemented polynomials over such discrete rings and fields (e.g. $\mathbb{Q}[X]$). Since polynomials are also a ring, the above implementation gives us rational fractions as the field of fractions of polynomials.
 
-In addition, we expose a way to generate taylor series of any math functions as long as coefficients are known. 
-Likewise, given we have a typed representation of rationals, we've been able to add continued fractions as part of the library. 
+In addition, we expose a way to generate the Taylor series of any math functions as long as coefficients are known. 
 
-`Aerobus` was designed to be used in high performance software, teaching purposes or embedded sotfware where as much as possible must be precomputed to shrink binary size. It compiles with major compilers : gcc, clang and msvc. It is quite easily configurable and extensible. 
+In addition, we added some useful additional features, such as known polynomials (Chebyshev), continued fractions, quotient rings and some Conway polynomials to define Galois finite fields.
+
+`Aerobus` was designed to be used in high-performance software, teaching purposes or embedded software where as much as possible must be precomputed to shrink binary size. It compiles with major compilers: gcc, clang and msvc. It is quite easily configurable and extensible. 
 
 # Statement of need
 By implementing general algebra concepts such as discrete rings, field of fractions and polynomials, `Aerobus` can serve multiple purposes. 
 
-The main application we want to express in this paper is the automatic (and configurable) generation or taylor approximation of usual transcendental functions such as `exp` or `sin`. The "generated" code is pure C++ and can be inspected. 
+The main application we want to express in this paper is the automatic (and configurable) generation or Taylor approximation of usual transcendental functions such as `exp` or `sin`. The "generated" code is pure C++ and can be inspected.
 
-This functions are usually exposed by the standard library (`<cmath>`) with high (guaranteed) precision. However, in high performance computing, when not compiled with `-Ofast`, evaluating `std::exp` has several flaws : 
+These functions are usually exposed by the standard library (`<cmath>`) with high (guaranteed) precision. However, in high-performance computing, when not compiled with `-Ofast`, evaluating `std::exp` has several flaws: 
 
-- it leads to a syscal which is very expensive
[…]
