+++ paper.bib
@@ -57,3 +57,10 @@
   pages={417--433},
   year={1846}
 }
+
+@book{holt2005handbook,
+  title={Handbook of computational group theory},
+  author={Holt, Derek F and Eick, Bettina and O'Brien, Eamonn A},
+  year={2005},
+  publisher={CRC Press}
+}
