### 2024-01-18 #6233 @danielskatz
A command needs to be the first thing in a new comment
