+++ paper.bib
@@ -64,3 +64,11 @@
   year={2005},
   publisher={CRC Press}
 }
+
+@book{heath1956thirteen,
+  title={The thirteen books of Euclid's Elements},
+  author={Heath, Thomas Little and others},
+  year={1956},
+  publisher={Courier Corporation}
+}
+
