### 2024-01-18 bdbd46d
add horner explanation and references
