+++ paper.md
@@ -45,13 +45,13 @@ These functions are usually exposed by the standard library (`<cmath>`) with hig
 
 Hardware vendors provide high-performance libraries such as [@wang2014intel], but implementation is often hidden and not extensible. 
 
-Some others can provide vectorized functions, such as [@vml] does. But libraries like VML are highly tight to one architecture by their use of intrinsics or inline assembly. In addition, they only provide a restricted list of math functions and do not expose capabilities to generate high-performance versions of other functions such as arctanh. It is the same for the standard library compiled with `-Ofast`: it generates a vectorized version of some functions (such as exp) but with no control of precision and no extensibility. 
+Some others can provide vectorized functions, such as @vml does. But libraries like VML are highly tight to one architecture by their use of intrinsics or inline assembly. In addition, they only provide a restricted list of math functions and do not expose capabilities to generate high-performance versions of other functions such as arctanh. It is the same for the standard library compiled with `-Ofast`: it generates a vectorized version of some functions (such as exp) but with no control of precision and no extensibility. 
 
 `Aerobus` provides automatic generation of such functions, in a hardware-independent way.
 In addition, `Aerobus` provides a way to control the precision of the generated function by changing the degree of Taylor expansion, which can't be used in competing libraries without reimplementing the whole function. 
 
 # Mathematic definitions
-For the sake of completeness, we give basic definitions of the mathematical concepts which the library deals with. However, readers desiring complete and rigorous definitions of the concepts explained below should refer to some mathematical books on algebra, such as [@lang2012algebra] or [@bourbaki2013algebra].
+For the sake of completeness, we give basic definitions of the mathematical concepts which the library deals with. However, readers desiring complete and rigorous definitions of the concepts explained below should refer to some mathematical books on algebra, such as @lang2012algebra or @bourbaki2013algebra.
 
 A `ring` $\mathbb{A}$ is a nonempty set with two internal laws, addition and multiplication. There is a neutral element for both, zero and one. 
 Addition is commutative and associative and every element $x$ has an inverse $-x$. Multiplication is commutative, associative and distributive over addition, meaning that $a(b+c) = ab+ac$ for every $a, b, c$ element. We call it `discrete` if it is countable. 
@@ -69,7 +69,7 @@ $$a_0 + a_1X + \ldots + a_nX^n$$
 
 where $a_n \neq 0$ if $n \neq 0$. 
 
-$(a_i)$, the coefficients, are elements of $\mathbb{A}$. The theory states that if $\mathbb{A}$ is a field, then $\mathbb{A}X$ is Euclidean. That means notions like division of greatest common divisor (gcd) have a meaning, yielding an arithmetic of polynomials. 
+$(a_i)$, the coefficients, are elements of $\mathbb{A}$. The theory states that if $\mathbb{A}$ is a field, then $\mathbb{A}[X]$ is Euclidean. That means notions like division of greatest common divisor (gcd) have a meaning, yielding an arithmetic of polynomials. 
 
 
 ## Field of fractions
@@ -81,7 +81,7 @@ Given polynomials over a field form an Euclidean ring, we can do the same constr
 ## Quotient rings
[…]
