+++ paper.bib
@@ -72,3 +72,22 @@
   publisher={Courier Corporation}
 }
 
+@inproceedings{horner1815new,
+  title={A New Method of Solving Numerical Equations of all Orders, by Continuous Approximation.},
+  author={Horner, WG},
+  booktitle={Abstracts of the Papers Printed in the Philosophical Transactions of the Royal Society of London},
+  volume={2},
+  pages={117--117},
+  year={1815},
+  organization={JSTOR}
+}
+
+@book{knuth2014art,
+  title={Art of computer programming, volume 2: Seminumerical algorithms},
+  author={Knuth, Donald E},
+  year={2014},
+  publisher={Addison-Wesley Professional}
+}
+
+
+
