### 2024-01-18 #6233 @JeWaVe
add missing references

freezing development - waiting for review
