+++ src/lib.h
@@ -71,6 +71,8 @@ namespace aerobus
 		typename R::template add_t<typename R::one, typename R::one>;
 		typename R::template sub_t<typename R::one, typename R::one>;
 		typename R::template mul_t<typename R::one, typename R::one>;
+		typename R::template minus_t<typename R::one>;
+		R::template eq_v<typename R::one, typename R::one> == true;
 	};
 	
 	/// Concept to express R is an euclidean domain
@@ -79,11 +81,9 @@ namespace aerobus
 		typename R::template div_t<typename R::one, typename R::one>;
 		typename R::template mod_t<typename R::one, typename R::one>;
 		typename R::template gcd_t<typename R::one, typename R::one>;
-		typename R::template eq_t<typename R::one, typename R::one>;
-		typename R::template pos_t<typename R::one>;
 
 		R::template pos_v<typename R::one> == true;
-		//typename R::template gt_t<typename R::one, typename R::zero>;
+		R::template gt_v<typename R::one, typename R::zero> == true;
 		R::is_euclidean_domain == true;
 	};
 
@@ -198,8 +198,7 @@ namespace aerobus {
 			// B = 0, A > 0
 			template<typename A, typename B>
 			struct gcd_helper<A, B, std::enable_if_t<
-				((B::is_zero_t::value) &&
-					(Ring::template gt_t<A, typename Ring::zero>::value))>>
+				B::is_zero_v && Ring::template pos_v<A>>>
 			{
 				using type = A;
 			};
@@ -207,16 +206,15 @@ namespace aerobus {
 			// B = 0, A < 0
 			template<typename A, typename B>
 			struct gcd_helper<A, B, std::enable_if_t<
-				((B::is_zero_t::value) &&
-					!(Ring::template gt_t<A, typename Ring::zero>::value))>>
+				B::is_zero_v && !Ring::template pos_v<A>>>
 			{
-				using type = typename Ring::template sub_t<typename Ring::zero, A>;
+				using type = typename Ring::template minus_t<A>;
 			};
 
 			// B != 0
 			template<typename A, typename B>
 			struct gcd_helper<A, B, std::enable_if_t<
-				(!B::is_zero_t::value)
+				(!B::is_zero_v)
 				>> {
 			private:
 				// A / B
@@ -251,7 +249,7 @@ namespace aerobus {
 			using type = std::conditional_t<
 				Ring::template pos_v<tmp>,
 				tmp, 
-				typename Ring::template sub_t<typename Ring::zero, tmp>
+				typename Ring::template minus_t<tmp>
 			>;
 		};
 
@@ -266,13 +264,12 @@ namespace aerobus {
 		using div_t = val<typename Ring::template div_t<typename v1::type, typename v2::type>>;
 		template<typename v1, typename v2>
 		using mod_t = val<typename Ring::template mod_t<typename v1::type, typename v2::type>>;
+
 		template<typename v1, typename v2>
-		using eq_t = typename Ring::template eq_t<typename v1::type, typename v2::type>;
-		template<typename v1>
-		using pos_t = std::true_type;
+		static constexpr bool eq_v = Ring::template eq_v<typename v1::type, typename v2::type>;
 
 		template<typename v>
-		static constexpr bool pos_v = pos_t<v>::value;
+		static constexpr bool pos_v = true;
 
 		static constexpr bool is_euclidean_domain = true;
 
@@ -432,7 +429,7 @@ namespace aerobus {
 			static constexpr valueType get() { return static_cast<valueType>(x); }
 
 			/// @brief is value zero 
[…]
