+++ src/main.cpp
@@ -302,35 +302,35 @@ int test_poly_eq() {
 	{
 		using A = polynomial<i32>::val<i32::val<1>>;
 		using B = polynomial<i32>::val<i32::val<1>>;
-		if (!polynomial<i32>::eq_t<A, B>::value) {
+		if (!polynomial<i32>::eq_v<A, B>) {
 			return 1;
 		}
 	}
 	{
 		using A = polynomial<i32>::val<i32::val<1>>;
 		using B = polynomial<i32>::val<i32::val<2>>;
-		if (polynomial<i32>::eq_t<A, B>::value) {
+		if (polynomial<i32>::eq_v<A, B>) {
 			return 1;
 		}
 	}
 	{
 		using A = polynomial<i32>::val<i32::val<1>, i32::val<1>>;
 		using B = polynomial<i32>::val<i32::val<1>, i32::val<1>>;
-		if (!polynomial<i32>::eq_t<A, B>::value) {
+		if (!polynomial<i32>::eq_v<A, B>) {
 			return 1;
 		}
 	}
 	{
 		using A = polynomial<i32>::val<i32::val<1>, i32::val<1>>;
 		using B = polynomial<i32>::val<i32::val<1>, i32::val<2>>;
-		if (polynomial<i32>::eq_t<A, B>::value) {
+		if (polynomial<i32>::eq_v<A, B>) {
 			return 1;
 		}
 	}
 	{
 		using A = polynomial<i32>::val<i32::val<1>, i32::val<1>>;
 		using B = polynomial<i32>::val<i32::val<1>, i32::val<1>, i32::val<2>>;
-		if (polynomial<i32>::eq_t<A, B>::value) {
+		if (polynomial<i32>::eq_v<A, B>) {
 			return 1;
 		}
 	}
@@ -419,7 +419,7 @@ int test_poly_div() {
 		// x + 1
 		using Q = polynomial<i32>::div_t<A, B>;
 		using R = polynomial<i32>::mod_t<A, B>;
-		if (!R::is_zero_t::value) {
+		if (!R::is_zero_v) {
 			return 1;
 		}
 		if (Q::degree != 1) {
@@ -438,7 +438,7 @@ int test_poly_div() {
 		using B = polynomial<i32>::val<i32::val<2>>;
 		using C = polynomial<i32>::div_t<A, B>;
 		using R = polynomial<i32>::mod_t<A, B>;
-		if (!R::is_zero_t::value) {
+		if (!R::is_zero_v) {
 			return 1;
 		}
 		if (C::degree != 2) {
@@ -614,11 +614,11 @@ int test_add_q32() {
 
 int test_is_zero_q32() {
 	using a = q32::zero;
-	if (!a::is_zero_t::value) {
+	if (!a::is_zero_v) {
 		return 1;
 	}
 	using b = q32::one;
-	if (b::is_zero_t::value) {
+	if (b::is_zero_v) {
 		return 1;
 	}
 
@@ -629,21 +629,21 @@ int test_gt_q32() {
 	{
 		using a = q32::inject_constant_t<1>;
 		using b = q32::zero;
-		if (!q32::gt_t<a, b>::value) {
+		if (!q32::gt_v<a, b>) {
 			return 1;
 		}
 	} 
 	{
 		using a = q32::zero;
 		using b = q32::inject_constant_t<2>;
-		if (q32::gt_t<a, b>::value) {
+		if (q32::gt_v<a, b>) {
 			return 1;
 		}
 	}
 	{
 		using a = q32::zero;
 		using b = q32::zero;
-		if (q32::gt_t<a, b>::value) {
+		if (q32::gt_v<a, b>) {
 			return 1;
[…]
