+++ src/Makefile
@@ -1,6 +1,5 @@
 clean:
 	rm -f *.exe
-	rm -rf ../documentation
 
 build: # to be sure both compile properly
 	clang++ -I. -ftemplate-depth=30000 -std=c++20 -ftemplate-backtrace-limit=0 -O3 main.cpp -o main.exe 
@@ -19,6 +18,7 @@ benchmarks: clean
 	./benchmarks_avx512.exe
 
 doc:
+	rm -rf ../documentation
 	cd .. && doxygen Doxyfile
 	cd ../documentation/latex && make
 
