### 2024-01-23 ffda390
add bigint shift right (necessary for division)
