### 2024-01-23 9b0be69
add division for bigint (not tested - real bad performance)
