+++ src/lib.h
@@ -1369,6 +1369,23 @@ namespace aerobus {
 
 	private:
 
+		template<signs s>
+		struct opposite {
+			static constexpr signs value = s == signs::positive ? signs::negative : signs::positive;
+		};
+
+		template<signs s>
+		static constexpr signs opposite_v = opposite<s>::value;
+
+		template<signs s1, signs s2>
+		static constexpr signs mul_sign() {
+			if constexpr (s1 == signs::positive) {
+				return s2;
+			}
+
+			return opposite_v<s2>;
+		}
+
 		template<size_t ss, signs s, uint32_t aN, uint32_t... as>
 		struct shift_left_helper {
 			using type = typename shift_left_helper<ss-1, s, aN, as..., 0>::type;
@@ -1379,20 +1396,6 @@ namespace aerobus {
 		{
 			using type = val<s, aN, as...>;
 		};
-
-		template<signs s>
-		static constexpr signs opposite() {
-			return s == signs::positive ? signs::negative : signs::positive;
-		}
-
-		template<signs s1, signs s2>
-		static constexpr signs mul_sign() {
-			if constexpr (s1 == signs::positive) {
-				return s2;
-			}
-
-			return opposite<s2>();
-		}
 
 	public:
 
@@ -1425,7 +1428,7 @@ namespace aerobus {
 
 			static constexpr bool is_zero_v = sizeof...(as) == 0 && an == 0;
 
-			using minus_t = val<opposite<s>(), an, as...>;
+			using minus_t = val<opposite_v<s>, an, as...>;
 		};
 
 		template<signs s, uint32_t a0>
@@ -1453,7 +1456,7 @@ namespace aerobus {
 
 			static constexpr bool is_zero_v = a0 == 0;
 
-			using minus_t = val<opposite<s>(), a0>;
+			using minus_t = val<opposite_v<s>, a0>;
 
 		};
 
@@ -1643,6 +1646,24 @@ namespace aerobus {
 		template<typename I1, typename I2, std::size_t... I>
 		struct mul_helper<I1, I2, std::index_sequence<I...>> {
 			using type = typename vadd<typename mul_row<I1, I2, I>::type...>::type;
+		};
+
+		template<typename I, size_t index>
+		struct div_2_digit {
+			static constexpr uint32_t value = ((I::template digit_at<index + 1>::value & 1) << 31) + (I::template digit_at<index>::value >> 1);
+		};
+
+		template<typename X, typename I>
+		struct div_2_helper {};
+
+		template<typename X, std::size_t... I>
+		struct div_2_helper<X, std::index_sequence<I...>> {
+			using type = val<signs::positive, div_2_digit<X, I>::value...>;
+		};
+
+		template<typename X>
+		struct div_2 {
+			using type = typename simplify<typename div_2_helper<X, internal::make_index_sequence_reverse<X::digits>>::type>::type;
 		};
 
 		template<typename I1, typename I2, typename E = void>
@@ -1992,9 +2013,99 @@ namespace aerobus {
 			using tmp = typename shift_right_helper<I, s + 1>::type;
 		public:
 			using type = typename add<
-				typename digit,
+				digit,
 				typename tmp::template shift_left<1>
 			>::type;
+		};
+
+		template<typename A, typename B, typename E = void>
+		struct floor_helper {};
+
+		template<typename A, typename B>
+		struct floor_helper<A, B, std::enable_if_t<gt_helper<B, A>::value>> {
+			using type = zero;
+		};
+
+		template<typename A, typename B>
+		struct floor_helper<A, B, std::enable_if_t<eq<A, B>::value>> {
+			using type = one;
+		}; 
+		
+		template<typename A, typename B>
[…]
