### 2024-01-23 bd9af06
add missing typename (CI detected)
