### 2024-01-24 19a2151
fix compilation issue (explicit template specialization out of namespace scope)
