### 2024-01-24 c2dd751
make bigint division OK TODO : add more test TODO 2 : speedup compilation TODO3 : gcd and ring for bigint
