+++ src/lib.h
@@ -1377,6 +1377,16 @@ namespace aerobus {
 		template<signs s>
 		static constexpr signs opposite_v = opposite<s>::value;
 
+		static std::string to_string(const signs& s) {
+			switch (s) {
+				case signs::negative:
+					return "-";
+				case signs::positive:
+				default:
+					return "+";
+			}
+		}
+
 		template<signs s1, signs s2>
 		static constexpr signs mul_sign() {
 			if constexpr (s1 == signs::positive) {
@@ -1423,7 +1433,7 @@ namespace aerobus {
 			static constexpr size_t digits = sizeof...(as) + 1;
 
 			static std::string to_string() {
-				return std::to_string(aN) + "B^" + std::to_string(digits-1) + " + " + strip::to_string();
+				return bigint::to_string(s) + std::to_string(aN) + "B^" + std::to_string(digits-1) + " + " + strip::to_string();
 			}
 
 			static constexpr bool is_zero_v = sizeof...(as) == 0 && an == 0;
@@ -1451,7 +1461,7 @@ namespace aerobus {
 			};
 
 			static std::string to_string() {
-				return std::to_string(a0);
+				return bigint::to_string(s) + std::to_string(a0);
 			}
 
 			static constexpr bool is_zero_v = a0 == 0;
@@ -2018,7 +2028,6 @@ namespace aerobus {
 			>::type;
 		};
 
-public:
 		template<typename A, typename B, typename E = void>
 		struct floor_helper {};
 
@@ -2039,10 +2048,13 @@ public:
 
 		template<typename A, typename B>
 		struct floor_helper<A, B, std::enable_if_t<gt_helper<A, B>::value && (A::digits != 1 || B::digits != 1)>> {
+			static_assert(A::sign == signs::positive);
+			static_assert(B::sign == signs::positive);
 			static constexpr size_t N = A::digits;
 			static constexpr size_t K = B::digits;
-			// TODO: find better heuristic for boundaries
-			using from = typename one::template shift_left<N - K>;
+			static constexpr size_t min_shift = N >= K + 1 ? N - K - 1 : 0;
+
+			using from = typename one::template shift_left<min_shift>;
 			using to = typename one::template shift_left<N - K + 1>;
 
 			template<typename X, typename Y>
@@ -2063,7 +2075,7 @@ public:
 				gt_helper<upperbound, typename add<lowerbound, one>::type>::value &&
 				gt_helper<typename mul<average_t<upperbound, lowerbound>, B>::type, A>::value
 				>> {
-				using type = typename inner<lowerbound, average_t<upperbound, lowerbound>>::type;
+				using type = typename simplify<typename inner<lowerbound, average_t<upperbound, lowerbound>>::type>::type;
 			};
 
 			template<typename lowerbound, typename upperbound>
@@ -2071,7 +2083,7 @@ public:
 				gt_helper<upperbound, typename add<lowerbound, one>::type>::value &&
 				!gt_helper<typename mul<average_t<upperbound, lowerbound>, B>::type, A>::value
 				>> {
-				using type = typename  inner<average_t<upperbound, lowerbound>, upperbound>::type;
+				using type = typename simplify<typename inner<average_t<upperbound, lowerbound>, upperbound>::type>::type;
 			};
 
 			using type = typename inner<from, to>::type;
@@ -2079,6 +2091,8 @@ public:
 
 		template<typename N, typename M, int64_t i>
 		struct div_helper_inner {
+			static_assert(N::sign == signs::positive);
+			static_assert(M::sign == signs::positive);
 			static constexpr size_t l = M::digits;
[…]
