### 2024-01-26 7895874
add bigint from_hex "constructor"
const string must be all uppercase, no prexif, no suffix, no separator
