### 2024-01-26 8daac9c
add to_hex for bigint (sign + 0X + uppcase hexa string)
