### 2024-01-26 8e23eb3
skip conway polynomials from templight trace
