### 2024-01-26 6bf8b0b
better compile time
better heuristic for floor (make it private)
