### 2024-01-26 e410fd0
fix bigint::to_hex (missing trailing zeros)
+ make bigint officialy an euclidean domain
