+++ src/main.cpp
@@ -10,7 +10,9 @@
 #endif
 
 #ifdef _MSC_VER
+#ifndef _USE_MATH_DEFINES
 #define _USE_MATH_DEFINES
+#endif
 #endif
 
 // conformance : https://godbolt.org/z/qnfP99KWv
