### 2024-02-01 b6b8307
add quadratic extension of rationals
