### 2024-02-01 bd73955
fix CI (missing typename for dependant name)
