### 2024-02-01 d4df8da
migrate boolean static properties to constexpr instead of conditional types
+ rename Quadratic to QuadraticExtension
