+++ src/lib.h
@@ -514,17 +514,17 @@ namespace aerobus {
 
 		template<typename v1, typename v2>
 		struct gt {
-			using type = std::conditional_t<(v1::v > v2::v), std::true_type, std::false_type>;
+			static constexpr bool value = (v1::v > v2::v);
 		};
 
 		template<typename v1, typename v2>
 		struct lt {
-			using type = std::conditional_t<(v1::v < v2::v), std::true_type, std::false_type>;
+			static constexpr bool value = (v1::v < v2::v);
 		};
 
 		template<typename v1, typename v2>
 		struct eq {
-			using type = std::conditional_t<(v1::v == v2::v), std::true_type, std::false_type>;
+			static constexpr bool value =(v1::v == v2::v);
 		};
 
 	public:
@@ -554,15 +554,15 @@ namespace aerobus {
 
 		/// @brief strictly greater operator (v1 > v2)
 		template<typename v1, typename v2>
-		static constexpr bool gt_v = gt<v1, v2>::type::value;
+		static constexpr bool gt_v = gt<v1, v2>::value;
 
 		/// @brief strict less operator (v1 < v2)
 		template<typename v1, typename v2>
-		using lt_t = typename lt<v1, v2>::type;
+		static constexpr bool lt_v = lt<v1, v2>::value;
 
 		/// @brief equality operator
 		template<typename v1, typename v2>
-		static constexpr bool eq_v = eq<v1, v2>::type::value;
+		static constexpr bool eq_v = eq<v1, v2>::value;
 
 		/// @brief positivity (v1 > 0)
 		template<typename v1>
@@ -652,17 +652,17 @@ namespace aerobus {
 
 		template<typename v1, typename v2>
 		struct gt {
-			using type = std::conditional_t<(v1::v > v2::v), std::true_type, std::false_type>;
+			static constexpr bool value = (v1::v > v2::v);
 		};
 
 		template<typename v1, typename v2>
 		struct lt {
-			using type = std::conditional_t<(v1::v < v2::v), std::true_type, std::false_type>;
+			static constexpr bool value = (v1::v < v2::v);
 		};
 
 		template<typename v1, typename v2>
 		struct eq {
-			using type = std::conditional_t<(v1::v == v2::v), std::true_type, std::false_type>;
+			static constexpr bool value = (v1::v == v2::v);
 		};
 
 	public:
@@ -692,15 +692,15 @@ namespace aerobus {
 
 		/// @brief strictly greater operator (v1 > v2)
 		template<typename v1, typename v2>
-		static constexpr bool gt_v = gt<v1, v2>::type::value;
+		static constexpr bool gt_v = gt<v1, v2>::value;
 
 		/// @brief strict less operator (v1 < v2)
 		template<typename v1, typename v2>
-		using lt_t = typename lt<v1, v2>::type;
+		static constexpr bool lt_v = lt<v1, v2>::value;
 
 		/// @brief equality operator
 		template<typename v1, typename v2>
-		static constexpr bool eq_v = eq<v1, v2>::type::value;
+		static constexpr bool eq_v = eq<v1, v2>::value;
 
 		/// weirdly enough, for clang, this must be declared before gcd_t
 		/// @brief is v posititive 
@@ -776,22 +776,22 @@ namespace aerobus {
 
 		template<typename v1, typename v2>
 		struct gt {
-			using type = std::conditional_t<(v1::v% p > v2::v% p), std::true_type, std::false_type>;
+			static constexpr bool value = (v1::v% p > v2::v% p);
 		};
 
 		template<typename v1, typename v2>
 		struct lt {
[…]
