### 2024-02-01 3135857
add missing simplifications in bigint
bigrationals and their extensions work well now
