+++ src/lib.h
@@ -524,7 +524,7 @@ namespace aerobus {
 
 		template<typename v1, typename v2>
 		struct eq {
-			static constexpr bool value =(v1::v == v2::v);
+			static constexpr bool value = (v1::v == v2::v);
 		};
 
 	public:
@@ -776,7 +776,7 @@ namespace aerobus {
 
 		template<typename v1, typename v2>
 		struct gt {
-			static constexpr bool value = (v1::v% p > v2::v% p);
+			static constexpr bool value = (v1::v % p > v2::v % p);
 		};
 
 		template<typename v1, typename v2>
@@ -926,7 +926,7 @@ namespace aerobus {
 		using minus_t = sub_t<zero, v>;
 
 		template<typename v1, typename v2>
-		static constexpr bool eq_v = 
+		static constexpr bool eq_v =
 			(Field::template eq_v<typename v1::x, typename v2::x>) &&
 			(Field::template eq_v<typename v1::y, typename v2::y>);
 
@@ -1309,9 +1309,9 @@ namespace aerobus {
 			using qq = typename add<Q, pT>::type;
 
 		public:
-			using q_type = typename div_helper<A, B, qq, rr>::q_type;
-			using mod_type = typename div_helper<A, B, qq, rr>::mod_type;
-			using gcd_type = rr;
+			using q_type = typename simplify<typename div_helper<A, B, qq, rr>::q_type>::type;
+			using mod_type = typename simplify<typename div_helper<A, B, qq, rr>::mod_type>::type;
+			using gcd_type = typename simplify<rr>::type;
 		};
 
 		template<typename A, typename B>
@@ -2020,10 +2020,37 @@ namespace aerobus {
 		template<typename I1, typename I2, typename E = void>
 		struct sub {};
 
+
+
+		// 0 + x -> x
+		template<typename I1, typename I2>
+		struct add<I1, I2, std::enable_if_t<
+			I1::is_zero_v && !I2::is_zero_v
+			>> {
+			using type = I2;
+		};
+
+		// x + 0 -> x
+		template<typename I1, typename I2>
+		struct add<I1, I2, std::enable_if_t<
+			I2::is_zero_v && !I1::is_zero_v
+			>> {
+			using type = I1;
+		};
+
+		// 0 + 0 -> x
+		template<typename I1, typename I2>
+		struct add<I1, I2, std::enable_if_t<
+			I2::is_zero_v && I1::is_zero_v
+			>> {
+			using type = zero;
+		};
+
 		// +x + +y -> x + y
 		template<typename I1, typename I2>
 		struct add<I1, I2, std::enable_if_t<
-			gt_helper<I1, zero>::value&&
+			!I2::is_zero_v && !I1::is_zero_v &&
+			gt_helper<I1, zero>::value &&
 			gt_helper<I2, zero>::value
 			>> {
 			using type = typename simplify<
@@ -2037,32 +2064,17 @@ namespace aerobus {
 		// -x + -y -> -(x+y)
 		template<typename I1, typename I2>
 		struct add<I1, I2, std::enable_if_t<
-			gt_helper<zero, I1>::value&&
+			!I2::is_zero_v && !I1::is_zero_v &&
+			gt_helper<zero, I1>::value &&
 			gt_helper<zero, I2>::value
 			>> {
 			using type = typename add<typename I1::minus_t, typename I2::minus_t>::type::minus_t;
 		};
 
-		// 0 + x -> x
-		template<typename I1, typename I2>
-		struct add<I1, I2, std::enable_if_t<
-			I1::is_zero_v
-			>> {
-			using type = I2;
-		};
-
-		// x + 0 -> x
-		template<typename I1, typename I2>
-		struct add<I1, I2, std::enable_if_t<
-			I2::is_zero_v
[…]
