### 2024-02-01 fc9d7a1
add missing concept requirement on QuadraticExtension
