### 2024-02-02 e1e3ebd
add Hermite polynomials
