+++ src/main.cpp
@@ -1257,6 +1257,24 @@ int test_hermite() {
 	return 0;
 }
 
+int test_legendre() {
+	{
+		using L2 = legendre<2>;
+		static_assert(L2::coeff_at_t<2>::get<float>() == 1.5F);
+		static_assert(L2::coeff_at_t<1>::get<float>() == 0.0F);
+		static_assert(L2::coeff_at_t<0>::get<float>() == -0.5F);
+	}
+	{
+		using L3 = legendre<3>;
+		static_assert(L3::coeff_at_t<3>::get<float>() == 2.5F);
+		static_assert(L3::coeff_at_t<2>::get<float>() == 0.0F);
+		static_assert(L3::coeff_at_t<1>::get<float>() == -1.5F);
+		static_assert(L3::coeff_at_t<0>::get<float>() == 0.0F);
+	}
+
+	return 0;
+}
+
 int test_value_at() {
 	{
 		constexpr uint16_t x = aerobus::internal::value_at<0, 1, 2, 3>::value;
@@ -1987,6 +2005,7 @@ int main(int argc, char* argv[]) {
 	RUN_TEST(test_chebyshev)
 	RUN_TEST(test_laguerre)
 	RUN_TEST(test_hermite)
+	RUN_TEST(test_legendre)
 	RUN_TEST(test_continued_fraction)
 	RUN_TEST(test_concept_ring)
 	RUN_TEST(test_type_list)
