### 2024-02-03 23f46ac
add install target (cmake) fix warning update doc
