### 2024-02-03 0aa8ba4
update benchmarks and readme
