+++ src/benchmarks.cpp
@@ -14,6 +14,57 @@ extern "C" __m512d _ZGVeN8v_sin(__m512d x);
 #define DEGREE 13
 #endif
 
+
+template<int deg>
+INLINED double std_laguerre(const double x) {
+    return 1.0 - std::laguerre(deg, x);
+}
+
+template<int deg>
+INLINED double aero_laguerre(const double x) {
+	// 1 - L_12
+	using H = typename aerobus::pq64::template sub_t<
+					typename aerobus::pq64::one, 
+					typename aerobus::laguerre<deg>>;
+    return H::template eval<double>(x);
+}
+
+template<int deg>
+INLINED double std_laguerre_12(const double x) {
+    return std_laguerre<deg>(std_laguerre<deg>(std_laguerre<deg>(std_laguerre<deg>(
+           std_laguerre<deg>(std_laguerre<deg>(std_laguerre<deg>(std_laguerre<deg>(
+           std_laguerre<deg>(std_laguerre<deg>(std_laguerre<deg>(std_laguerre<deg>(x))))))))))));
+}
+
+template<int deg>
+INLINED double aero_laguerre_12(const double x) {
+    return aero_laguerre<deg>(aero_laguerre<deg>(aero_laguerre<deg>(aero_laguerre<deg>(
+           aero_laguerre<deg>(aero_laguerre<deg>(aero_laguerre<deg>(aero_laguerre<deg>(
+           aero_laguerre<deg>(aero_laguerre<deg>(aero_laguerre<deg>(aero_laguerre<deg>(x))))))))))));
+}
+
+template<size_t N>
+void stdh4(double* const __restrict output, const double* const __restrict input) {
+	double* d_o = (double*)__builtin_assume_aligned(output, 64);
+	double* d_i = (double*)__builtin_assume_aligned(input, 64);
+	#pragma omp parallel for
+	for (int64_t i = 0; i < N; ++i) {
+		d_o[i] = std_laguerre_12<12>(d_i[i]);
+	}
+}
+
+
+template<size_t N>
+void aerobush4(double* const __restrict output, const double* const __restrict input) {
+	double* d_o = (double*)__builtin_assume_aligned(output, 64);
+	double* d_i = (double*)__builtin_assume_aligned(input, 64);
+	#pragma omp parallel for
+	for (int64_t i = 0; i < N; ++i) {
+		d_o[i] = aero_laguerre_12<12>(d_i[i]);
+	}
+}
+
+
 template<int deg>
 INLINED
 double sin_12(const double x) {
@@ -24,14 +75,14 @@ double sin_12(const double x) {
 
 template<size_t N, int deg>
 INLINED
-void vsin_12(double * in, double * out) {
+void vsin_12(const double * const __restrict in, double * const __restrict out) {
 	static_assert(N % 32 == 0);
-	in = (double*) __builtin_assume_aligned(in, 64);
-	out = (double*) __builtin_assume_aligned(out, 64);
+	double *d_in = (double*) __builtin_assume_aligned(in, 64);
+	double *d_out = (double*) __builtin_assume_aligned(out, 64);
 #pragma omp parallel for
 	for (size_t i = 0; i < N; i += 8) {
 		for(size_t j = i; j < i+8; ++j) {
-			out[j] = sin_12<deg>(in[j]);
+			d_out[j] = sin_12<deg>(d_in[j]);
 		}
 	}
 }
@@ -44,14 +95,14 @@ double sin_12_slow(const double x) {
 
 template<size_t N>
 INLINED
-void vsin_slow(double * in, double * out) {
+void vsin_slow(const double * const __restrict in, double * const __restrict out) {
 	static_assert(N % 32 == 0);
[…]
