### 2024-02-04 ff57c62
better benchmarks (avx2 support) better report and better windows support
