+++ src/benchmarks.cpp
@@ -5,23 +5,93 @@
 #include "lib.h"
 #include <immintrin.h>
 
-const size_t N = 134217728/2;
+constexpr size_t N = 32*1024*1024;
 
-#ifndef _MSC_VER
-extern "C" __m512d _ZGVeN8v_sin(__m512d x);
+
+// sometimes, gcc and clang don't use fast math lib calls and stay scalar
+// using that, we force them to use vectorized version, no matter the optimization flags
+#ifdef __AVX512F__
+#ifdef __MSC_VER
+extern "C" __m512d __vdecl_sin8(__m512d x);
+INLINED __m512d vec_sin(__m512d x) {
+	return __vdecl_sin8(x);
+}
+#else
+extern "C" __m512d _ZGVeN8v_sin(__m512d x); 
 INLINED __m512d vec_sin(__m512d x) {
 	return _ZGVeN8v_sin(x);
 }
+#endif
 #else
+	#ifdef __AVX2__
+#ifdef _MSC_VER
 extern "C" __m256d __vdecl_sin4(__m256d x);
 INLINED __m256d vec_sin(__m256d x) {
 	return __vdecl_sin4(x);
 }
+#else 
+extern "C" __m256d _ZGVdN4v_sin(__m256d x);
+INLINED __m256d vec_sin(__m256d x) {
+	return _ZGVdN4v_sin(x);
+}
+#endif
+	#else
+#error "benchmark require avx2 or avx512"
+	#endif
 #endif
 
-#ifndef DEGREE 
-#define DEGREE 13
-#endif
+
+
+template<int deg>
+INLINED double std_laguerre(const double x) {
+	return 1.0 - std::laguerre(deg, x);
+}
+
+template<int deg>
+INLINED double aero_laguerre(const double x) {
+	// 1 - L_12
+	using H = typename aerobus::pq64::template sub_t<
+		typename aerobus::pq64::one,
+		typename aerobus::laguerre<deg>>;
+	return H::template eval<double>(x);
+}
+
+template<int deg>
+INLINED double std_laguerre_12(const double x) {
+	return std_laguerre<deg>(std_laguerre<deg>(std_laguerre<deg>(std_laguerre<deg>(
+		std_laguerre<deg>(std_laguerre<deg>(std_laguerre<deg>(std_laguerre<deg>(
+			std_laguerre<deg>(std_laguerre<deg>(std_laguerre<deg>(std_laguerre<deg>(x))))))))))));
+}
+
+template<int deg>
+INLINED double aero_laguerre_12(const double x) {
+	return aero_laguerre<deg>(aero_laguerre<deg>(aero_laguerre<deg>(aero_laguerre<deg>(
+		aero_laguerre<deg>(aero_laguerre<deg>(aero_laguerre<deg>(aero_laguerre<deg>(
+			aero_laguerre<deg>(aero_laguerre<deg>(aero_laguerre<deg>(aero_laguerre<deg>(x))))))))))));
+}
+
+template<size_t N>
+void stdh4(double* const __restrict output, const double* const __restrict input) {
+	double* d_o = (double*)__builtin_assume_aligned(output, 64);
+	double* d_i = (double*)__builtin_assume_aligned(input, 64);
+#pragma omp parallel for
+	for (int64_t i = 0; i < N; ++i) {
+		d_o[i] = std_laguerre_12<12>(d_i[i]);
+	}
+}
+
+
+template<size_t N>
+void aerobush4(double* const __restrict output, const double* const __restrict input) {
+	double* d_o = (double*)__builtin_assume_aligned(output, 64);
+	double* d_i = (double*)__builtin_assume_aligned(input, 64);
[…]
