### 2024-02-05 6534044
fix cmake on linux + script to run cmake + fix all warnings in benchmarks and tests
