+++ cmake.sh
@@ -1,0 +1,3 @@
+rm -rf build
+mkdir build
+cd build && cmake ..
