+++ CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 3.1...3.27)
+cmake_minimum_required(VERSION 3.27)
 
 project(
     Aerobus
@@ -8,51 +8,37 @@ project(
 
 find_package(OpenMP REQUIRED)
 
+set(AVX_FLAGS)
+set(CMAKE_REQUIRED_FLAGS)
 include(CheckCXXSourceRuns)
 
-set(CMAKE_REQUIRED_FLAGS)
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -mavx2 -mfma")
 
 check_cxx_source_runs("
         #include <immintrin.h>
         int main()
         {
-          __m512d a, b, c;
-          const double src[8] = { 1., 2., 3., 4., 5., 6., 7., 8. };
-          double dst[8];
-          a =  _mm512_loadu_pd(src);
-          b =  _mm512_loadu_pd(src);
-          c =  _mm512_loadu_pd(src);
-          d = _mm512_fmadd_pd(a, b, c);
-          _mm512_storeu_pd(dst, c);
-          for(int i = 0; i < 8; i++) {
-            if((src[i] * src[i] + src[i]) != dst[i]) {
-              return -1;
-            }
-          }
+          __m256 a = _mm256_fmadd_ps(
+            _mm256_set1_ps(0.0F),
+            _mm256_set1_ps(0.0F),
+            _mm256_set1_ps(0.0F));
+          return 0;
+        }"
+        HAVE_AVX_EXTENSIONS)
+
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -mavx512f -mavx512vl")
+
+check_cxx_source_runs("
+        #include <immintrin.h>
+        int main()
+        {
+          __m512 a = _mm512_fmadd_ps(
+            _mm512_set1_ps(0.0F),
+            _mm512_set1_ps(0.0F),
+            _mm512_set1_ps(0.0F));
           return 0;
         }"
         HAVE_AVX512_EXTENSIONS)
-
-check_cxx_source_runs("
-        #include <immintrin.h>
-        int main()
-        {
-          __m256d a, b, c, d;
-          const double src[4] = { 1., 2., 3., 4. };
-          double dst[4];
-          a =  _mm256_loadu_pd(src);
-          b =  _mm256_loadu_pd(src);
-          c =  _mm256_loadu_pd(src);
-          d = _mm256_fmadd_pd(a, b, c);
-          _mm256_storeu_pd(dst, d);
-          for(int i = 0; i < 4; i++) {
-            if((src[i] * src[i] + src[i]) != dst[i]) {
-              return -1;
-            }
-          }
-          return 0;
-        }"
-        HAVE_AVX2_FMA_EXTENSIONS)
 
 
 set(CMAKE_CXX_STANDARD 20)
@@ -66,7 +52,7 @@ if(HAVE_AVX512_EXTENSIONS)
     else()
       set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -mavx512f -mavx512vl -mfma")
     endif()
-elseif(HAVE_AVX2_FMA_EXTENSIONS)
+elseif(HAVE_AVXFMA_EXTENSIONS)
     message("AVX2+FMA detected, adding flags")
     if(MSVC)
[…]
