+++ src/lib.h
@@ -29,28 +29,30 @@
 
 // aligned allocation
 namespace aerobus {
-	/**
-	 * aligned allocation of count elements of type T
-	 * @tparam T the type of elements to store
-	 * @param count the number of elements
-	 * @param alignment alignment boundary
-	*/
-	template<typename T>
-	T* aligned_malloc(size_t count, size_t alignment) {
-#ifdef _MSC_VER
-		return static_cast<T*>(_aligned_malloc(count * sizeof(T), alignment));
-#else
-		return static_cast<T*>(aligned_alloc(alignment, count * sizeof(T)));
-#endif
-	}
+	namespace memory {
+		/**
+		 * aligned allocation of count elements of type T
+		 * @tparam T the type of elements to store
+		 * @param count the number of elements
+		 * @param alignment alignment boundary
+		*/
+		template<typename T>
+		T* aligned_malloc(size_t count, size_t alignment) {
+	#ifdef _MSC_VER
+			return static_cast<T*>(_aligned_malloc(count * sizeof(T), alignment));
+	#else
+			return static_cast<T*>(aligned_alloc(alignment, count * sizeof(T)));
+	#endif
+		}
 
-	template<typename T>
-	void aligned_free(T* memblock) {
-#ifdef _MSC_VER
-		_aligned_free((void*)memblock);
-#else
-		free((void*)memblock);
-#endif
+		template<typename T>
+		void aligned_free(T* memblock) {
+	#ifdef _MSC_VER
+			_aligned_free((void*)memblock);
+	#else
+			free((void*)memblock);
+	#endif
+		}
 	}
 
 	namespace internal {
@@ -60,10 +62,10 @@ namespace aerobus {
 	/// <summary>
 	/// checks if v is in arr
 	/// </summary>
-	/// <typeparam name="T"></typeparam>
-	/// <typeparam name="N"></typeparam>
-	/// <param name="arr"></param>
-	/// <param name="v"></param>
+	/// <typeparam name="T">type of elements in arr</typeparam>
+	/// <typeparam name="N">size of arr</typeparam>
+	/// <param name="arr">array of N elements of type T</param>
+	/// <param name="v">value to search</param>
 	/// <returns></returns>
 	template<typename T, size_t N>
 	constexpr bool contains(const std::array<T, N>& arr, const T& v) {
@@ -76,6 +78,8 @@ namespace aerobus {
 		return false;
 	}
 
+	/// @brief a constexpr "string" utility
+	/// @tparam N sizeof string
 	template <size_t N>
 	struct string_literal {
 		constexpr string_literal(const char(&str)[N]) {
@@ -84,10 +88,11 @@ namespace aerobus {
 
 		template<size_t i>
 		constexpr char char_at() const {
-			if constexpr (i < N) {
-				return this->value[i];
+			// first char is always \0
+			if constexpr (i + 1 < N) {
+				return this->value[i + 1];
 			}
-			return 0;
+			return '\0';
 		}
 
 		constexpr size_t len() const { return N; }
@@ -510,7 +515,7 @@ namespace aerobus {
 
 		template<typename v1, typename v2>
 		struct mul {
-			using type = val<v1::v* v2::v>;
+			using type = val<v1::v * v2::v>;
 		};
 
 		template<typename v1, typename v2>
@@ -2347,21 +2352,21 @@ namespace aerobus {
 			static constexpr size_t N = S.len();
 
 			template<size_t i>
-			static constexpr char char_at = (i < N) ? S.template char_at<i>() : '0';
+			static constexpr char char_at = (i + 1 < N) ? S.template char_at<i>() : '0';
 
 			template <char c>
[…]
