### 2024-02-06 7437687
solve #18
static_assert in tests instead of if which make no sense
