+++ src/main.cpp
@@ -20,15 +20,9 @@
 using namespace aerobus;
 
 int test_type_at() {
-	if (!std::is_same<internal::type_at_t<0, float, int, long>, float>::value) {
-		return 1;
-	}
-	if (!std::is_same<internal::type_at_t<1, float, int, long>, int>::value) {
-		return 1;
-	}
-	if (!std::is_same<internal::type_at_t<2, float, int, long>, long>::value) {
-		return 1;
-	}
+	static_assert(std::is_same<internal::type_at_t<0, float, int, long>, float>::value);
+	static_assert(std::is_same<internal::type_at_t<1, float, int, long>, int>::value);
+	static_assert(std::is_same<internal::type_at_t<2, float, int, long>, long>::value);
 
 	return 0;
 }
@@ -37,15 +31,11 @@ int test_poly_simplify() {
 	using poly1 = polynomial<i32>::val<i32::val<0>, i32::val<1>, i32::val<2>>;
 	using simplified1 = polynomial<i32>::simplify_t<poly1>;
 	using expected1 = polynomial<i32>::val<i32::val<1>, i32::val<2>>;
-	if (!std::is_same<expected1, simplified1>::value) {
-		return 1;
-	}
+	static_assert(std::is_same<expected1, simplified1>::value);
 
 	using poly2 = polynomial<i32>::val<i32::val<12>>;
 	using simplified2 = polynomial<i32>::simplify_t<poly2>;
-	if (!std::is_same<poly2, simplified2>::value) {
-		return 1;
-	}
+	static_assert(std::is_same<poly2, simplified2>::value);
 
 	return 0;
 }
@@ -54,24 +44,16 @@ int test_poly_eval() {
 	// 1 + 2x + 3x^2
 	using poly = polynomial<i32>::val<i32::val<3>, i32::val<2>, i32::val<1>>;
 	constexpr int v = poly::eval(1);
-	if (v != 6) {
-		return 1;
-	}
+	static_assert(v == 6);
 	constexpr float vv = poly::eval(1.0f);
-	if (vv != 6.0f) {
-		return 1;
-	}
+	static_assert(vv == 6.0F);
 
 	// 1/2 + 3x/2
 	using polyf = polynomial<q32>::val<q32::val<i32::val<3>, i32::val<2>>, q32::val<i32::val<1>, i32::val<2>>>;
 	constexpr float vvv = polyf::eval(1.0f);
-	if (vvv != 2.0f) {
-		return 1;
-	}
+	static_assert(vvv == 2.0F);
 	constexpr double vvvv = polyf::eval(-1.0);
-	if (vvvv != -1.0) {
-		return 1;
-	}
+	static_assert(vvvv == -1.0);
 
 	return 0;
 }
@@ -79,33 +61,23 @@ int test_poly_eval() {
 int test_fraction_field_eval() {
 	using half = q32::val<i32::one, i32::val<2>>;
 	constexpr float x = half::eval(2.0f);
-	if (x != 0.5f) {
-		return 1;
-	}
+	static_assert(x == 0.5F);
 	using thirdhalf = q32::val<i32::val<3>, i32::val<2>>;
 	constexpr float y = thirdhalf::eval(1.0f);
-	if (y != 1.5f) {
-		return 1;
-	}
+	static_assert(y == 1.5F);
 
 	// 3/2 + x / 2
 	using polyA = polynomial<q32>::val<half, thirdhalf>;
[…]
