### 2024-04-12 #6233 @JeWaVe
Indeed. Thats me in both cases.
For some obscure reason, when I commit from one computer, it uses a différent id.

I dont know how to prove it.

Envoyé depuis Proton Mail pour Android

-------- Message d'origine --------
Le 12/04/2024 17:33, Nikos Pitsianis  a écrit :

> The submitted paper lists a single author; however, the git repository has non-trivial contributions from two different ids. Is this correct?
>
