### 2024-04-12 #6233 @pitsianis
Good to know ;-)
I believe you can resolve this for your next project. Find the file `~/.git-credentials` and rename it.
Then issue `git config credential.helper store`
With the next `git` interaction, it will ask you for the username/password and store it.
