### 2024-05-28 #6233 @pitsianis
Here is a mixed bag of comments I collected over my trials.

The building instructions are incomplete.
 * After `cmake` I also need to issue `make`.
 * `make benchmarks` has to be issued in the `src` directory
 * The benchmark output does not match the one shown in the `HOW TO`

The installation instructions do not list any OS/Software/Hardware dependencies.
Building the library failed on an M2 MacBookPro with the error
> Could NOT find OpenMP_CXX (missing: OpenMP_CXX_FLAGS OpenMP_CXX_LIB_NAMES)

The summary and paper have areas that can be improved. In its current form, the summary consists of several paragraphs, many of which repeat "We also" and "In addition" four times. I suggest reorganizing it and separating the library's features from its demonstrations. The paper also reads like a list of items without a clear purpose. I do not find the listings of the generated assembly code useful.

I find the benchmark with the function `sin` that trades off accuracy for speed *and* it outperforms the standard library intriguing. I wish *the use* of the library was described in detail and not the implementation of macros of the library, like the implementation of Horner's rule for polynomial evaluation. In this case, wouldn't a name like `horners_rule` be more informative than `eval_helper`?

Writing macros --especially recursive ones-- is very hard! Even using such macros is hard. This work deserves publication, but this paper could be improved to make the presentation of the Aerobus macros more attractive.
