### 2024-06-02 8ba0aa5
improve readme - install instructions and how to run benchmarks and tests
