### 2024-06-02 #6233 @JeWaVe
Hi and thank you for the comments.
I just pushed a "fixed" readme with better instructions for tests/bench.

The rest (reorganization) requires quite a large amount of work so I prefer waiting for the opinion of others reviewers before engaging it.
