### 2024-06-04 #6233 @gkthiruvathukal
@mmoelle1, @lucaferranti Can either of you please provide additional comments based on your experience with the submission?

@JeWaVe I suggest working to address the input. In general, I consider reviewer input "stateless". That is, everyone is supposed to independently review the submission.  @pitsianis has made some helpful comments that don't look controversial to me. In the end, the goal of getting reviewer input is to improve the JOSS submission. The more you can do to improve the submission, the sooner it can move to the next stages.
