### 2024-06-05 #6233 @lucaferranti
Hi all,

first of all, I am deeply sorry for the delay in the review. This kept slipping through in my TODO list, but all explanations are in the end just excuses.

Overall, I find the work very valuable and worth publishing and appropriate for the forum. I work on computational algebraic geometry and see value in this.

Now some concrete feedback

## Paper

My main comment for the paper is that at the moment I think it does not conform to the JOSS format. Mainly, the paper is currently 12 pages, while generally JOSS expects a much shorter (~ 2 pages) paper, which should give a general motivation and overview of timeless features of the sofware. Currently, the paper presents several examples and code snippets, which I believe would better fit the documentation. The reason for this is that software can change and things written in the paper are frozen, while documentation can evolve and update (see also my comment on documentation below). However I let the editor @danielskatz overwrite my comment if he thinks the paper is appropriate this way.

## Documentation

- [ ] Currently the documentation is a pdf file which is built on CI, I believe hosting the documentation on github pages e.g. using doxygen would improve usability significantly. A lot of the examples in the paper should be moved there too
- [ ] contribution guidelines are missing. As a minimum, some notes in the readme are needed. Bonus points to set-up e.g. a PR template and issue template.
- [ ] **Installation instructions**: I find the installation instructions (particularly the "How to" section) a bit minimal. Mainly
> Clone or download the repo somewhere
> include "lib.h" in code

is very prone to relative path issues. I think it would be more user friendly to expand this section a bit

## Testing

- Tests at the moment are only run locally, have you considered setting up a CI to run the tests on every PR and on the main branch? Would make the workflow slicker I think, especially to review external contributions.

Overall, very nice work and I hope to see this published. Hopefully you find these comments useful

O
