### 2024-06-06 #6233 @gkthiruvathukal
@lucaferranti
These are excellent comments. I'd like to see all on your checklist addressed.

Briefly:
- The paper should NOT be 12 pages. Please shorten, @JeWaVe. While I have seen longer than 2 pages, it should not be *much* longer than that. @danielskatz can override me, of course.
- Docs should *not* be PDF only. Consider the options provided by @lucaferranti. There are so many good tools to build doc sites but doxygen is one of the ones most ideal for C++ API docs.
- Although the points about "testing" are not strictly required by JOSS, I have advocated for testing via CI to be required. Ideally, mathematically focused codes should use some form of "property" checking, e.g. https://github.com/emil-e/rapidcheck. For other possibilities, see the Awesome C++ resources at https://github.com/fffaraz/awesome-cpp.
