### 2024-06-11 9d71d12
add cppcheck & cpplint + please them + integrate in CI + better test report + update readme accordingly
