+++ src/main.cpp
@@ -4,7 +4,7 @@
 #include <cmath>
 #include <cstdlib>
 
-#include "lib.h"
+#include "./lib.h"
 #include "../imports/conwaypolynomials.h"
 
 #ifdef _MSC_VER
@@ -13,124 +13,124 @@
 
 // conformance : https://godbolt.org/z/qnfP99KWv
 
-using namespace aerobus;
+using namespace aerobus;  // NOLINT
 
 int test_type_at() {
-	if (!std::is_same<internal::type_at_t<0, float, int, long>, float>::value) {
-		return 1;
-	}
-	if (!std::is_same<internal::type_at_t<1, float, int, long>, int>::value) {
-		return 1;
-	}
-	if (!std::is_same<internal::type_at_t<2, float, int, long>, long>::value) {
-		return 1;
-	}
+    if (!std::is_same<internal::type_at_t<0, float, int, int64_t>, float>::value) {
+        return 1;
+    }
+    if (!std::is_same<internal::type_at_t<1, float, int, int64_t>, int>::value) {
+        return 1;
+    }
+    if (!std::is_same<internal::type_at_t<2, float, int, int64_t>, int64_t>::value) {
+        return 1;
+    }
 
-	return 0;
+    return 0;
 }
 
 int test_poly_simplify() {
-	using poly1 = polynomial<i32>::val<i32::val<0>, i32::val<1>, i32::val<2>>;
-	using simplified1 = polynomial<i32>::simplify_t<poly1>;
-	using expected1 = polynomial<i32>::val<i32::val<1>, i32::val<2>>;
-	if (!std::is_same<expected1, simplified1>::value) {
-		return 1;
-	}
+    using poly1 = polynomial<i32>::val<i32::val<0>, i32::val<1>, i32::val<2>>;
+    using simplified1 = polynomial<i32>::simplify_t<poly1>;
+    using expected1 = polynomial<i32>::val<i32::val<1>, i32::val<2>>;
+    if (!std::is_same<expected1, simplified1>::value) {
+        return 1;
+    }
 
-	using poly2 = polynomial<i32>::val<i32::val<12>>;
-	using simplified2 = polynomial<i32>::simplify_t<poly2>;
-	if (!std::is_same<poly2, simplified2>::value) {
-		return 1;
-	}
+    using poly2 = polynomial<i32>::val<i32::val<12>>;
+    using simplified2 = polynomial<i32>::simplify_t<poly2>;
+    if (!std::is_same<poly2, simplified2>::value) {
+        return 1;
+    }
 
-	return 0;
+    return 0;
 }
 
 int test_poly_eval() {
-	// 1 + 2x + 3x^2
-	using poly = polynomial<i32>::val<i32::val<3>, i32::val<2>, i32::val<1>>;
-	constexpr int v = poly::eval(1);
-	if (v != 6) {
-		return 1;
-	}
-	constexpr float vv = poly::eval(1.0f);
-	if (vv != 6.0f) {
-		return 1;
-	}
+    // 1 + 2x + 3x^2
+    using poly = polynomial<i32>::val<i32::val<3>, i32::val<2>, i32::val<1>>;
+    constexpr int v = poly::eval(1);
+    if (v != 6) {
+        return 1;
+    }
+    constexpr float vv = poly::eval(1.0f);
+    if (vv != 6.0f) {
+        return 1;
+    }
 
-	// 1/2 + 3x/2
-	using polyf = polynomial<q32>::val<q32::val<i32::val<3>, i32::val<2>>, q32::val<i32::val<1>, i32::val<2>>>;
[…]
