+++ src/lib.h
@@ -1,6 +1,8 @@
 // -*- lsst-c++ -*-
+#ifndef __INC_AEROBUS__ // NOLINT
+#define __INC_AEROBUS__
 
-#include <cstdint> // NOLINT(clang-diagnostic-pragma-pack)
+#include <cstdint>
 #include <cstddef>
 #include <cstring>
 #include <type_traits>
@@ -8,311 +10,302 @@
 #include <algorithm>
 #include <functional>
 #include <string>
-#include <concepts>
+#include <concepts> // NOLINT
 #include <array>
 
 
 #ifdef _MSC_VER
 #define ALIGNED(x) __declspec(align(x))
-#define INLINED __forceinline 
-#else 
+#define INLINED __forceinline
+#else
 #define ALIGNED(x) __attribute__((aligned(x)))
 #define INLINED __attribute__((always_inline)) inline
 #endif
 
 // aligned allocation
 namespace aerobus {
-	/**
-	 * aligned allocation of count elements of type T
-	 * @tparam T the type of elements to store
-	 * @param count the number of elements
-	 * @param alignment alignment boundary
-	*/
-	template<typename T>
-	T* aligned_malloc(size_t count, size_t alignment) {
-		#ifdef _MSC_VER
-		return static_cast<T*>(_aligned_malloc(count * sizeof(T), alignment));
-		#else
-		return static_cast<T*>(aligned_alloc(alignment, count * sizeof(T)));
-		#endif
-	}
+    /**
+     * aligned allocation of count elements of type T
+     * @tparam T the type of elements to store
+     * @param count the number of elements
+     * @param alignment alignment boundary
+    */
+    template<typename T>
+    T* aligned_malloc(size_t count, size_t alignment) {
+        #ifdef _MSC_VER
+        return static_cast<T*>(_aligned_malloc(count * sizeof(T), alignment));
+        #else
+        return static_cast<T*>(aligned_alloc(alignment, count * sizeof(T)));
+        #endif
+    }
 
[…]
