+++ CPPLINT.cfg
@@ -1,0 +1,5 @@
+filter=-legal/copyright
+
+filter=-runtime/indentation_namespace
+
+linelength=120
