+++ .github/workflows/c-cpp.yml
@@ -14,8 +14,15 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - name: Install compilers and lib
-      run: sudo apt-get install -y clang g++ libomp-dev
+      run: sudo apt-get install -y clang g++ libomp-dev cppcheck
+    - name: cppcheck
+      if: success()
+      run: cppcheck --std=c++20 --language=c++ src/lib.h && cppcheck --std=c++20 --language=c++ imports/conwaypolynomials.h && cppcheck --std=c++20 --language=c++ src/main.cpp
+    - name: cpplint
+      if: success()
+      run: cpplint src/lib.h && cpplint src/main.cpp
     - name: make
+      if: success()
       run: cd src && make build run
   paper:
     runs-on: ubuntu-latest
