### 2024-06-11 3336de7
fix CI + split build and run tests
