+++ .github/workflows/c-cpp.yml
@@ -14,7 +14,7 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - name: Install compilers and lib
-      run: sudo apt-get install -y clang g++ libomp-dev cppcheck
+      run: sudo apt-get install -y clang g++ libomp-dev cppcheck cpplint
     - name: cppcheck
       if: success()
       run: cppcheck --std=c++20 --language=c++ src/lib.h && cppcheck --std=c++20 --language=c++ imports/conwaypolynomials.h && cppcheck --std=c++20 --language=c++ src/main.cpp
@@ -24,6 +24,9 @@ jobs:
     - name: make
       if: success()
       run: cd src && make build run
+    - name: test
+      if: success()
+      run: cd src && make run
   paper:
     runs-on: ubuntu-latest
     name: Paper Draft
