### 2024-06-11 96394ed
test if CI is effective - should fail when lint fails
