+++ .github/workflows/c-cpp.yml
@@ -20,10 +20,10 @@ jobs:
       run: cppcheck --std=c++20 --language=c++ src/lib.h && cppcheck --std=c++20 --language=c++ imports/conwaypolynomials.h && cppcheck --std=c++20 --language=c++ src/main.cpp
     - name: cpplint
       if: success()
-      run: cpplint src/lib.h && cpplint src/main.cpp
+      run: cpplint src/lib.h && cpplint src/main.cpp && cpplint imports/conwaypolynomials.h
     - name: make
       if: success()
-      run: cd src && make build run
+      run: cd src && make build
     - name: test
       if: success()
       run: cd src && make run
