### 2024-06-11 #6233 @JeWaVe
Hi

Thanks for the very insightful comments and thanks @gkthiruvathukal for the summary.

- I added a check and a linter to the CI which now reports tests correctly. However, I can't really use a "classic" testing library/framework because everything is typed-based. That's why I did my own code in main.cpp.
Now it reports all the passed tests and fails at the first in error, reporting proper status in console. Please tell me if it's OK for you.

- I'll update doc and paper according to your suggestions (documentation will take me some time though).
