### 2024-06-12 #6233 @gkthiruvathukal
@JeWaVe Thanks for your response & support for my remarks. I always want submissions to be great, even if not all feedback can be immediately addressed.

Can you please be clearer about why you cannot make your tests report all test failures instead of stopping on the first? Why can't you you use Google tests to do what you are describing? when it comes to type-based testing? If you can point me to a range of lines of code, I may be able to suggest a concrete direction.

In any event, we're making nice progress here!
