### 2024-08-13 #6233 @pitsianis
Let's add an extra week, for instance, until Aug. 23. In Europe, especially in the Mediterranean, July 15-Aug. 15 is the "dead" zone summer vacation period. Otherwise, I agree with the assessment of this submission.
