### 2024-08-13 #6233 @lucaferranti
Indeed 3 days notice in holiday period does sound a bit tight...
