### 2024-08-31 #6233 @JeWaVe
first things first : use google tests.
