### 2024-08-31 aa49d58
attempt to introduce google tests
