+++ src/lib_tests.cpp
@@ -3,6 +3,7 @@
 #include <array>
 #include <cmath>
 #include <cstdlib>
+#include <gtest/gtest.h>
 
 #include "./lib.h"
 #include "../imports/conwaypolynomials.h"
@@ -15,95 +16,61 @@
 
 using namespace aerobus;  // NOLINT
 
-int test_type_at() {
-    if (!std::is_same<internal::type_at_t<0, float, int, int64_t>, float>::value) {
-        return 1;
-    }
-    if (!std::is_same<internal::type_at_t<1, float, int, int64_t>, int>::value) {
-        return 1;
-    }
-    if (!std::is_same<internal::type_at_t<2, float, int, int64_t>, int64_t>::value) {
-        return 1;
-    }
-
-    return 0;
+TEST(utilies, type_at) {
+    EXPECT_TRUE((std::is_same<internal::type_at_t<0, float, int, int64_t>, float>::value));
+    EXPECT_TRUE((std::is_same<internal::type_at_t<1, float, int, int64_t>, int>::value));
+    EXPECT_TRUE((std::is_same<internal::type_at_t<2, float, int, int64_t>, int64_t>::value));
 }
 
-int test_poly_simplify() {
+TEST(polynomials, simplify) {
     using poly1 = polynomial<i32>::val<i32::val<0>, i32::val<1>, i32::val<2>>;
     using simplified1 = polynomial<i32>::simplify_t<poly1>;
     using expected1 = polynomial<i32>::val<i32::val<1>, i32::val<2>>;
-    if (!std::is_same<expected1, simplified1>::value) {
-        return 1;
-    }
+    EXPECT_TRUE((std::is_same<expected1, simplified1>::value));
 
     using poly2 = polynomial<i32>::val<i32::val<12>>;
     using simplified2 = polynomial<i32>::simplify_t<poly2>;
-    if (!std::is_same<poly2, simplified2>::value) {
-        return 1;
-    }
-
-    return 0;
+    EXPECT_TRUE((std::is_same<poly2, simplified2>::value));
 }
 
-int test_poly_eval() {
+TEST(polynomials, eval) {
     // 1 + 2x + 3x^2
     using poly = polynomial<i32>::val<i32::val<3>, i32::val<2>, i32::val<1>>;
     constexpr int v = poly::eval(1);
-    if (v != 6) {
-        return 1;
-    }
+    EXPECT_EQ(v, 6);
     constexpr float vv = poly::eval(1.0f);
-    if (vv != 6.0f) {
-        return 1;
-    }
+    EXPECT_EQ(vv, 6.0f);
 
     // 1/2 + 3x/2
     using polyf = polynomial<q32>::val<q32::val<i32::val<3>, i32::val<2>>, q32::val<i32::val<1>, i32::val<2>>>;
     constexpr float vvv = polyf::eval(1.0f);
-    if (vvv != 2.0f) {
-        return 1;
-    }
-    constexpr double vvvv = polyf::eval(-1.0);
-    if (vvvv != -1.0) {
-        return 1;
-    }
+    EXPECT_EQ(vvv, 2.0f);
 
-    return 0;
+    constexpr double vvvv = polyf::eval(-1.0);
+    EXPECT_EQ(vvvv, -1.0);
 }
 
-int test_fraction_field_eval() {
+TEST(fraction_field, eval) {
     using half = q32::val<i32::one, i32::val<2>>;
     constexpr float x = half::eval(2.0f);
[…]
