### 2024-08-31 4f58c50
fix is_prime + move tests to gtests
