+++ src/lib_tests.cpp
@@ -17,20 +17,20 @@
 using namespace aerobus;  // NOLINT
 
 TEST(utilies, type_at) {
-    EXPECT_TRUE((std::is_same<internal::type_at_t<0, float, int, int64_t>, float>::value));
-    EXPECT_TRUE((std::is_same<internal::type_at_t<1, float, int, int64_t>, int>::value));
-    EXPECT_TRUE((std::is_same<internal::type_at_t<2, float, int, int64_t>, int64_t>::value));
+    EXPECT_TRUE((std::is_same_v<internal::type_at_t<0, float, int, int64_t>, float>));
+    EXPECT_TRUE((std::is_same_v<internal::type_at_t<1, float, int, int64_t>, int>));
+    EXPECT_TRUE((std::is_same_v<internal::type_at_t<2, float, int, int64_t>, int64_t>));
 }
 
 TEST(polynomials, simplify) {
     using poly1 = polynomial<i32>::val<i32::val<0>, i32::val<1>, i32::val<2>>;
     using simplified1 = polynomial<i32>::simplify_t<poly1>;
     using expected1 = polynomial<i32>::val<i32::val<1>, i32::val<2>>;
-    EXPECT_TRUE((std::is_same<expected1, simplified1>::value));
+    EXPECT_TRUE((std::is_same_v<expected1, simplified1>));
 
     using poly2 = polynomial<i32>::val<i32::val<12>>;
     using simplified2 = polynomial<i32>::simplify_t<poly2>;
-    EXPECT_TRUE((std::is_same<poly2, simplified2>::value));
+    EXPECT_TRUE((std::is_same_v<poly2, simplified2>));
 }
 
 TEST(polynomials, eval) {
@@ -73,7 +73,7 @@ TEST(fraction_field, eval) {
     EXPECT_EQ(zz, -2.0f);
 }
 
-int test_coeff_at() {
+TEST(polynomials, coeff_at) {
     // 1 + 2x + 3x^2
     using poly = polynomial<i32>::val<i32::val<3>, i32::val<2>, i32::val<1>>;
     using at0 = poly::coeff_at_t<0>;
@@ -81,23 +81,10 @@ int test_coeff_at() {
     using at2 = poly::coeff_at_t<2>;
     using at3 = poly::coeff_at_t<3>;
 
-    if (!std::is_same<at0, i32::val<1>>::value) {
-        return 1;
-    }
-
-    if (!std::is_same<at1, i32::val<2>>::value) {
-        return 1;
-    }
-
-    if (!std::is_same<at2, i32::val<3>>::value) {
-        return 1;
-    }
-
-    if (!std::is_same<at3, i32::val<0>>::value) {
-        return 1;
-    }
-
-    return 0;
+    EXPECT_TRUE((std::is_same_v<at0, i32::val<1>>));
+    EXPECT_TRUE((std::is_same_v<at1, i32::val<2>>));
+    EXPECT_TRUE((std::is_same_v<at2, i32::val<3>>));
+    EXPECT_TRUE((std::is_same_v<at3, i32::val<0>>));
 }
 
 template<typename... coeffs>
@@ -110,7 +97,7 @@ using add_ix = polynomial<i32>::add_t< P1, P2>;
 template<typename P1, typename P2>
 using sub_ix = polynomial<i32>::sub_t<P1, P2>;
 
-int test_poly_add() {
+TEST(polynomials, add) {
     {
         // 1 + x
         using P1 = IX<Int<1>, Int<1>>;
@@ -119,18 +106,10 @@ int test_poly_add() {
         // 2 + 2x + x^2
         using A = polynomial<i32>::add_t<P1, P2>;
 
-        if (A::coeff_at_t<0>::v != 2) {
[…]
