### 2024-08-31 #6233 @JeWaVe
I moved tests to google tests in main branch.
Now : update readme + install / usage / tests instructions
