### 2024-08-31 1f6205c
rename main file to aerobus.h\nmove conwaypolynomials.h into main include\nplease cpplint everywhere\nfix benchmarks on linux -- avx512 detection\nupdate readme
