+++ src/lib.h
@@ -1,2412 +1,0 @@
-// -*- lsst-c++ -*-
-#ifndef __INC_AEROBUS__ // NOLINT
-#define __INC_AEROBUS__
-
-#include <cstdint>
-#include <cstddef>
-#include <cstring>
-#include <type_traits>
-#include <utility>
-#include <algorithm>
-#include <functional>
-#include <string>
-#include <concepts> // NOLINT
-#include <array>
-
-
-#ifdef _MSC_VER
-#define ALIGNED(x) __declspec(align(x))
-#define INLINED __forceinline
-#else
-#define ALIGNED(x) __attribute__((aligned(x)))
-#define INLINED __attribute__((always_inline)) inline
-#endif
-
-// aligned allocation
-namespace aerobus {
-    /**
-     * aligned allocation of count elements of type T
-     * @tparam T the type of elements to store
-     * @param count the number of elements
-     * @param alignment alignment boundary
-    */
-    template<typename T>
-    T* aligned_malloc(size_t count, size_t alignment) {
-        #ifdef _MSC_VER
-        return static_cast<T*>(_aligned_malloc(count * sizeof(T), alignment));
-        #else
-        return static_cast<T*>(aligned_alloc(alignment, count * sizeof(T)));
-        #endif
-    }
-
-    /// <summary>
-    /// checks if v is in arr
-    /// </summary>
-    /// <typeparam name="T"></typeparam>
-    /// <typeparam name="N"></typeparam>
-    /// <param name="arr"></param>
-    /// <param name="v"></param>
-    /// <returns></returns>
-    template<typename T, size_t N>
-    constexpr bool contains(const std::array<T, N>& arr, const T& v) {
-        for (const auto& vv : arr) {
-            if (v == vv) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-}  // namespace aerobus
-
-// concepts
-namespace aerobus {
-    /// Concept to express R is a Ring (ordered)
-    template <typename R>
-    concept IsRing = requires {
-        typename R::one;
-        typename R::zero;
-        typename R::template add_t<typename R::one, typename R::one>;
-        typename R::template sub_t<typename R::one, typename R::one>;
-        typename R::template mul_t<typename R::one, typename R::one>;
-    };
-
-    /// Concept to express R is an euclidean domain
-    template <typename R>
-    concept IsEuclideanDomain = IsRing<R> && requires {
-        typename R::template div_t<typename R::one, typename R::one>;
-        typename R::template mod_t<typename R::one, typename R::one>;
-        typename R::template gcd_t<typename R::one, typename R::one>;
-        typename R::template eq_t<typename R::one, typename R::one>;
-        typename R::template pos_t<typename R::one>;
-
-        R::template pos_v<typename R::one> == true;
-        // typename R::template gt_t<typename R::one, typename R::zero>;
-        R::is_euclidean_domain == true;
-    };
-
-    /// Concept to express R is a field
-    template<typename R>
-    concept IsField = IsEuclideanDomain<R> && requires {
-        R::is_field == true;
-    };
-}  // namespace aerobus
-
-// utilities
-namespace aerobus {
-    namespace internal {
-        template<template<typename...> typename TT, typename T>
-        struct is_instantiation_of : std::false_type { };
-
-        template<template<typename...> typename TT, typename... Ts>
-        struct is_instantiation_of<TT, TT<Ts...>> : std::true_type { };
-
-        template<template<typename...> typename TT, typename T>
-        inline constexpr bool is_instantiation_of_v = is_instantiation_of<TT, T>::value;
-
-        template <int64_t i, typename T, typename... Ts>
-        struct type_at {
[…]
