### 2024-08-31 #6233 @JeWaVe
usage / tests / bench instructions updated -- should be oK now but didn't test on windows yet (I updated the cmakelists.txt)
