### 2024-09-03 a8369aa
more documentation - more boolean properties (helpers to save some <tynemame>_t::value - some more testing
