+++ src/aerobus.h
@@ -332,7 +332,7 @@ namespace aerobus {
             using head = T;
         };
 
-        template <uint64_t index, typename L1, typename L2>
+        template <size_t index, typename L1, typename L2>
         struct split_h {
          private:
             static_assert(index <= L2::length, "index ouf of bounds");
@@ -351,7 +351,7 @@ namespace aerobus {
             using tail = L2;
         };
 
-        template <uint64_t index, typename L, typename T>
+        template <size_t index, typename L, typename T>
         struct insert_h {
             static_assert(index <= L::length, "index ouf of bounds");
             using s = typename L::template split<index>;
@@ -361,7 +361,7 @@ namespace aerobus {
             using type = typename ll::template concat<right>;
         };
 
-        template <uint64_t index, typename L>
+        template <size_t index, typename L>
         struct remove_h {
             using s = typename L::template split<index>;
             using left = typename s::head;
@@ -371,6 +371,9 @@ namespace aerobus {
         };
     }  // namespace internal
 
+
+    /// @brief A list of types
+    /// @tparam ...Ts
     template <typename... Ts>
     struct type_list {
      private:
@@ -383,26 +386,40 @@ namespace aerobus {
         };
 
      public:
+        /// @brief length of list
         static constexpr size_t length = sizeof...(Ts);
 
+        /// @brief Adds T to front of the list
+        /// @tparam T
         template <typename T>
         using push_front = type_list<T, Ts...>;
 
-        template <uint64_t index>
+        /// @brief returns type at index
+        /// @tparam index
+        template <size_t index>
         using at = internal::type_at_t<index, Ts...>;
 
+        /// @brief removes types from head of the list
         struct pop_front {
+            /// @brief type that was previously head of the list
             using type = typename internal::pop_front_h<Ts...>::head;
+            /// @brief remaining types in parent list when front is removed
             using tail = typename internal::pop_front_h<Ts...>::tail;
         };
 
+        /// @brief pushes T at the tail of the list
+        /// @tparam T
         template <typename T>
         using push_back = type_list<Ts..., T>;
 
+        /// @brief concatenates two list into one
+        /// @tparam U
         template <typename U>
         using concat = typename concat_h<U>::type;
 
-        template <uint64_t index>
+        /// @brief splits list at index
+        /// @tparam index
+        template <size_t index>
         struct split {
          private:
             using inner = internal::split_h<index, type_list<>, type_list<Ts...>>;
@@ -412,10 +429,15 @@ namespace aerobus {
             using tail = typename inner::tail;
         };
 
-        template <uint64_t index, typename T>
+        /// @brief inserts type at index
+        /// @tparam index
+        /// @tparam T
+        template <typename T, size_t index>
         using insert = typename internal::insert_h<index, type_list<Ts...>, T>::type;
 
-        template <uint64_t index>
+        /// @brief removes type at index
+        /// @tparam index
+        template <size_t index>
         using remove = typename internal::remove_h<index, type_list<Ts...>>::type;
     };
 
@@ -433,7 +455,7 @@ namespace aerobus {
         using concat = U;
 
         // TODO(jewave): assert index == 0
-        template <uint64_t index, typename T>
+        template <typename T, size_t index>
         using insert = type_list<T>;
     };
 }  // namespace aerobus
@@ -443,10 +465,11 @@ namespace aerobus {
     /// @brief 32 bits signed integers, seen as a algebraic ring with related operations
[…]
