### 2024-09-04 #6233 @JeWaVe
I'm extending and cleaning the documentation, sorry it's a long process
