### 2024-09-04 bf8490d
slightly more doc + integrate known polynomials from bigint-branch
