+++ src/aerobus.h
@@ -708,65 +708,84 @@ namespace aerobus {
 
      public:
         /// @brief addition operator
+        /// @tparam v1 : an element of aerobus::i64::val
+        /// @tparam v2 : an element of aerobus::i64::val
         template<typename v1, typename v2>
         using add_t = typename add<v1, v2>::type;
 
         /// @brief substraction operator
+        /// @tparam v1 : an element of aerobus::i64::val
+        /// @tparam v2 : an element of aerobus::i64::val
         template<typename v1, typename v2>
         using sub_t = typename sub<v1, v2>::type;
 
         /// @brief multiplication operator
+        /// @tparam v1 : an element of aerobus::i64::val
+        /// @tparam v2 : an element of aerobus::i64::val
         template<typename v1, typename v2>
         using mul_t = typename mul<v1, v2>::type;
 
         /// @brief division operator
+        /// @tparam v1 : an element of aerobus::i64::val
+        /// @tparam v2 : an element of aerobus::i64::val
         template<typename v1, typename v2>
         using div_t = typename div<v1, v2>::type;
 
         /// @brief modulus operator
+        /// @tparam v1 : an element of aerobus::i64::val
+        /// @tparam v2 : an element of aerobus::i64::val
         template<typename v1, typename v2>
         using mod_t = typename remainder<v1, v2>::type;
 
         /// @brief strictly greater operator (v1 > v2) - type
+        /// @tparam v1 : an element of aerobus::i64::val
+        /// @tparam v2 : an element of aerobus::i64::val
         template<typename v1, typename v2>
         using gt_t = typename gt<v1, v2>::type;
 
         /// @brief strictly greater operator (v1 > v2) - boolean value
-        /// @tparam v1
-        /// @tparam v2
+        /// @tparam v1 : an element of aerobus::i64::val
+        /// @tparam v2 : an element of aerobus::i64::val
         template<typename v1, typename v2>
         static constexpr bool gt_v = gt_t<v1, v2>::value;
 
         /// @brief strict less operator (v1 < v2)
+        /// @tparam v1 : an element of aerobus::i64::val
+        /// @tparam v2 : an element of aerobus::i64::val
         template<typename v1, typename v2>
         using lt_t = typename lt<v1, v2>::type;
 
         /// @brief strictly smaller operator (v1 < v2) - boolean value
-        /// @tparam v1
-        /// @tparam v2
+        /// @tparam v1 : an element of aerobus::i64::val
+        /// @tparam v2 : an element of aerobus::i64::val
         template<typename v1, typename v2>
         static constexpr bool lt_v = lt_t<v1, v2>::value;
 
         /// @brief equality operator (type)
+        /// @tparam v1 : an element of aerobus::i64::val
+        /// @tparam v2 : an element of aerobus::i64::val
         template<typename v1, typename v2>
         using eq_t = typename eq<v1, v2>::type;
 
         /// @brief equality operator (boolean value)
-        /// @tparam v1
-        /// @tparam v2
+        /// @tparam v1 : an element of aerobus::i64::val
+        /// @tparam v2 : an element of aerobus::i64::val
         template<typename v1, typename v2>
         static constexpr bool eq_v = eq_t<v1, v2>::value;
 
         /// @brief greatest common divisor
+        /// @tparam v1 : an element of aerobus::i64::val
+        /// @tparam v2 : an element of aerobus::i64::val
         template<typename v1, typename v2>
         using gcd_t = gcd_t<i64, v1, v2>;
 
         /// @brief is v posititive (type)
+        /// @tparam v1 : an element of aerobus::i64::val
         template<typename v>
         using pos_t = typename pos<v>::type;
 
         /// @brief positivity (boolean value)
-        /// @tparam v
+        /// @tparam v : an element of aerobus::i64::val
         template<typename v>
[…]
