+++ .github/workflows/c-cpp.yml
@@ -17,7 +17,7 @@ jobs:
       run: sudo apt-get install -y clang g++ libomp-dev cppcheck cpplint
     - name: cppcheck
       if: success()
-      run: cppcheck --std=c++20 --language=c++ src/lib.h && cppcheck --std=c++20 --language=c++ imports/conwaypolynomials.h && cppcheck --std=c++20 --language=c++ src/lib_tests.cpp
+      run: cppcheck --std=c++20 --language=c++ src/aerobus.h && cppcheck --std=c++20 --language=c++ src/lib_tests.cpp
     - name: cpplint
       if: success()
       run: cpplint src/lib.h && cpplint src/lib_tests.cpp
