### 2024-09-04 bc10860
add stirling numbers (signed - first kind) and optimize pow
