### 2024-09-05 f9d7aec
add absolute value, more stirling values ++ fix compilation with clang / msvc
