+++ src/aerobus.h
@@ -268,6 +268,12 @@ namespace aerobus {
     /// @tparam T Ring type (must be euclidean domain)
     template<typename T, typename A, typename B>
     using gcd_t = typename internal::gcd<T>::template type<A, B>;
+
+    template<typename val>
+    requires IsEuclideanDomain<typename val::ring_type>
+    using abs_t = std::conditional_t<
+                    val::ring_type::template pos_v<val>,
+                    val, typename val::ring_type::template sub_t<typename val::ring_type::zero, val>>;
 }  // namespace aerobus
 
 // quotient ring by the principal ideal generated by X
@@ -469,6 +475,8 @@ namespace aerobus {
         /// @tparam x an actual integer
         template<int32_t x>
         struct val {
+            /// @brief Enclosing ring type
+            using ring_type = i32;
             /// @brief actual value stored in val type
             static constexpr int32_t v = x;
 
@@ -614,11 +622,15 @@ namespace aerobus {
 namespace aerobus {
     /// @brief 64 bits signed integers, seen as a algebraic ring with related operations
     struct i64 {
+        /// @brief type for actual values
         using inner_type = int64_t;
         /// @brief values in i64
         /// @tparam x an actual integer
         template<int64_t x>
         struct val {
+            /// @brief enclosing ring type
+            using ring_type = i64;
+            /// @brief actual value
             static constexpr int64_t v = x;
 
             /// @brief cast value in valueType
@@ -802,6 +814,9 @@ namespace aerobus {
         using inner_type = int32_t;
         template<int32_t x>
         struct val {
+            /// @brief enclosing ring type
+            using ring_type = zpz<p>;
+            /// @brief actual value
             static constexpr int32_t v = x % p;
 
             template<typename valueType>
@@ -938,7 +953,7 @@ namespace aerobus {
      * polynomial with coefficients in Ring
      * Ring must be an integral domain
     */
-    template<typename Ring, char variable_name = 'x'>
+    template<typename Ring>
     requires IsEuclideanDomain<Ring>
     struct polynomial {
         static constexpr bool is_field = false;
@@ -949,6 +964,8 @@ namespace aerobus {
         /// @tparam ...coeffs lower degree coefficients
         template<typename coeffN, typename... coeffs>
         struct val {
+            /// @brief enclosing ring type
+            using ring_type = polynomial<Ring>;
             /// @brief degree of the polynomial
             static constexpr size_t degree = sizeof...(coeffs);
             /// @brief heavy weight coefficient (non zero)
@@ -1002,6 +1019,9 @@ namespace aerobus {
         /// @tparam coeffN
         template<typename coeffN>
         struct val<coeffN> {
+            /// @brief enclosing ring type
+            using ring_type = polynomial<Ring>;
+            /// @brief degree
             static constexpr size_t degree = 0;
             using aN = coeffN;
             using strip = val<coeffN>;
@@ -1371,17 +1391,16 @@ namespace aerobus {
                     return tail;
                 } else if (Ring::template eq_t<coeff, typename Ring::one>::value) {
                     if (sizeof...(coeffs) == 1) {
-                        result += std::string(1, variable_name);
+                        result += "x";
                     } else {
-                        result += std::string(1, variable_name) + "^" + std::to_string(sizeof...(coeffs));
+                        result += "x^" + std::to_string(sizeof...(coeffs));
                     }
                 } else {
                     if (sizeof...(coeffs) == 1) {
-                        result += coeff::to_string() + " " + std::string(1, variable_name);
+                        result += coeff::to_string() + " x";
                     } else {
[…]
