### 2024-09-05 ec5085b
run tests with both g++ and clang
