### 2024-09-05 57080ac
fix CI (missing typename for clang) + a bit more documentation
