+++ src/aerobus.h
@@ -99,50 +99,55 @@ namespace aerobus {
         using type_at_t = typename type_at<i, Ts...>::type;
 
 
-        template<int32_t n, int32_t i, typename E = void>
+        template<size_t n, size_t i, typename E = void>
         struct _is_prime {};
 
-        template<int32_t i>
+        template<size_t i>
+        struct _is_prime<0, i> {
+            static constexpr bool value = false;
+        };
+
+        template<size_t i>
         struct _is_prime<1, i> {
             static constexpr bool value = false;
         };
 
-        template<int32_t i>
+        template<size_t i>
         struct _is_prime<2, i> {
             static constexpr bool value = true;
         };
 
-        template<int32_t i>
+        template<size_t i>
         struct _is_prime<3, i> {
             static constexpr bool value = true;
         };
 
-        template<int32_t i>
+        template<size_t i>
         struct _is_prime<5, i> {
             static constexpr bool value = true;
         };
 
-        template<int32_t i>
+        template<size_t i>
         struct _is_prime<7, i> {
             static constexpr bool value = true;
         };
 
-        template<int32_t n, int32_t i>
+        template<size_t n, size_t i>
         struct _is_prime<n, i, std::enable_if_t<(n != 2 && n % 2 == 0)>> {
             static constexpr bool value = false;
         };
 
-        template<int32_t n, int32_t i>
+        template<size_t n, size_t i>
         struct _is_prime<n, i, std::enable_if_t<(n != 2 && n != 3 && n % 2 != 0 && n % 3 == 0)>> {
             static constexpr bool value = false;
         };
 
-        template<int32_t n, int32_t i>
+        template<size_t n, size_t i>
         struct _is_prime<n, i, std::enable_if_t<(n >= 9 && i * i > n)>> {
             static constexpr bool value = true;
         };
 
-        template<int32_t n, int32_t i>
+        template<size_t n, size_t i>
         struct _is_prime<n, i, std::enable_if_t<(
             n % i == 0 &&
             n >= 9 &&
@@ -152,7 +157,7 @@ namespace aerobus {
             static constexpr bool value = true;
         };
 
-        template<int32_t n, int32_t i>
+        template<size_t n, size_t i>
         struct _is_prime<n, i, std::enable_if_t<(
             n % (i+2) == 0 &&
             n >= 9 &&
@@ -162,7 +167,7 @@ namespace aerobus {
             static constexpr bool value = true;
         };
 
-        template<int32_t n, int32_t i>
+        template<size_t n, size_t i>
         struct _is_prime<n, i, std::enable_if_t<(
                 n % (i+2) != 0 &&
                 n % i != 0 &&
@@ -177,13 +182,16 @@ namespace aerobus {
 
     /// @brief checks if n is prime
     /// @tparam n
-    template<int32_t n>
+    template<size_t n>
     struct is_prime {
         /// @brief true iff n is prime
         static constexpr bool value = internal::_is_prime<n, 5>::value;
     };
 
-    template<int32_t n>
+    /// @brief checks if n is prime
+    /// true if and only if is n is prime
+    /// @tparam n
+    template<size_t n>
     static constexpr bool is_prime_v = is_prime<n>::value;
 
     namespace internal {
@@ -195,7 +203,7 @@ namespace aerobus {
         using make_index_sequence_reverse
             = decltype(index_sequence_reverse(std::make_index_sequence<N>{}));
 
-        /**
+        /** @brief greatest common divisor
          * computes the greatest common divisor
          * exposes it in gcd<A, B>::type
          * as long as Ring type is an integral domain
@@ -249,6 +257,9 @@ namespace aerobus {
     template<typename T, typename A, typename B>
     using gcd_t = typename internal::gcd<T>::template type<A, B>;
 
[…]
