+++ src/aerobus.h
@@ -22,6 +22,9 @@
 #define INLINED __attribute__((always_inline)) inline
 #endif
 
+/// \file
+/// aerobus.h
+
 // aligned allocation
 namespace aerobus {
     /**
@@ -341,13 +344,13 @@ namespace aerobus {
         static constexpr bool is_euclidean_domain = true;
 
         /// @brief inject a 'constant' in quotient ring
-        /// for example : QuotientRing<i32, i32::val<2>>::inject_constant_t<1>
+        /// @example QuotientRing<i32, i32::val<2>>::inject_constant_t<1>
         /// @tparam x a 'constant' from Ring point of view
         template<auto x>
         using inject_constant_t = val<typename Ring::template inject_constant_t<x>>;
 
         /// @brief projects a value of Ring onto the quotient
-        /// for example : QuotientRing<i32, i32::val<2>>::inject_ring_t<i32::val<1>>
+        /// @example QuotientRing<i32, i32::val<2>>::inject_ring_t<i32::val<1>>
         /// @tparam v a value in Ring
         template<typename v>
         using inject_ring_t = val<v>;
@@ -407,7 +410,7 @@ namespace aerobus {
     }  // namespace internal
 
     /// @brief A list of types
-    /// for example type_list<int, double, float>
+    /// @example type_list<int, double, float>
     /// @tparam ...Ts types to store and manipulate at compile time
     template <typename... Ts>
     struct type_list {
@@ -596,53 +599,95 @@ namespace aerobus {
 
      public:
         /// @brief addition operator
+        /// yields v1 + v2
+        /// @example i32::add_t<i32::val<2>, i32::val<3>>
+        /// @tparam v1 a value in i32
+        /// @tparam v2 a value in i32
         template<typename v1, typename v2>
         using add_t = typename add<v1, v2>::type;
 
         /// @brief substraction operator
+        /// yields v1 - v2
+        /// @example i32::sub_t<i32::val<3>, i32::val<2>>
+        /// @tparam v1 a value in i32
+        /// @tparam v2 a value in i32
         template<typename v1, typename v2>
         using sub_t = typename sub<v1, v2>::type;
 
         /// @brief multiplication operator
+        /// yields v1 * v2
+        /// @example i32::mul_t<i32::val<3>, i32::val<2>>
+        /// @tparam v1 a value in i32
+        /// @tparam v2 a value in i32
         template<typename v1, typename v2>
         using mul_t = typename mul<v1, v2>::type;
 
         /// @brief division operator
+        /// yields v1 / v2
+        /// @example i32::div_t<i32::val<7>, i32::val<2>> -> i32::val<3>
+        /// @tparam v1 a value in i32
+        /// @tparam v2 a value in i32
         template<typename v1, typename v2>
         using div_t = typename div<v1, v2>::type;
 
         /// @brief modulus operator
+        /// yields v1 % v2
+        /// for example : i32::mod_t<i32::val<7>, i32::val<2>>
+        /// @tparam v1 a value in i32
+        /// @tparam v2 a value in i32
         template<typename v1, typename v2>
         using mod_t = typename remainder<v1, v2>::type;
 
         /// @brief strictly greater operator (v1 > v2)
+        /// yields v1 > v2
+        /// @example i32::gt_t<i32::val<7>, i32::val<2>>
+        /// @tparam v1 a value in i32
+        /// @tparam v2 a value in i32
         template<typename v1, typename v2>
         using gt_t = typename gt<v1, v2>::type;
 
         /// @brief strict less operator (v1 < v2)
+        /// yields v1 < v2
+        /// @example i32::div_t<i32::val<7>, i32::val<2>>
+        /// @tparam v1 a value in i32
[…]
