### 2024-09-09 65ae0b7
add generic vadd and vmul types
