+++ src/aerobus.h
@@ -197,6 +197,7 @@ namespace aerobus {
     template<size_t n>
     static constexpr bool is_prime_v = is_prime<n>::value;
 
+    // gcd
     namespace internal {
         template <std::size_t... Is>
         constexpr auto index_sequence_reverse(std::index_sequence<Is...> const&)
@@ -255,19 +256,60 @@ namespace aerobus {
         };
     }  // namespace internal
 
+    // vadd and vmul
+    namespace internal {
+        template<typename... vals>
+        struct vmul {};
+
+        template<typename v1, typename... vals>
+        struct vmul<v1, vals...> {
+            using type = typename v1::enclosing_type::template mul_t<v1, typename vmul<vals...>::type>;
+        };
+
+        template<typename v1>
+        struct vmul<v1> {
+            using type = v1;
+        };
+
+        template<typename... vals>
+        struct vadd {};
+
+        template<typename v1, typename... vals>
+        struct vadd<v1, vals...> {
+            using type = typename v1::enclosing_type::template add_t<v1, typename vadd<vals...>::type>;
+        };
+
+        template<typename v1>
+        struct vadd<v1> {
+            using type = v1;
+        };
+    }  // namespace internal
+
     /// @brief computes the greatest common divisor or A and B
     /// @tparam T Ring type (must be euclidean domain)
     template<typename T, typename A, typename B>
     using gcd_t = typename internal::gcd<T>::template type<A, B>;
 
+    /// @brief adds multiple values (v1 + v2 + ... + vn)
+    /// vals must have same "enclosing_type" and "enclosing_type" must have an add_t binary operator
+    /// @tparam ...vals
+    template<typename... vals>
+    using vadd_t = typename internal::vadd<vals...>::type;
+
+    /// @brief multiplies multiple values (v1 + v2 + ... + vn)
+    /// vals must have same "enclosing_type" and "enclosing_type" must have an mul_t binary operator
+    /// @tparam ...vals
+    template<typename... vals>
+    using vmul_t = typename internal::vmul<vals...>::type;
+
     /// @brief computes absolute value of 'val'
     /// val must be a 'value' in a Ring satisfying 'IsEuclideanDomain' concept
     /// @tparam val a value in a RIng, such as i64::val<-2>
     template<typename val>
-    requires IsEuclideanDomain<typename val::ring_type>
+    requires IsEuclideanDomain<typename val::enclosing_type>
     using abs_t = std::conditional_t<
-                    val::ring_type::template pos_v<val>,
-                    val, typename val::ring_type::template sub_t<typename val::ring_type::zero, val>>;
+                    val::enclosing_type::template pos_v<val>,
+                    val, typename val::enclosing_type::template sub_t<typename val::enclosing_type::zero, val>>;
 }  // namespace aerobus
 
 namespace aerobus {
@@ -509,7 +551,7 @@ namespace aerobus {
         template<int32_t x>
         struct val {
             /// @brief Enclosing ring type
-            using ring_type = i32;
+            using enclosing_type = i32;
             /// @brief actual value stored in val type
             static constexpr int32_t v = x;
 
@@ -704,7 +746,7 @@ namespace aerobus {
         template<int64_t x>
         struct val {
             /// @brief enclosing ring type
-            using ring_type = i64;
+            using enclosing_type = i64;
             /// @brief actual value
             static constexpr int64_t v = x;
 
@@ -917,7 +959,7 @@ namespace aerobus {
         template<int32_t x>
         struct val {
             /// @brief enclosing ring type
-            using ring_type = zpz<p>;
+            using enclosing_type = zpz<p>;
             /// @brief actual value
             static constexpr int32_t v = x % p;
 
@@ -1093,7 +1135,7 @@ namespace aerobus {
         template<typename coeffN, typename... coeffs>
         struct val {
[…]
