### 2024-09-09 cdb5720
add bernoulli and legendre polynomials
