+++ src/aerobus.h
@@ -2088,6 +2088,20 @@ namespace aerobus {
     template<typename Ring, typename v1, typename v2>
     using makefraction_t = typename FractionField<Ring>::template val<v1, v2>;
 
+    /// @brief helper type : make a fraction from numerator and denominator
+    /// @tparam p numerator
+    /// @tparam q denominator
+    template<int64_t p, int64_t q>
+    using make_q64_t = typename q64::template simplify_t<
+                typename q64::val<i64::inject_constant_t<p>, i64::inject_constant_t<q>>>;
+
+    /// @brief helper type : make a fraction from numerator and denominator
+    /// @tparam p numerator
+    /// @tparam q denominator
+    template<int32_t p, int32_t q>
+    using make_q32_t = typename q32::template simplify_t<
+                typename q32::val<i32::inject_constant_t<p>, i32::inject_constant_t<q>>>;
+
     /// @brief helper type : adds two fractions
     /// @tparam Ring
     /// @tparam v1 belongs to FractionField<Ring>
@@ -2102,7 +2116,7 @@ namespace aerobus {
     using mulfractions_t = typename FractionField<Ring>::template mul_t<v1, v2>;
 }  // namespace aerobus
 
-// taylor series and common integers (factorial, bernouilli...) appearing in taylor coefficients
+// taylor series and common integers (factorial, bernoulli...) appearing in taylor coefficients
 namespace aerobus {
     namespace internal {
         template<typename T, size_t x, typename E = void>
@@ -2181,11 +2195,11 @@ namespace aerobus {
 
     namespace internal {
         template<typename T, size_t m>
-        struct bernouilli;
+        struct bernoulli;
 
         template<typename T, typename accum, size_t k, size_t m>
-        struct bernouilli_helper {
-            using type = typename bernouilli_helper<
+        struct bernoulli_helper {
+            using type = typename bernoulli_helper<
                 T,
                 addfractions_t<T,
                     accum,
@@ -2193,7 +2207,7 @@ namespace aerobus {
                         makefraction_t<T,
                             combination_t<T, k, m + 1>,
                             typename T::one>,
-                        typename bernouilli<T, k>::type
+                        typename bernoulli<T, k>::type
                     >
                 >,
                 k + 1,
@@ -2201,16 +2215,16 @@ namespace aerobus {
         };
 
         template<typename T, typename accum, size_t m>
-        struct bernouilli_helper<T, accum, m, m> {
+        struct bernoulli_helper<T, accum, m, m> {
             using type = accum;
         };
 
 
 
         template<typename T, size_t m>
-        struct bernouilli {
+        struct bernoulli {
             using type = typename FractionField<T>::template mul_t<
-                typename internal::bernouilli_helper<T, typename FractionField<T>::zero, 0, m>::type,
+                typename internal::bernoulli_helper<T, typename FractionField<T>::zero, 0, m>::type,
                 makefraction_t<T,
                 typename T::template val<static_cast<typename T::inner_type>(-1)>,
                 typename T::template val<static_cast<typename T::inner_type>(m + 1)>
@@ -2222,7 +2236,7 @@ namespace aerobus {
         };
 
         template<typename T>
-        struct bernouilli<T, 0> {
+        struct bernoulli<T, 0> {
             using type = typename FractionField<T>::one;
 
             template<typename floatType>
@@ -2230,18 +2244,18 @@ namespace aerobus {
         };
     }  // namespace internal
 
-    /// @brief nth Bernouilli number as type in T
+    /// @brief nth bernoulli number as type in T
     /// @tparam T Ring type (i64)
     /// @tparam n
     template<typename T, size_t n>
-    using bernouilli_t = typename internal::bernouilli<T, n>::type;
[…]
