### 2024-09-10 c160af2
more documentation + fix atanh (typo in coefficient)
