+++ .vscode/ltex.hiddenFalsePositives.en-US.txt
@@ -1,0 +1,3 @@
+{"rule":"NON_STANDARD_WORD","sentence":"^\\Qadd_t : a type (specialization of val), representing addition between two values\\E$"}
+{"rule":"MORFOLOGIK_RULE_EN_US","sentence":"^\\Qadd_t : a type (specialization of val), representing addition between two values\nsub_t : a type (specialization of val), representing subtraction between two values\nmul_t : a type (specialization of val), representing multiplication between two values\ndiv_t : a type (specialization of val), representing division between two values\nadd_t : a type (specialization of val), representing addition between two values\\E$"}
+{"rule":"NON_STANDARD_WORD","sentence":"^\\Qadd_t : a type (specialization of val), representing addition between two values\nsub_t : a type (specialization of val), representing subtraction between two values\nmul_t : a type (specialization of val), representing multiplication between two values\ndiv_t : a type (specialization of val), representing division between two values\nadd_t : a type (specialization of val), representing addition between two values\\E$"}
