+++ paper.md
@@ -18,12 +18,12 @@ date: 12 October 2023
 bibliography: paper.bib
 ---
 
-# Summary
+## Summary
 
 C++ comes with high compile-time computations capability, also known as metaprogramming with templates.
-Templates are a language-in-the-language which is Turing-complete, meaning we can run every computation at compile time instead of runtime, as long as input data is known at compile time. 
+Templates are a language-in-the-language which is Turing-complete, meaning we can run every computation at compile time instead of runtime, as long as input data is known at compile time.
 
-Using these capabilities, vastly extended with the latest versions of the standard, we implemented a library for discrete Euclidean domains, such as $\mathbb{Z}$. We also provide a way to generate the fraction field of such rings (e.g. $\mathbb{Q}$). 
+Using these capabilities, vastly extended with the latest versions of the standard, we implemented a library for discrete Euclidean domains, such as $\mathbb{Z}$. We also provide a way to generate the fraction field of such rings (e.g. $\mathbb{Q}$).
 
 We also implemented polynomials over such discrete rings and fields (e.g. $\mathbb{Q}[X]$). Since polynomials are also a ring, the above implementation gives us rational fractions as the field of fractions of polynomials.
 
@@ -33,37 +33,39 @@ In addition, we added some useful additional features, such as known polynomials
 
 `Aerobus` was designed to be used in high-performance software, teaching purposes or embedded software where as much as possible must be precomputed to shrink binary size. It compiles with major compilers: gcc, clang and msvc. It is quite easily configurable and extensible. 
 
-# Statement of need
-By implementing general algebra concepts such as discrete rings, field of fractions and polynomials, `Aerobus` can serve multiple purposes. 
+## Statement of need
+By implementing general algebra concepts such as discrete rings, field of fractions and polynomials, `Aerobus` can serve multiple purposes.
 
 The main application we want to express in this paper is the automatic (and configurable) generation or Taylor approximation of usual transcendental functions such as `exp` or `sin`. The "generated" code is pure C++ and can be inspected.
 
 These functions are usually exposed by the standard library (`<cmath>`) with high (guaranteed) precision. However, in high-performance computing, when not compiled with `-Ofast`, evaluating `std::exp` has several flaws: 
 
-- it leads to a `syscall` which is very expensive
-- it doesn't leverage vector units (avx, avx2, avx512 or equivalent in non-intel hardware). 
+- It leads to a `syscall` which is very expensive;
+- It doesn't leverage vector units (AVX, AVX2, AVX512 or equivalent in non-intel hardware).
 
-Hardware vendors provide high-performance libraries such as [@wang2014intel], but implementation is often hidden and not extensible. 
+Hardware vendors provide high-performance libraries such as [@wang2014intel], but implementation is often hidden and not extensible.
 
-Some others can provide vectorized functions, such as @vml does. But libraries like VML are highly tight to one architecture by their use of intrinsics or inline assembly. In addition, they only provide a restricted list of math functions and do not expose capabilities to generate high-performance versions of other functions such as arctanh. It is the same for the standard library compiled with `-Ofast`: it generates a vectorized version of some functions (such as exp) but with no control of precision and no extensibility. 
[…]
