+++ main_page.md
@@ -12,7 +12,8 @@ The library serves two main purposes :
 
 - Express algebra structures and associated operations in type arithmetic, compile-time;
 - Provide portable and fast evaluation functions for polynomials.
-It is designed to be 'easily' extensible.
+
+It is designed to be 'quite easily' extensible.
 
 Given these functions are "generated" at compile time and do not rely on inline assembly, they are actually platform independent, yielding exact same results if processors have same capabilities (such as Fused-Multiply-Add instructions).
 
@@ -61,6 +62,17 @@ using B23 = aerobus::known_polynomials::bernstein<2, 3>; // 3X^2(1-X)
 constexpr float x = B32::eval(2.0F); // -12
 ```
 
+Complete list is:
+
+- chebyshev_T
+- chebyshev_U
+- laguerre
+- hermite_prob
+- hermite_phys
+- bernstein
+- legendre
+- bernoulli
+
 ### Conway polynomials
 
 When the tag `AEROBUS_CONWAY_IMPORTS` is defined at compile time (`-DAEROBUS_CONWAY_IMPORTS`), `aerobus` provides definition for all Conway polynomials `CP(p, n)` for `p` up to 997 and low values for `n` (usually less than 10).
@@ -85,6 +97,24 @@ using aero_atanh = atanh<i64, 6>;
 constexpr float val = aero_atanh::eval(0.1F); // approximation of arctanh(0.1) using taylor expansion of degree 6
 ```
 
+Exposed functions are:
+
+- exp
+- expm1 \f$e^x - 1\f$
+- lnp1 (\f$\ln(x+1)\f$)
+- geom (\f$\frac{1}{1-x}\f$)
+- sin
+- cos
+- tan
+- sh
+- cosh
+- tanh
+- asin
+- acos
+- acosh
+- asinh
+- atanh
+
 Having the capacity of specifying the degree is very important, as users may use other formats than `float64` or `float32` which require higher or lower degree to achieve correct or acceptable precision.
 
 It's possible to define Taylor expansion by implementing a `coeff_at` structure which must meet the following requirement :
@@ -104,8 +134,116 @@ template<typename Integers, typename degree>
 using my_serie = taylor<Integers, my_coeff_at, degree>;
 ```
 
+On x86-64 and CUDA platforms at least, using proper compiler directives, these functions yield very performant assembly, similar or better than standard library implementation in fast math. For example, this code:
+
+```cpp
+double compute_expm1(const size_t N, double* in, double* out) {
+    using V = aerobus::expm1<aerobus::i64, 13>;
+    for (size_t i = 0; i < N; ++i) {
+        out[i] = V::eval(in[i]);
+    }
+}
+```
+
+Yields this assembly (clang 17, `-ffast-math -mavx512 -O3`):
+
+```nasm
+compute_expm1(unsigned long, double const*, double*):
+  lea     rax, [rdi-1]
+  cmp     rax, 2
+  jbe     .L5
+  mov     rcx, rdi
+  xor     eax, eax
+  vxorpd  xmm1, xmm1, xmm1
+  vbroadcastsd    ymm14, QWORD PTR .LC1[rip]
+  vbroadcastsd    ymm13, QWORD PTR .LC3[rip]
+  shr     rcx, 2
+  vbroadcastsd    ymm12, QWORD PTR .LC5[rip]
+  vbroadcastsd    ymm11, QWORD PTR .LC7[rip]
+  sal     rcx, 5
[…]
