+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,33 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+1. Install aerobus as described in README.md
+2. write this code
+3. compile and run
+4. see result
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**Desktop (please complete the following information):**
+ - OS: [e.g. linux]
+ - Platform [x86 / ARM / ...]
+ - Compiler [c++, clang++, msvc, icc, ...]
+ - Versions of OS / compiler [e.g. clang18]
+
+**Additional context**
+Add any other context about the problem here.
