+++ .github/pull_request_template.md
@@ -1,0 +1,10 @@
+## Brief description of changes
+Is that a bug fix, a new feature ? 
+
+## Defailed descriptions
+Describe what's new. 
+In case of bug fix, what was the buggy behavior and how it's now fixed (if relevant, specify compiler, OS, platform). 
+In case of new feature, explain what it does and how it will be used
+
+## Mentions
+mention reviewers and contributors
