### 2024-09-11 a964d9e
update paper.md + more doc
