+++ paper.md
@@ -27,18 +27,19 @@ Using these capabilities, vastly extended with the latest versions of the standa
 
 We also implemented polynomials over such discrete rings and fields (e.g. $\mathbb{Q}[X]$). Since polynomials are also a ring, the above implementation gives us rational fractions as the field of fractions of polynomials.
 
-In addition, we expose a way to generate the Taylor series of any math functions as long as coefficients are known. 
+In addition, we expose a way to generate the Taylor series of any math functions as long as coefficients are known.
 
 In addition, we added some useful additional features, such as known polynomials (Chebyshev), continued fractions, quotient rings and some Conway polynomials to define Galois finite fields.
 
-`Aerobus` was designed to be used in high-performance software, teaching purposes or embedded software where as much as possible must be precomputed to shrink binary size. It compiles with major compilers: gcc, clang and msvc. It is quite easily configurable and extensible. 
+`Aerobus` was designed to be used in high-performance software, teaching purposes or embedded software where as much as possible must be precomputed to shrink binary size. It compiles with major compilers: `gcc`, `clang` and `msvc`. It is quite easily configurable and extensible.
 
 ## Statement of need
+
 By implementing general algebra concepts such as discrete rings, field of fractions and polynomials, `Aerobus` can serve multiple purposes.
 
-The main application we want to express in this paper is the automatic (and configurable) generation or Taylor approximation of usual transcendental functions such as `exp` or `sin`. The "generated" code is pure C++ and can be inspected.
+The main application we want to express in this paper is the automatic (and configurable) generation or Taylor approximation of usual transcendental functions such as `exp` or `sin`.
 
-These functions are usually exposed by the standard library (`<cmath>`) with high (guaranteed) precision. However, in high-performance computing, when not compiled with `-Ofast`, evaluating `std::exp` has several flaws: 
+These functions are usually exposed by the standard library (`<cmath>`) with high (guaranteed) precision. However, in high-performance computing, when not compiled with `-Ofast`, evaluating `std::exp` has several flaws:
 
 - It leads to a `syscall` which is very expensive;
 - It doesn't leverage vector units (AVX, AVX2, AVX512 or equivalent in non-intel hardware).
@@ -59,7 +60,7 @@ Addition is commutative and associative and every element $x$ has an inverse $-x
 
 An `integral domain` is a ring with one additional property. For every elements $a, b, c$ such as $ab = ac$, then either $a = 0$ or $b = c$. Such a ring is not always a field, such as $\mathbb{Z}$ shows it.
 
-A `euclidean domain` is an integral domain that can be endowed with an euclidean division.
+A `euclidean domain` is an integral domain that can be endowed with a euclidean division.
 
 For such a euclidean domain, we can build two important structures:
 
@@ -69,22 +70,22 @@ Polynomials over $\mathbb{A}$ is the free module generated by a base noted $(X^k
 
 $$a_0 + a_1X + \ldots + a_nX^n$$
 
-where $a_n \neq 0$ if $n \neq 0$. 
+where $a_n \neq 0$ if $n \neq 0$.
 
[…]
