+++ .vscode/ltex.hiddenFalsePositives.en-US.txt
@@ -1,3 +1,6 @@
 {"rule":"NON_STANDARD_WORD","sentence":"^\\Qadd_t : a type (specialization of val), representing addition between two values\\E$"}
 {"rule":"MORFOLOGIK_RULE_EN_US","sentence":"^\\Qadd_t : a type (specialization of val), representing addition between two values\nsub_t : a type (specialization of val), representing subtraction between two values\nmul_t : a type (specialization of val), representing multiplication between two values\ndiv_t : a type (specialization of val), representing division between two values\nadd_t : a type (specialization of val), representing addition between two values\\E$"}
 {"rule":"NON_STANDARD_WORD","sentence":"^\\Qadd_t : a type (specialization of val), representing addition between two values\nsub_t : a type (specialization of val), representing subtraction between two values\nmul_t : a type (specialization of val), representing multiplication between two values\ndiv_t : a type (specialization of val), representing division between two values\nadd_t : a type (specialization of val), representing addition between two values\\E$"}
+{"rule":"WHERE_AS","sentence":"^\\Q\\E(?:Dummy|Ina|Jimmy-)[0-9]+\\Q was designed to be used in high-performance software, teaching purposes or embedded software where as much as possible must be precomputed to shrink binary size.\\E$"}
+{"rule":"EN_A_VS_AN","sentence":"^\\QAn \\E(?:Dummy|Ina|Jimmy-)[0-9]+\\Q is a ring with one additional property.\\E$"}
+{"rule":"EVERY_EACH_SINGULAR","sentence":"^\\QFor every elements \\E(?:Dummy|Ina|Jimmy-)[0-9]+\\Q such as \\E(?:Dummy|Ina|Jimmy-)[0-9]+\\Q, then either \\E(?:Dummy|Ina|Jimmy-)[0-9]+\\Q or \\E(?:Dummy|Ina|Jimmy-)[0-9]+\\Q.\\E$"}
