### 2024-09-11 f12ec8c
update README and documentation + make them compatible
