+++ main_page.md
@@ -1,261 +1,0 @@
-# Introduction
-
-`Aerobus` is a C++-20 pure header library for general algebra on polynomials, discrete rings and associated structures.
-
-Everything in `Aerobus` is expressed as types.
-
-We say that again as it is the most fundamental characteristic of `Aerobus` :
-
-***Everything is expressed as types***
-
-The library serves two main purposes :
-
-- Express algebra structures and associated operations in type arithmetic, compile-time;
-- Provide portable and fast evaluation functions for polynomials.
-
-It is designed to be 'quite easily' extensible.
-
-Given these functions are "generated" at compile time and do not rely on inline assembly, they are actually platform independent, yielding exact same results if processors have same capabilities (such as Fused-Multiply-Add instructions).
-
-## Structures
-
-### Predefined discrete euclidean domains
-
-`Aerobus` predefines several simple euclidean domains, such as :
-
-- `aerobus::i32` : integers (32 bits)
-- `aerobus::i64` : integers (64 bits)
-- `aerobus::zpz<p>` : integers modulo p (prime number) on 32 bits
-
-All these types represent the Ring, meaning the algebraic structure. They have a nested type `val<i>` where `i` is a scalar native value (int32_t or int64_t) to represent actual values in the ring.
-They have the following "operations", required by the IsEuclideanDomain concept :
-
-- `add_t` : a type (specialization of val), representing addition between two values
-- `sub_t` : a type (specialization of val), representing subtraction between two values
-- `mul_t` : a type (specialization of val), representing multiplication between two values
-- `div_t` : a type (specialization of val), representing division between two values
-- `mod_t` : a type (specialization of val), representing modulus between two values
-
-and the following "elements" :
-
-- one : the neutral element for multiplication, val<1>
-- zero : the neutral element for addition, val<0>
-
-### Polynomials
-
-`Aerobus` defines polynomials as a variadic template structure, with coefficient in an arbitrary discrete euclidean domain. As `i32` or `i64`, they are given same operations and elements, which make them a euclidean domain by themselves. Similarly, `aerobus::polynomial` represents the algebraic structure, actual values are in `aerobus::polynomial::val`.
-
-In addition, values have an evaluation function :
-
-```cpp
-template<typename valueRing> static constexpr valueRing eval(const valueRing& x) {...}
-```
-
-Which can be used at compile time (constexpr evaluation) or runtime.
-
-### Known polynomials
-
-`Aerobus` predefines some well known families of polynomials, such as Hermite or Bernstein :
-
-```cpp
-using B23 = aerobus::known_polynomials::bernstein<2, 3>; // 3X^2(1-X)
-constexpr float x = B32::eval(2.0F); // -12
-```
-
-They have their coefficients either in `aerobus::i64` or `aerobus::q64`.
-Complete list is (but is meant to be extended):
-
-- `chebyshev_T`
-- `chebyshev_U`
-- `laguerre`
-- `hermite_prob`
-- `hermite_phys`
-- `bernstein`
-- `legendre`
-- `bernoulli`
-
[…]
