+++ README.md
@@ -1,51 +1,63 @@
-# Aerobus
-Aerobus is a C++20 template library which expresses algebra concepts in types
+# Introduction
 
-It defines some concepts, such as Ring, IntegralDomain or Field which can be used to construct the field of fractions of a Ring, polynomials with coefficients in such a set or rational fractions, all at compile time. It allows the definition of polynomials over any discrete integral domain (and therefore the corresponding field of fractions), as long as your Ring implementation satisfies the IsIntegralDomain concept. 
+`Aerobus` is a C++-20 pure header library for general algebra on polynomials, discrete rings and associated structures.
 
-It defines Integers (32 or 64 bits) as types, therefore rationals (field of fractions) and modular arithmetic (Z/pZ). Polynomials with integer or rational coefficients are exposed as types, and so are rational fractions (field of fractions of polynomials). 
+Everything in `Aerobus` is expressed as types.
 
-As an interesting application, it provides predefined polynomials, such as the Taylor series of usual functions. Given polynomials can then be evaluated at floating point values, Aerobus therefore defines compile-time versions of usual functions and very efficient implementations for runtime.
+We say that again as it is the most fundamental characteristic of `Aerobus` :
 
-Unlike most competing libraries, it's quite easy to add a custom function as Aerobus provides mechanisms to easily define coefficients and Taylor series. 
+***Everything is expressed as types***
 
-Code is tested against MSVC, CLANG and GCC, see report [here](https://godbolt.org/z/qnfP99KWv)
+The library serves two main purposes :
 
+- Express algebra structures and associated operations in type arithmetic, compile-time;
+- Provide portable and fast evaluation functions for polynomials.
+
+It is designed to be 'quite easily' extensible.
+
+Given these functions are "generated" at compile time and do not rely on inline assembly, they are actually platform independent, yielding exact same results if processors have same capabilities (such as Fused-Multiply-Add instructions).
 
 ## HOW TO
+
 - Clone or download the repository somewhere, or just download the aerobus.h
 - In your code, add : `#include "aerobus.h"`
 - Compile with -std=c++20 (at least) -I<install_location>
   
-Aerobus provide a definition for low-degree (up to 997) conway polynomials. To use them, define AEROBUS_CONWAY_IMPORTS before including aerobus.h.
+`Aerobus` provides a definition for low-degree (up to 997) Conway polynomials. To use them, define `AEROBUS_CONWAY_IMPORTS` before including `aerobus.h`.
 
 ### Unit Test
-Install [Cmake](https://cmake.org/download/)
+
+Install [`Cmake`](https://cmake.org/download/)
 Install a recent compiler (supporting c++20), such as MSVC, G++ or Clang++
 
-Move to the top directory then : 
+Move to the top directory then :
+
 ```bash
 cmake -S . -B build
 cmake --build build
 cd build && ctest
 ```
 
[…]
