+++ .github/workflows/c-cpp.yml
@@ -26,13 +26,13 @@ jobs:
       run: make clean tests
     - name: documentation
       if: success()
-      run: rm -rd docs/ && make doc && mkdir docs && mv -v documentation/html/* docs/
+      run: rm -rf docs/ && make doc && mkdir docs && mv -v documentation/html/* docs/
   paper:
     runs-on: ubuntu-latest
     name: Paper Draft
     steps:
       - name: Checkout
-        uses: actions/checkout@v2
+        uses: actions/checkout@v3
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
@@ -40,7 +40,7 @@ jobs:
           # This should be the path to the paper within your repo.
           paper-path: paper.md
       - name: Upload
-        uses: actions/upload-artifact@v1
+        uses: actions/upload-artifact@v4
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled
