### 2024-09-11 #6233 @JeWaVe
It's now published at https://aerobus-open-source.github.io/aerobus/
README is updated an imported in documentation
added issue and PR template
I cleaned up the paper, tell me if I go in the right direction
