### 2024-09-11 b7835a2
add two example files - hermite polynomial and custom taylor serie + add bell numbers
