+++ examples/hermite.cpp
@@ -1,0 +1,34 @@
+/**
+ * Example demonstating how to use aerobus "pow" function and "hermite" polynomials
+ */
+
+#include <cmath>
+#include <iostream>
+#include "../src/aerobus.h"
+
+namespace standardlib {
+    double H3(double x) {
+        return 8 * std::pow(x, 3) - 12 * x;
+    }
+
+    double H4(double x) {
+        return 16 * std::pow(x, 4) - 48 * x * x + 12;
+    }
+}
+
+namespace aerobuslib {
+    double H3(double x) {
+        return 8 * aerobus::pow_scalar<double, 3>(x) - 12 * x;
+    }
+
+    double H4(double x) {
+        return 16 * aerobus::pow_scalar<double, 4>(x) - 48 * x * x + 12;
+    }
+}
+
+int main() {
+    std::cout << std::hermite(3, 10) << '=' << standardlib::H3(10) << '\n'
+              << std::hermite(4, 10) << '=' << standardlib::H4(10) << '\n';
+    std::cout << aerobus::known_polynomials::hermite_phys<3>::eval(10) << '=' << aerobuslib::H3(10) << '\n'
+              << aerobus::known_polynomials::hermite_phys<4>::eval(10) << '=' << aerobuslib::H4(10) << '\n';
+}
