### 2024-09-12 aeb3cb1
make functions CUDA compatible
