### 2024-09-13 f4b0914
install doxygen for CI
