+++ .github/workflows/c-cpp.yml
@@ -14,7 +14,7 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - name: Install compilers and lib
-      run: sudo apt-get install -y clang g++ libomp-dev cppcheck cpplint
+      run: sudo apt install -y clang g++ libomp-dev cppcheck cpplint doxygen
     - name: cppcheck
       if: success()
       run: cppcheck --std=c++20 --language=c++ src/aerobus.h && cppcheck --std=c++20 --language=c++ src/lib_tests.cpp
