### 2024-09-13 9f25941
install pdflatex on CI
