+++ .github/workflows/c-cpp.yml
@@ -14,7 +14,7 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - name: Install compilers and lib
-      run: sudo apt install -y clang g++ libomp-dev cppcheck cpplint doxygen
+      run: sudo apt install -y clang g++ libomp-dev cppcheck cpplint doxygen texlive-latex-base texlive-fonts-recommended texlive-fonts-extra texlive-latex-extra
     - name: cppcheck
       if: success()
       run: cppcheck --std=c++20 --language=c++ src/aerobus.h && cppcheck --std=c++20 --language=c++ src/lib_tests.cpp
