+++ .github/workflows/c-cpp.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4
     - name: Install compilers and lib
       run: sudo apt install -y clang g++ libomp-dev cppcheck cpplint doxygen texlive-latex-base texlive-fonts-recommended texlive-fonts-extra texlive-latex-extra
     - name: cppcheck
@@ -32,7 +32,7 @@ jobs:
     name: Paper Draft
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
