### 2024-09-16 84c25f9
make  known polynomial dependant on Integers ring (to allow them to live in i32) but defaults to aerobus::i64 so no breaking change
