+++ src/aerobus.h
@@ -2894,121 +2894,136 @@ namespace aerobus {
 namespace aerobus {
     // CChebyshev
     namespace internal {
-        template<int kind, size_t deg>
+        template<int kind, size_t deg, typename I>
         struct chebyshev_helper {
-            using type = typename pi64::template sub_t<
-                typename pi64::template mul_t<
-                    typename pi64::template mul_t<
-                        pi64::inject_constant_t<2>,
-                        typename pi64::X>,
-                    typename chebyshev_helper<kind, deg - 1>::type
+            using type = typename polynomial<I>::template sub_t<
+                typename polynomial<I>::template mul_t<
+                    typename polynomial<I>::template mul_t<
+                        typename polynomial<I>::template inject_constant_t<2>,
+                        typename polynomial<I>::X>,
+                    typename chebyshev_helper<kind, deg - 1, I>::type
                 >,
-                typename chebyshev_helper<kind, deg - 2>::type
+                typename chebyshev_helper<kind, deg - 2, I>::type
             >;
         };
 
-        template<>
-        struct chebyshev_helper<1, 0> {
-            using type = typename pi64::one;
+        template<typename I>
+        struct chebyshev_helper<1, 0, I> {
+            using type = typename polynomial<I>::one;
         };
 
-        template<>
-        struct chebyshev_helper<1, 1> {
-            using type = typename pi64::X;
+        template<typename I>
+        struct chebyshev_helper<1, 1, I> {
+            using type = typename polynomial<I>::X;
         };
 
-        template<>
-        struct chebyshev_helper<2, 0> {
-            using type = typename pi64::one;
+        template<typename I>
+        struct chebyshev_helper<2, 0, I> {
+            using type = typename polynomial<I>::one;
         };
 
-        template<>
-        struct chebyshev_helper<2, 1> {
-            using type = typename pi64::template mul_t<
-                typename pi64::inject_constant_t<2>,
-                typename pi64::X>;
+        template<typename I>
+        struct chebyshev_helper<2, 1, I> {
+            using type = typename polynomial<I>::template mul_t<
+                typename polynomial<I>::template inject_constant_t<2>,
+                typename polynomial<I>::X>;
         };
     }  // namespace internal
 
     // Laguerre
     namespace internal {
-        template<size_t deg>
+        template<size_t deg, typename I>
         struct laguerre_helper {
+            using Q = FractionField<I>;
+            using PQ = polynomial<Q>;
+
          private:
             // Lk = (1 / k) * ((2 * k - 1 - x) * lkm1 - (k - 2)Lkm2)
-            using lnm2 = typename laguerre_helper<deg - 2>::type;
-            using lnm1 = typename laguerre_helper<deg - 1>::type;
+            using lnm2 = typename laguerre_helper<deg - 2, I>::type;
+            using lnm1 = typename laguerre_helper<deg - 1, I>::type;
             // -x + 2k-1
-            using p = typename pq64::template val<
-                typename q64::template inject_constant_t<-1>,
-                typename q64::template inject_constant_t<2 * deg - 1>>;
+            using p = typename PQ::template val<
+                typename Q::template inject_constant_t<-1>,
+                typename Q::template inject_constant_t<2 * deg - 1>>;
             // 1/n
-            using factor = typename pq64::template inject_ring_t<
-                q64::val<typename i64::one, typename i64::template inject_constant_t<deg>>>;
+            using factor = typename PQ::template inject_ring_t<
+                typename Q::template val<typename I::one, typename I::template inject_constant_t<deg>>>;
 
          public:
-            using type = typename pq64::template mul_t <
+            using type = typename PQ::template mul_t <
                 factor,
[…]
