+++ CPPLINT.cfg
@@ -1,5 +1,5 @@
 filter=-legal/copyright
 
-filter=-runtime/indentation_namespace
+filter=-runtime/indentation_namespace,-runtime/references
 
 linelength=120
