### 2024-09-16 cc5201d
run CI on docker image with preinstalled tooling
