+++ .github/workflows/Dockerfile
@@ -1,0 +1,5 @@
+FROM ubuntu:latest
+
+RUN apt update
+RUN apt install  -y clang g++ libomp-dev cppcheck cpplint doxygen texlive-latex-base texlive-fonts-recommended texlive-fonts-extra texlive-latex-extra
+
