### 2024-09-16 82d229b
add make to CI container
