### 2024-09-16 0383af4
add cmake to CI container
