+++ .github/workflows/Dockerfile
@@ -3,3 +3,4 @@ FROM ubuntu:latest
 RUN apt update
 RUN apt install  -y clang g++ libomp-dev cppcheck cpplint doxygen texlive-latex-base texlive-fonts-recommended texlive-fonts-extra texlive-latex-extra
 RUN apt install -y build-essential
+RUN apt install -y cmake
