### 2024-09-17 c2c5dee
add embeddings for algebraic structures (map from z/2Z[X] to Z[X] for example) plus some helper type alias
