+++ src/aerobus.h
@@ -323,6 +323,12 @@ namespace aerobus {
                     val, typename val::enclosing_type::template sub_t<typename val::enclosing_type::zero, val>>;
 }  // namespace aerobus
 
+// embedding
+namespace aerobus {
+    template<typename Small, typename Large, typename E = void>
+    struct Embed;
+}  // namespace aerobus
+
 namespace aerobus {
     /// @brief Quotient ring by the principal ideal generated by 'X'
     /// With i32 as Ring and i32::val<2> as X, Quotient is Z/2Z
@@ -336,6 +342,7 @@ namespace aerobus {
         template <typename V>
         struct val {
          public:
+            using raw_t = V;
             using type = abs_t<typename Ring::template mod_t<V, X>>;
         };
 
@@ -411,6 +418,12 @@ namespace aerobus {
         /// @tparam v a value in Ring
         template<typename v>
         using inject_ring_t = val<v>;
+    };
+
+    template<typename Ring, typename X>
+    struct Embed<Quotient<Ring, X>, Ring> {
+        template<typename val>
+        using type = typename val::raw_t;
     };
 }  // namespace aerobus
 
@@ -964,6 +977,12 @@ namespace aerobus {
         template<typename v>
         static constexpr bool pos_v = pos_t<v>::value;
     };
+
+    template<>
+    struct Embed<i32, i64> {
+        template<typename val>
+        using type = i64::val<static_cast<int64_t>(val::v)>;
+    };
 }  // namespace aerobus
 
 // z/pz
@@ -1133,6 +1152,12 @@ namespace aerobus {
         template<typename v>
         static constexpr bool pos_v = pos_t<v>::value;
     };
+
+    template<int32_t x>
+    struct Embed<zpz<x>, i32> {
+        template <typename val>
+        using type = i32::val<val::v>;
+    };
 }  // namespace aerobus
 
 // polynomial
@@ -1153,6 +1178,8 @@ namespace aerobus {
         /// @tparam ...coeffs lower degree coefficients
         template<typename coeffN, typename... coeffs>
         struct val {
+            /// @brief ring coefficients live in
+            using ring_type = Ring;
             /// @brief enclosing ring type
             using enclosing_type = polynomial<Ring>;
             /// @brief degree of the polynomial
@@ -1208,6 +1235,8 @@ namespace aerobus {
         /// @tparam coeffN
         template<typename coeffN>
         struct val<coeffN> {
+            /// @brief ring coefficients live in
+            using ring_type = Ring;
             /// @brief enclosing ring type
             using enclosing_type = polynomial<Ring>;
             /// @brief degree
@@ -2083,31 +2112,56 @@ namespace aerobus {
     template<typename Ring>
     requires IsEuclideanDomain<Ring>
     using FractionField = typename internal::FractionFieldImpl<Ring>::type;
+
+    template<typename Ring>
+    struct Embed<Ring, FractionField<Ring>> {
+        template<typename v>
+        using type = typename FractionField<Ring>::template val<v, typename Ring::one>;
+    };
 }  // namespace aerobus
+
 
 // short names for common types
 namespace aerobus {
     /// @brief 32 bits rationals
     /// rationals with 32 bits numerator and denominator
     using q32 = FractionField<i32>;
+
     /// @brief rational fractions with 32 bits rational coefficients
     /// rational fractions with rationals coefficients (32 bits numerator and denominator)
     using fpq32 = FractionField<polynomial<q32>>;
+
     /// @brief 64 bits rationals
     /// rationals with 64 bits numerator and denominator
     using q64 = FractionField<i64>;
+
     /// @brief polynomial with 64 bits integers coefficients
     using pi64 = polynomial<i64>;
+
[…]
