+++ src/aerobus.h
@@ -325,6 +325,10 @@ namespace aerobus {
 
 // embedding
 namespace aerobus {
+    /// @brief embedding - struct forward declaration
+    /// @tparam Small a ring which can be embedded in Large
+    /// @tparam Large a ring in which Small can be embedded
+    /// @tparam E some default type (unused -- implementation related)
     template<typename Small, typename Large, typename E = void>
     struct Embed;
 }  // namespace aerobus
@@ -420,8 +424,13 @@ namespace aerobus {
         using inject_ring_t = val<v>;
     };
 
+    /// @brief embeds Quotient<Ring, X> into Ring
+    /// @tparam Ring a Euclidean ring
+    /// @tparam X a value in Ring
     template<typename Ring, typename X>
     struct Embed<Quotient<Ring, X>, Ring> {
+        /// @brief Ring reprensentation of val
+        /// @tparam val a value in Quotient<Ring, X>
         template<typename val>
         using type = typename val::raw_t;
     };
@@ -978,8 +987,11 @@ namespace aerobus {
         static constexpr bool pos_v = pos_t<v>::value;
     };
 
+    /// @brief embeds i32 into i64
     template<>
     struct Embed<i32, i64> {
+        /// @brief the i64 representation of val
+        /// @tparam val a value in i32
         template<typename val>
         using type = i64::val<static_cast<int64_t>(val::v)>;
     };
@@ -987,13 +999,18 @@ namespace aerobus {
 
 // z/pz
 namespace aerobus {
-    /**
-     * congruence classes of integers for a modulus
-     * if p is prime, zpz is a field, otherwise an integral domain with all related operations
-    */
+    /// @brief congruence classes of integers modulo p (32 bits)
+    ///
+    /// if p is prime, zpz<p> is a field
+    ///
+    /// @tparam p a integer
     template<int32_t p>
     struct zpz {
+        /// @brief underlying type for values
         using inner_type = int32_t;
+
+        /// @brief values in zpz<p>
+        /// @tparam x an integer
         template<int32_t x>
         struct val {
             /// @brief enclosing ring type
@@ -1001,10 +1018,19 @@ namespace aerobus {
             /// @brief actual value
             static constexpr int32_t v = x % p;
 
+            /// @brief get value as valueType
+            /// @tparam valueType an arithmetic type, such as float
             template<typename valueType>
             static constexpr DEVICE INLINED valueType get() { return static_cast<valueType>(x % p); }
 
-            using is_zero_t = std::bool_constant<x% p == 0>;
+            /// @brief true_type if zero
+            using is_zero_t = std::bool_constant<v == 0>;
+
+            /// @brief true if zero
+            using is_zero_v = v == 0;
+
+            /// @brief string representation
+            /// @return a string representation
             static std::string to_string() {
                 return std::to_string(x % p);
             }
@@ -1015,12 +1041,21 @@ namespace aerobus {
             }
         };
 
+        /// @brief injects a constant integer into zpz
+        /// @tparam x an integer
         template<auto x>
         using inject_constant_t = val<static_cast<int32_t>(x)>;
 
+        /// @brief zero value
         using zero = val<0>;
+
+        /// @brief one value
         using one = val<1>;
+
+        /// @brief true iff p is prime
         static constexpr bool is_field = is_prime<p>::value;
+
+        /// @brief always true
         static constexpr bool is_euclidean_domain = true;
 
      private:
@@ -1153,8 +1188,12 @@ namespace aerobus {
         static constexpr bool pos_v = pos_t<v>::value;
     };
 
[…]
