### 2024-09-17 71c7f70
add generic arithmetic operators
