### 2024-09-20 3495f2c
support CUDA fp16 (partially) -- unfortunately NVIDIA does not help that much + update benchmarks
