+++ src/benchmarks.cpp
@@ -1,14 +1,11 @@
-#include <immintrin.h>
 #include <cstdio>
 #include <cstdlib>
 #include <chrono> // NOLINT
 #include <cmath>
 #include "./aerobus.h"
 
-const size_t N = 134217728/2;
+const size_t N = 134217728/8;
 
-
-extern "C" __m512d _ZGVeN8v_sin(__m512d x);
 
 #ifndef DEGREE
 #define DEGREE 13
@@ -16,60 +13,43 @@ extern "C" __m512d _ZGVeN8v_sin(__m512d x);
 
 template<int deg>
 INLINED
-double sin_12(const double x) {
-    using AERO_SIN = aerobus::sin<aerobus::i64, deg>;
-    return AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(
-        AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(AERO_SIN::eval(x))))))))))));
+double exp_12(const double x) {
+    using EXP = aerobus::expm1<aerobus::i64, deg>;
+    return EXP::eval(EXP::eval(EXP::eval(EXP::eval(EXP::eval(EXP::eval(
+        EXP::eval(EXP::eval(EXP::eval(EXP::eval(EXP::eval(EXP::eval(x))))))))))));
 }
 
 template<size_t N, int deg>
 INLINED
-void vsin_12(double * in, double * out) {
+void vexp_12(double * in, double * out) {
     static_assert(N % 32 == 0);
     in = reinterpret_cast<double*>(__builtin_assume_aligned(in, 64));
     out = reinterpret_cast<double*>(__builtin_assume_aligned(out, 64));
 #pragma omp parallel for
     for (size_t i = 0; i < N; i += 8) {
         for (size_t j = i; j < i+8; ++j) {
-            out[j] = sin_12<deg>(in[j]);
+            out[j] = exp_12<deg>(in[j]);
         }
     }
 }
 
 INLINED
-double sin_12_slow(const double x) {
-    return ::sin(::sin(::sin(::sin(::sin(::sin(
-        ::sin(::sin(::sin(::sin(::sin(::sin(x))))))))))));
+double exp_12_slow(const double x) {
+    return ::expm1(::expm1(::expm1(::expm1(::expm1(::expm1(
+        ::expm1(::expm1(::expm1(::expm1(::expm1(::expm1(x))))))))))));
 }
 
 template<size_t N>
 INLINED
-void vsin_slow(double * in, double * out) {
+void vexp_12_slow(double * in, double * out) {
     static_assert(N % 32 == 0);
     in = reinterpret_cast<double*>(__builtin_assume_aligned(in, 64));
     out = reinterpret_cast<double*>(__builtin_assume_aligned(out, 64));
 #pragma omp parallel for
     for (size_t i = 0; i < N; i += 8) {
         for (size_t j = i; j < i+8; ++j) {
-            out[j] = sin_12_slow(in[j]);
+            out[j] = exp_12_slow(in[j]);
         }
-    }
-}
-
-
-template<size_t N>
-INLINED
-void fast_math_vsin_12(double * in, double * out) {
-    static_assert(N % 32 == 0);
-    in = reinterpret_cast<double*>(__builtin_assume_aligned(in, 64));
-    out = reinterpret_cast<double*>(__builtin_assume_aligned(out, 64));
-    #pragma omp parallel for
-    for (size_t i = 0; i < N; i += 8) {
[…]
