### 2024-09-20 327675e
add cuda fp16 in README
