+++ .vscode/ltex.disabledRules.en-US.txt
@@ -1,0 +1,1 @@
+SENT_START_CONJUNCTIVE_LINKING_ADVERB_COMMA
