### 2024-09-30 d93a23c
CUDA __half and __half2 supported with very low performance due to NVIDIA bugs 4872028 and 4863696
