+++ examples/fp16.cu
@@ -4,21 +4,26 @@
 #define WITH_CUDA_FP16
 #include "../src/aerobus.h"
 
-using P = aerobus::known_polynomials::hermite_phys<8, aerobus::i16>;
-using P2 = aerobus::known_polynomials::hermite_phys<8, aerobus::i64>;
+/*
+change int_type to aerobus::i32 (or i64) and float_type to float (resp. double)
+to see how good is the generated assembly compared to what nvcc generates for 16 bits 
+*/
+using int_type = aerobus::i16;
+using float_type = __half2;
 
 
-__device__ __forceinline__ __half f(__half x) {
-    return P::eval(x);
+using EXPM1 = aerobus::expm1<int_type, 6>;
+
+
+__device__ INLINED float_type f(float_type x) {
+    return EXPM1::eval(x);
 }
 
-__global__ void run(size_t N, __half* in, __half* out) {
+__global__ void run(size_t N, float_type* in, float_type* out) {
     for(size_t i = threadIdx.x + blockDim.x * blockIdx.x; i < N; i += blockDim.x * gridDim.x) {
-        out[i] = f(in[i]);
+        out[i] = f(f(f(f(f(f(in[i]))))));
     }
 }
 
 int main() {
-    printf("hermite<8, int16>(1) == %f\n", __half2float(P::eval<__half>(__float2half(1.0F))));
-    printf("hermite<8, int64>(1) == %f\n", P2::eval(1.0));
 }
