+++ src/aerobus.h
@@ -95,7 +95,7 @@ namespace aerobus {
 // all this shit is required because of NVIDIA bug https://developer.nvidia.com/bugs/4863696
 namespace aerobus {
     namespace internal {
-        static consteval uint16_t my_internal_float2half(
+        static consteval DEVICE uint16_t my_internal_float2half(
             const float f, uint32_t &sign, uint32_t &remainder) {
             uint32_t x;
             uint32_t u;
@@ -129,7 +129,7 @@ namespace aerobus {
             return static_cast<uint16_t>(result);
         }
 
-        static consteval __half my_float2half_rn(const float a) {
+        static consteval DEVICE __half my_float2half_rn(const float a) {
             __half val;
             __half_raw r;
             uint32_t sign = 0U;
@@ -143,13 +143,70 @@ namespace aerobus {
             return val;
         }
 
-        template<int16_t i>
+        template <int16_t i>
         static constexpr __half convert_int16_to_half = my_float2half_rn(static_cast<float>(i));
+
+
+        template <typename Out, int16_t x, typename E = void>
+        struct int16_convert_helper;
+
+        template <typename Out, int16_t x>
+        struct int16_convert_helper<Out, x,
+            std::enable_if_t<!std::is_same_v<Out, __half> && !std::is_same_v<Out, __half2>>> {
+            static constexpr Out value() {
+                return static_cast<Out>(x);
+            }
+        };
+
+        template <int16_t x>
+        struct int16_convert_helper<__half, x> {
+            static constexpr __half value() {
+                return convert_int16_to_half<x>;
+            }
+        };
+
+        template <int16_t x>
+        struct int16_convert_helper<__half2, x> {
+            static constexpr __half2 value() {
+                return __half2(convert_int16_to_half<x>, convert_int16_to_half<x>);
+            }
+        };
     }  // namespace internal
 }  // namespace aerobus
 #endif
 
-// fma_helper, required because nvidia fails to reconstruct fma
+//  cast
+namespace aerobus {
+    namespace internal {
+        template<typename Out, typename In>
+        struct staticcast {
+            template<auto x>
+            static consteval INLINED DEVICE Out func() {
+                return static_cast<Out>(x);
+            }
+        };
+
+        #ifdef WITH_CUDA_FP16
+        template<>
+        struct staticcast<__half, int16_t> {
+            template<int16_t x>
+            static consteval INLINED DEVICE __half func() {
+                return int16_convert_helper<__half, x>::value();
+            }
+        };
+
+        template<>
+        struct staticcast<__half2, int16_t> {
+            template<int16_t x>
+            static consteval INLINED DEVICE __half2 func() {
+                return int16_convert_helper<__half2, x>::value();
+            }
+        };
+        #endif
+    }  // namespace internal
+}  // namespace aerobus
+
+// fma_helper, required because nvidia fails to reconstruct fma for fp16 types
 namespace aerobus {
     namespace internal {
         template<typename T>
@@ -171,6 +228,13 @@ namespace aerobus {
 
         template<>
         struct fma_helper<int32_t> {
+            static constexpr INLINED DEVICE int16_t eval(const int16_t x, const int16_t y, const int16_t z) {
+                return x * y + z;
+            }
+        };
+
+        template<>
+        struct fma_helper<int16_t> {
             static constexpr INLINED DEVICE int32_t eval(const int32_t x, const int32_t y, const int32_t z) {
                 return x * y + z;
             }
@@ -703,32 +767,6 @@ namespace aerobus {
 #ifdef WITH_CUDA_FP16
 // i16
 namespace aerobus {
-    namespace internal {
-        template<int16_t x, typename Out, typename E = void>
[…]
