+++ src/lib_tests.cpp
@@ -57,11 +57,11 @@ TEST(polynomials, eval) {
 
 TEST(fraction_field, get) {
     using half = q32::val<i32::one, i32::val<2>>;
-    constexpr float x = half::template get<float>;
+    constexpr float x = half::template get<float>();
     EXPECT_EQ(x, 0.5f);
 
     using thirdhalf = q32::val<i32::val<3>, i32::val<2>>;
-    constexpr float y = thirdhalf::template get<float>;
+    constexpr float y = thirdhalf::template get<float>();
     EXPECT_EQ(y, 1.5f);
 
     // 3/2 + x / 2
@@ -707,6 +707,15 @@ TEST(utilities, exp) {
     EXPECT_TRUE((std::abs(std::exp(0.1f) - e01) <= 1E-7F));
 }
 
+TEST(utilities, expm1) {
+    using E = aerobus::expm1<i32, 12>;
+    constexpr float e0 = E::eval(0.0F);
+    constexpr float e01 = E::eval(0.1F);
+    EXPECT_EQ(e0, 0.0f);
+
+    EXPECT_TRUE((std::abs(std::expm1(0.1f) - e01) <= 1E-7F));
+}
+
 TEST(utilities, alternate) {
     constexpr int a0 = internal::alternate<i32, 0>::value;
     EXPECT_EQ(a0, 1);
@@ -861,17 +870,17 @@ TEST(known_polynomials, chebyshev) {
 TEST(known_polynomials, laguerre) {
     {
         using L2 = known_polynomials::laguerre<2>;
-        EXPECT_EQ((L2::coeff_at_t<2>::template get<float>), 0.5F);
-        EXPECT_EQ((L2::coeff_at_t<1>::template get<float>),  -2.0F);
-        EXPECT_EQ((L2::coeff_at_t<0>::template get<float>),  1.0F);
+        EXPECT_EQ((L2::coeff_at_t<2>::template get<float>()), 0.5F);
+        EXPECT_EQ((L2::coeff_at_t<1>::template get<float>()),  -2.0F);
+        EXPECT_EQ((L2::coeff_at_t<0>::template get<float>()),  1.0F);
     }
     {
         using L3 = known_polynomials::laguerre<3>;
         EXPECT_EQ((L3::coeff_at_t<3>::x::v), -1);
         EXPECT_EQ((L3::coeff_at_t<3>::y::v), 6);
-        EXPECT_EQ((L3::coeff_at_t<2>::template get<float>), 1.5F);
-        EXPECT_EQ((L3::coeff_at_t<1>::template get<float>), -3.0F);
-        EXPECT_EQ((L3::coeff_at_t<0>::template get<float>), 1.0F);
+        EXPECT_EQ((L3::coeff_at_t<2>::template get<float>()), 1.5F);
+        EXPECT_EQ((L3::coeff_at_t<1>::template get<float>()), -3.0F);
+        EXPECT_EQ((L3::coeff_at_t<0>::template get<float>()), 1.0F);
     }
 }
 
@@ -879,29 +888,29 @@ TEST(known_polynomials, laguerre) {
 TEST(known_polynomials, hermite) {
     {
         using H2 = known_polynomials::hermite_prob<2>;
-        EXPECT_EQ((H2::coeff_at_t<2>::get<float>), 1.0F);
-        EXPECT_EQ((H2::coeff_at_t<1>::get<float>), 0.0F);
-        EXPECT_EQ((H2::coeff_at_t<0>::get<float>), -1.0F);
[…]
