### 2024-09-30 #6233 @lucaferranti
Hi @JeWaVe (cc @gkthiruvathukal )

The repository looks much better now! Some comments:

- In the [examples section](https://aerobus-open-source.github.io/aerobus/examples.html) of the documentation, there are several broken links (QuotientField, FractionField, PI_Fraction::val). For the working links, those seem more like API documentation and not example usages.
- Contributing guidelines are still missing

Also, my previous comment about the paper length stands:

> My main comment for the paper is that at the moment I think it does not conform to the JOSS format. Mainly, the paper is currently 12 pages, while generally JOSS expects a much shorter (~ 2 pages) paper, which should give a general motivation and overview of timeless features of the sofware. Currently, the paper presents several examples and code snippets, which I believe would better fit the documentation. The reason for this is that software can change and things written in the paper are frozen, while documentation can evolve and update (see also my comment on documentation below). However I let the editor decide and overwrite my comment if he thinks the paper is appropriate this way.
