### 2024-10-01 e0a6d60
Create CONTRIBUTING.md
