+++ .github/CONTRIBUTING.md
@@ -1,0 +1,40 @@
+There are several ways to contribute to Aerobus, and all are equally welcomed. 
+
+Whatever the reason, please follow our Code of Conduct. 
+
+# No code
+
+## Raise a bug
+
+Raising bug is crucial when it comes to contributing to software development, particularly in open source world. 
+When you have questions, encounter issues, see discrepancies in API or behavior, please take the time to open an issue. 
+
+## Propose features or enhancements
+
+If you think about something of interest that aerobus can help solve, or something it should integrate, please open an issue and tag it "enhancement". 
+
+# Code
+
+For all that follows : 
+
+- implementations must have an associated issue
+- develop in dedicated branch, labelled with issue ID and name, for example : "bug-41-undefined-behavior-with-intel-compiler" or "feature-28-compensated-horner". 
+- write **clean code** (explicit names, short functions, no deep nesting, ...)
+- comment and document your code
+- do not hesitate to engage discussions with others
+- test your code
+- rebase your branch on main
+- propose pull request with detailed explanations of what's going on
+
+## Fix bugs
+
+Fixing bugs is one of the best ways to get onboarded on complex code bases. As every project on earth, aerobus warmly welcome bug fixes.
+
+## Write tests
+
+Unit testing is a never-finished task. The more tests, the more stable the code. Tests are always welcomed.
+
+## Implement features
+
+Before starting development, please engage discussion in an issue and open it if it's not already. 
+
