### 2024-10-01 d1bf661
Create CODE_OF_CONDUCT.md
