### 2024-10-01 #6233 @JeWaVe
Hi,

I added contributing guidelines and code of conduct.
I understood why examples are broken and that is because I misunderstood the @example tag. Going to fix them all now.
