### 2024-10-01 #6233 @pitsianis
I am not satisfied with the changes. The draft has no coherent message; it is an enumeration of the contents of a bag of different things that only have in common that they have been included in the Aerobus macro library.

Let's start from the title, which specifies "polynomial algebra over **discrete** Euclidean domains", however The Statement of Need only dedicates a single line:
> By implementing general algebra concepts such as discrete rings, field of fractions and polynomials, Aerobus can serve multiple purposes.

that says nothing to address the need for a polynomial algebra over discrete domains.

The rest of The Statement of Need talks only about the Taylor expansion for the approximation of transcendental functions:
> The main application we want to express in this paper is the automatic (and configurable)
generation or Taylor approximation of usual transcendental functions such as exp or sin.

Taylor expansion relies on the smoothness of functions and their derivatives. Are you considering the floating point representation of real numbers as the discrete domain? Or, is this macro package doing something with finite differences to express function changes?

Back to the first line. What does it mean to "implement the concept of discrete ring"? Given by the user an addition and a multiplication operator, does it verify the ring or the field axioms like associativity, distribution of the multiplication over addition, existence of zero (and one) identity element(s), and inverses?  What would happen if the user attempted to use operators that do not satisfy the corresponding axioms, for instance, using the semiring operators `min(int32,int32)` and `plus(int32,int32)` as the addition and multiplication?

Why is it advantageous to precompute at compile-time the addition and multiplication tables of a Galois field using macros instead of "normal" run-time codes? Why is it advantageous to have a macro to form continued fractions instead of a function at run time?

Some of the claims might also be incorrect. Why precomputing shrinks binary size? For example if I implement Horner's rule as a loop, and compare it against the code with the same loop unrolled, which code will have the smaller binary for a polynomial of degree 20?

I will reiterate that I do find this work interesting overall. However the article fails to provide a clear motivation why such a macro library is needed beyond the Taylor expansion of transcendental functions that allows the precomputation of specialized fast approximations.
