### 2024-10-01 6565325
more documentations - update JOSS paper
