+++ paper.md
@@ -46,19 +46,19 @@ These functions are usually exposed by the standard library (`<cmath>`) with hig
 
 Hardware vendors provide high-performance libraries such as [@wang2014intel], but implementation is often hidden and not extensible.
 
-Some others can provide vectorized functions, such as @vml does. But libraries like VML are highly tight to one architecture by their use of intrinsic or inline assembly. In addition, they only provide a restricted list of math functions and do not expose capabilities to generate high-performance versions of other functions such as `arctanh`. It is the same for the standard library compiled with `-Ofast`: it generates a vectorized version of some functions (such as exp) but with no control of precision and no extensibility.
+Some others can provide vectorized functions, such as @vml does. But libraries like VML are highly tight to one architecture by their use of intrinsic or inline assembly. In addition, they only provide a restricted list of math functions and do not expose capabilities to generate high-performance versions of other functions such as `arctanh`. It is the same for the standard library compiled with `-Ofast`: it generates a vectorized version of some functions (such as exp) but with no control of precision and no extensibility. In addition, `fast-math` versions are compiler and architecture dependent, which can be a problem for results reproducibility.
 
-`Aerobus` provides automatic generation of such functions, in a hardware-independent way.
+`Aerobus` provides automatic generation of such functions, in a hardware-independent way, as tested on x86 and CUDA platforms.
 In addition, `Aerobus` provides a way to control the precision of the generated function by changing the degree of Taylor expansion, which can't be used in competing libraries without reimplementing the whole function.
 
 ## Mathematic definitions
 
-For the sake of completeness, we give basic definitions of the mathematical concepts which the library deals with. However, readers desiring complete and rigorous definitions of the concepts explained below should refer to some mathematical books on algebra, such as @lang2012algebra or @bourbaki2013algebra.
+For the sake of completeness, we give basic definitions of the mathematical concepts which the library deals with. However, readers desiring complete and rigorous definitions of the concepts explained below should refer to a mathematical book on algebra, such as @lang2012algebra or @bourbaki2013algebra.
 
 A `ring` $\mathbb{A}$ is a nonempty set with two internal laws, addition and multiplication. There is a neutral element for both, zero and one.
 Addition is commutative and associative and every element $x$ has an inverse $-x$. Multiplication is commutative, associative and distributive over addition, meaning that $a(b+c) = ab+ac$ for every $a, b, c$ element. We call it `discrete` if it is countable.
 
-An `integral domain` is a ring with one additional property. For every elements $a, b, c$ such as $ab = ac$, then either $a = 0$ or $b = c$. Such a ring is not always a field, such as $\mathbb{Z}$ shows it.
+A `integral domain` is a ring with one additional property. For every element $a, b, c$ such as $ab = ac$, then either $a = 0$ or $b = c$. Such a ring is not always a field, such as $\mathbb{Z}$ shows it.
 
 A `euclidean domain` is an integral domain that can be endowed with a euclidean division.
 
@@ -125,13 +125,6 @@ The same is done for the field of fractions: implementation does not rely on the
 
 Doing that way, $\mathbb{Q}$ has the same implementation as rational fractions of polynomials. Users could also get the field of fractions of any ring of their convenience, as long as they implement the required concepts.
 
-For example, rationals and rational fractions with rational coefficients are exposed through type aliases:
-
-```cpp
-using q32 = FractionField<i32>;
[…]
