+++ examples/make_polynomial.cpp
@@ -1,0 +1,45 @@
+#include <iostream>
+#include "../src/aerobus.h"
+
+
+// let's build Abel polynomials from scratch using Aerobus
+
+template<typename I = aerobus::i64>
+struct AbelHelper {
+ private:
+    using P = aerobus::polynomial<I>;
+
+ public:
+    // to keep recursion working, we need to operate on a*n and not just a
+    template<size_t deg, I::inner_type an>
+    struct Inner {
+        // abel(n, a) = (x-an) * abel(n-1, a)
+        using type = typename aerobus::mul_t<
+            typename Inner<deg-1, an>::type,
+            typename aerobus::sub_t<typename P::X, typename P::template inject_constant_t<an>>
+        >;
+    };
+
+    // abel(0, a) = 1
+    template<I::inner_type an>
+    struct Inner<0, an> {
+        using type = P::one;
+    };
+
+    // abel(1, a) = X
+    template<I::inner_type an>
+    struct Inner<1, an> {
+        using type = P::X;
+    };
+};
+
+template<size_t n, auto a, typename I = aerobus::i64>
+using AbelPolynomials = typename AbelHelper<I>::template Inner<n, a*n>::type;
+
+using A2_3 = AbelPolynomials<3, 2>;
+
+int main() {
+    std::cout << "expected = x^3 - 12 x^2 + 36 x" << std::endl;
+    std::cout << "aerobus  = " << A2_3::to_string() << std::endl;
+    return 0;
+}
