+++ examples/continued_fractions.cpp
@@ -1,0 +1,17 @@
+
+#include <cmath>
+#include <iostream>
+#include <iomanip>
+#include "../src/aerobus.h"
+
+static constexpr double PHI = aerobus::ContinuedFraction<
+        1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
+        1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1>::val;
+
+static const double phi = (std::sqrt(5.0) + 1.0)/2.0;
+
+int main() {
+    std::cout << std::setprecision(15) << "Aerobus  PHI : " << PHI << std::endl;
+    std::cout << std::setprecision(15) << "Computed PHI : " << phi << std::endl;
+    return 0;
+}
