### 2024-10-02 e378684
rephrase statement of need and add bibliography
