+++ paper.bib
@@ -89,5 +89,28 @@
   publisher={Addison-Wesley Professional}
 }
 
+@article{karney2013algorithms,
+  title={Algorithms for geodesics},
+  author={Karney, Charles FF},
+  journal={Journal of Geodesy},
+  volume={87},
+  pages={43--55},
+  year={2013},
+  publisher={Springer}
+}
 
+@inproceedings{chevillard2010sollya,
+  title={Sollya: An environment for the development of numerical codes},
+  author={Chevillard, Sylvain and Jolde{\c{s}}, Mioara and Lauter, Christoph},
+  booktitle={Mathematical Software--ICMS 2010: Third International Congress on Mathematical Software, Kobe, Japan, September 13-17, 2010. Proceedings 3},
+  pages={28--31},
+  year={2010},
+  organization={Springer}
+}
 
+@article{graillat2006compensated,
+  title={Compensated horner scheme},
+  author={Graillat, Stef and Langlois, Philippe and Louvet, Nicolas},
+  journal={Algebraic and Numerical Algorithms and},
+  year={2006}
+}
