+++ paper.md
@@ -23,33 +23,37 @@ bibliography: paper.bib
 C++ comes with high compile-time computations capability, also known as metaprogramming with templates.
 Templates are a language-in-the-language which is Turing-complete, meaning we can run every computation at compile time instead of runtime, as long as input data is known at compile time.
 
-Using these capabilities, vastly extended with the latest versions of the standard, we implemented a library for discrete Euclidean domains, such as $\mathbb{Z}$. We also provide a way to generate the fraction field of such rings (e.g. $\mathbb{Q}$).
+Using these capabilities, vastly extended with the latest versions of the standard, we implemented a library for discrete Euclidean domains (commutative and associative), such as $\mathbb{Z}$. We also provide a way to generate the fraction field of such rings (e.g. $\mathbb{Q}$).
 
 We also implemented polynomials over such discrete rings and fields (e.g. $\mathbb{Q}[X]$). Since polynomials are also a ring, the above implementation gives us rational fractions as the field of fractions of polynomials.
 
-In addition, we expose a way to generate the Taylor series of any math functions as long as coefficients are known.
+In addition, we expose a way to generate Taylor series of analytic functions, known polynomials (e.g. Chebyshev), continued fractions, quotient rings and small degree Conway polynomials to define Galois finite fields.
 
-In addition, we added some useful additional features, such as known polynomials (Chebyshev), continued fractions, quotient rings and some Conway polynomials to define Galois finite fields.
-
-`Aerobus` was designed to be used in high-performance software, teaching purposes or embedded software where as much as possible must be precomputed to shrink binary size. It compiles with major compilers: `gcc`, `clang` and `msvc`. It is quite easily configurable and extensible.
+`Aerobus` was designed to be used in high-performance software, teaching purposes or embedded software. It compiles with major compilers: `gcc`, `clang` and `msvc`. It is quite easily configurable and extensible.
 
 ## Statement of need
 
-By implementing general algebra concepts such as discrete rings, field of fractions and polynomials, `Aerobus` can serve multiple purposes.
+By implementing general algebra concepts such as discrete rings, field of fractions and polynomials, `Aerobus` can serve multiple purposes, mainly polynomial arithmetic at compile time and efficient polynomial evaluation, regardless of the coefficients `ring`.
 
-The main application we want to express in this paper is the automatic (and configurable) generation or Taylor approximation of usual transcendental functions such as `exp` or `sin`.
+The main application we want to express in this paper is the automatic (and configurable) generation or Taylor approximation of analytic functions such as `exp` or `sin`, by using polynomial arithmetic at compile time.
 
-These functions are usually exposed by the standard library (`<cmath>`) with high (guaranteed) precision. However, in high-performance computing, when not compiled with `-Ofast`, evaluating `std::exp` has several flaws:
+Some important software, such as `geographiclib` [karney2013algorithms] evaluate polynomials with a simple loop, expected to be unrolled by compiler. It works really well (on arithmetic types such as `double`) but does not provide a way to manipulate polynomials (addition, multiplication, division, modulus) automatically where `aerobus` does at no runtime cost.
+
+Notable libraries such as [boost](https://live.boost.org/doc/libs/1_86_0/libs/math/doc/html/math_toolkit/polynomials.html) provide polynomial arithmetic, but arithmetic is done at runtime with memory allocations while `aerobus` does it at compile time.
+
+Common analytic functions are usually exposed by the standard library (`<cmath>`) with high (guaranteed) precision. However, in high-performance computing, when not compiled with `-Ofast`, evaluating `std::exp` has several flaws:
 
 - It leads to a `syscall` which is very expensive;
[…]
