### 2024-10-02 #6233 @JeWaVe
Hi everyone and thanks for your comments.

I fixed documentation by adding actual examples and tried to rephrase the statement of need.

Basically the point is that aerobus provide both efficient polynomial evaluation and polynomial arithmetic at no runtime cost.

I hope it addresses your respective points.

Latest pdf above
