### 2024-10-02 8b4d6d3
fix bibliography
