+++ paper.md
@@ -37,7 +37,7 @@ By implementing general algebra concepts such as discrete rings, field of fracti
 
 The main application we want to express in this paper is the automatic (and configurable) generation or Taylor approximation of analytic functions such as `exp` or `sin`, by using polynomial arithmetic at compile time.
 
-Some important software, such as `geographiclib` [karney2013algorithms] evaluate polynomials with a simple loop, expected to be unrolled by compiler. It works really well (on arithmetic types such as `double`) but does not provide a way to manipulate polynomials (addition, multiplication, division, modulus) automatically where `aerobus` does at no runtime cost.
+Some important software, such as `geographiclib` [@karney2013algorithms] evaluate polynomials with a simple loop, expected to be unrolled by compiler. It works really well (on arithmetic types such as `double`) but does not provide a way to manipulate polynomials (addition, multiplication, division, modulus) automatically where `aerobus` does at no runtime cost.
 
 Notable libraries such as [boost](https://live.boost.org/doc/libs/1_86_0/libs/math/doc/html/math_toolkit/polynomials.html) provide polynomial arithmetic, but arithmetic is done at runtime with memory allocations while `aerobus` does it at compile time.
 
@@ -53,11 +53,11 @@ Some others can provide vectorized functions, such as [@wang2014intel] does. But
 
 `Aerobus` provides automatic generation of such functions, in a hardware-independent way, as tested on x86 and CUDA platforms. In addition, `Aerobus` provides a way to control the precision of the generated function by changing the degree of Taylor expansion, which can't be used in competing libraries without reimplementing the whole function or changing the array of coefficients.
 
-Please note that, `Aerobus` does not provide optimal approximation polynomials the way [chevillard2010sollya] does. However, `Sollya` can be used beforehand to feed `aerobus` with appropriate coefficients. Similarly, `Aerobus` does not perform (yet) compensated Horner scheme like in [graillat2006compensated] or floating point manipulations (domain normalization) to extend domain of approximation, like it is done in standard library.
+Please note that, `Aerobus` does not provide optimal approximation polynomials the way [@chevillard2010sollya] does. However, `Sollya` can be used beforehand to feed `aerobus` with appropriate coefficients. Similarly, `Aerobus` does not perform (yet) compensated Horner scheme like in [@graillat2006compensated] or floating point manipulations (domain normalization) to extend domain of approximation, like it is done in standard library.
 
 ## Mathematic definitions
 
-For the sake of completeness, we give basic definitions of the mathematical concepts which the library deals with. However, readers desiring complete and rigorous definitions of the concepts explained below should refer to a mathematical book on algebra, such as @lang2012algebra or @bourbaki2013algebra.
+For the sake of completeness, we give basic definitions of the mathematical concepts which the library deals with. However, readers desiring complete and rigorous definitions of the concepts explained below should refer to a mathematical book on algebra, such as [@lang2012algebra] or [@bourbaki2013algebra].
 
 A `ring` $\mathbb{A}$ is a nonempty set with two internal laws, addition and multiplication. There is a neutral element for both, zero and one.
 Addition is commutative and associative and every element $x$ has an inverse $-x$. Multiplication is commutative, associative and distributive over addition, meaning that $a(b+c) = ab+ac$ for every $a, b, c$ element. We call it `discrete` if it is countable.
@@ -91,7 +91,7 @@ Given polynomials over a field form an Euclidean ring, we can do the same constr
 
[…]
