+++ src/aerobus.h
@@ -2889,32 +2889,37 @@ namespace aerobus {
     template<typename T, int k>
     using alternate_t = typename internal::alternate<T, k>::type;
 
+    /// @brief (-1)^k as value from T
+    /// @tparam T Ring type, aerobus::i64 for example, then result will be an int64_t
+    template<typename T, size_t k>
+    inline constexpr typename T::inner_type alternate_v = internal::alternate<T, k>::value;
+
     namespace internal {
         template<typename T, int n, int k, typename E = void>
-        struct stirling_helper {};
+        struct stirling_1_helper {};
 
         template<typename T>
-        struct stirling_helper<T, 0, 0> {
+        struct stirling_1_helper<T, 0, 0> {
             using type = typename T::one;
         };
 
         template<typename T, int n>
-        struct stirling_helper<T, n, 0, std::enable_if_t<(n > 0)>> {
+        struct stirling_1_helper<T, n, 0, std::enable_if_t<(n > 0)>> {
             using type = typename T::zero;
         };
 
         template<typename T, int n>
-        struct stirling_helper<T, 0, n, std::enable_if_t<(n > 0)>> {
+        struct stirling_1_helper<T, 0, n, std::enable_if_t<(n > 0)>> {
             using type = typename T::zero;
         };
 
         template<typename T, int n, int k>
-        struct stirling_helper<T, n, k, std::enable_if_t<(k > 0) && (n > 0)>> {
+        struct stirling_1_helper<T, n, k, std::enable_if_t<(k > 0) && (n > 0)>> {
             using type = typename T::template sub_t<
-                            typename stirling_helper<T, n-1, k-1>::type,
+                            typename stirling_1_helper<T, n-1, k-1>::type,
                             typename T::template mul_t<
                                 typename T::template inject_constant_t<n-1>,
-                                typename stirling_helper<T, n-1, k>::type
+                                typename stirling_1_helper<T, n-1, k>::type
                             >>;
         };
     }  // namespace internal
@@ -2924,34 +2929,72 @@ namespace aerobus {
     /// @tparam n (integer)
     /// @tparam k (integer)
     template<typename T, int n, int k>
-    using stirling_signed_t = typename internal::stirling_helper<T, n, k>::type;
+    using stirling_1_signed_t = typename internal::stirling_1_helper<T, n, k>::type;
 
     /// @brief Stirling number of first king (unsigned) -- as types
     /// @tparam T (ring type, such as aerobus::i64)
     /// @tparam n (integer)
     /// @tparam k (integer)
     template<typename T, int n, int k>
-    using stirling_unsigned_t = abs_t<typename internal::stirling_helper<T, n, k>::type>;
-
-    /// @brief Stirling number of first king (signed) -- as value
-    /// @tparam T (ring type, such as aerobus::i64)
-    /// @tparam n (integer)
-    /// @tparam k (integer)
-    template<typename T, int n, int k>
-    static constexpr typename T::inner_type stirling_signed_v = stirling_signed_t<T, n, k>::v;
-
+    using stirling_1_unsigned_t = abs_t<typename internal::stirling_1_helper<T, n, k>::type>;
 
     /// @brief Stirling number of first king (unsigned) -- as value
     /// @tparam T (ring type, such as aerobus::i64)
     /// @tparam n (integer)
     /// @tparam k (integer)
     template<typename T, int n, int k>
-    static constexpr typename T::inner_type stirling_unsigned_v = stirling_unsigned_t<T, n, k>::v;
+    static constexpr typename T::inner_type stirling_1_unsigned_v = stirling_1_unsigned_t<T, n, k>::v;
 
-    /// @brief (-1)^k as value from T
-    /// @tparam T Ring type, aerobus::i64 for example, then result will be an int64_t
[…]
