+++ src/benchmarks.cpp
@@ -1,57 +1,8 @@
-#include <cstdio>
-#include <cstdlib>
-#include <chrono> // NOLINT
-#include <cmath>
 #include "./aerobus.h"
+#include <benchmark/benchmark.h>  // NOLINT
+#include <cmath>  // NOLINT
 
-const size_t N = 134217728/8;
-
-
-#ifndef DEGREE
-#define DEGREE 13
-#endif
-
-template<int deg>
-INLINED
-double exp_12(const double x) {
-    using EXP = aerobus::expm1<aerobus::i64, deg>;
-    return EXP::eval(EXP::eval(EXP::eval(EXP::eval(EXP::eval(EXP::eval(
-        EXP::eval(EXP::eval(EXP::eval(EXP::eval(EXP::eval(EXP::eval(x))))))))))));
-}
-
-template<size_t N, int deg>
-INLINED
-void vexp_12(double * in, double * out) {
-    static_assert(N % 32 == 0);
-    in = reinterpret_cast<double*>(__builtin_assume_aligned(in, 64));
-    out = reinterpret_cast<double*>(__builtin_assume_aligned(out, 64));
-#pragma omp parallel for
-    for (size_t i = 0; i < N; i += 8) {
-        for (size_t j = i; j < i+8; ++j) {
-            out[j] = exp_12<deg>(in[j]);
-        }
-    }
-}
-
-INLINED
-double exp_12_slow(const double x) {
-    return ::expm1(::expm1(::expm1(::expm1(::expm1(::expm1(
-        ::expm1(::expm1(::expm1(::expm1(::expm1(::expm1(x))))))))))));
-}
-
-template<size_t N>
-INLINED
-void vexp_12_slow(double * in, double * out) {
-    static_assert(N % 32 == 0);
-    in = reinterpret_cast<double*>(__builtin_assume_aligned(in, 64));
-    out = reinterpret_cast<double*>(__builtin_assume_aligned(out, 64));
-#pragma omp parallel for
-    for (size_t i = 0; i < N; i += 8) {
-        for (size_t j = i; j < i+8; ++j) {
-            out[j] = exp_12_slow(in[j]);
-        }
-    }
-}
+static constexpr size_t N = 1 << 23;
 
 double rand(double min, double max) {
   double range = (max - min);
@@ -59,103 +10,76 @@ double rand(double min, double max) {
   return min + (rand() / div);  // NOLINT
 }
 
-template<int deg>
-void run_aero(double* in, double* out_aero) {
-    // warmup
-    vexp_12<N, deg>(in, out_aero);
-    vexp_12<N, deg>(in, out_aero);
-    vexp_12<N, deg>(in, out_aero);
-    double best = 1.0E9;
-    for (int i = 0; i < 10; ++i) {
-        auto start = std::chrono::steady_clock::now();
-        vexp_12<N, deg>(in, out_aero);
-        auto stop = std::chrono::steady_clock::now();
-        std::chrono::duration<double> time = stop - start;
-        if (time.count() < best) {
-            best = time.count();
-        }
-    }
-
-    printf("[aerobus deg %d] %.3e Gexp/s\n", deg, (12.0E-9 * static_cast<double>(N)) / best);
+INLINED double aero_expm1_12(const double x) {
+    using EXPM1 = aerobus::expm1<aerobus::i64, 13>;
+    return EXPM1::eval(EXPM1::eval(EXPM1::eval(EXPM1::eval(EXPM1::eval(EXPM1::eval(
+                EXPM1::eval(EXPM1::eval(EXPM1::eval(EXPM1::eval(EXPM1::eval(EXPM1::eval(x))))))))))));
 }
 
[…]
