### 2024-10-02 #6233 @JeWaVe
One more comment if I may.

I rewrote my benchmarks to use google benchmarks but I wonder if I should include results in the paper, I mean they are OK but I usually don't trust benchmarks
