+++ CONTRIBUTING.md
@@ -1,40 +1,41 @@
-There are several ways to contribute to Aerobus, and all are equally welcomed. 
+# Contributing
 
-Whatever the reason, please follow our Code of Conduct. 
+There are several ways to contribute to `Aerobus`, and all are equally welcomed.
 
-# No code
+Whatever the reason, please follow our Code of Conduct.
 
-## Raise a bug
+## No code
 
-Raising bug is crucial when it comes to contributing to software development, particularly in open source world. 
-When you have questions, encounter issues, see discrepancies in API or behavior, please take the time to open an issue. 
+### Raise a bug
 
-## Propose features or enhancements
+Raising bug is crucial when it comes to contributing to software development, particularly in open source world.
+When you have questions, encounter issues, see discrepancies in API or behavior, please take the time to open an issue.
 
-If you think about something of interest that aerobus can help solve, or something it should integrate, please open an issue and tag it "enhancement". 
+### Propose features or enhancements
 
-# Code
+If you think about something of interest that `Aerobus` can help solve, or something it should integrate, please open an issue and tag it "enhancement".
 
-For all that follows : 
+## Code
 
-- implementations must have an associated issue
-- develop in dedicated branch, labelled with issue ID and name, for example : "bug-41-undefined-behavior-with-intel-compiler" or "feature-28-compensated-horner". 
-- write **clean code** (explicit names, short functions, no deep nesting, ...)
-- comment and document your code
-- do not hesitate to engage discussions with others
-- test your code
-- rebase your branch on main
-- propose pull request with detailed explanations of what's going on
+For all that follows :
 
-## Fix bugs
+- Implementations must have an associated issue;
+- Develop in dedicated branch, labelled with issue ID and name, for example : "bug-41-undefined-behavior-with-intel-compiler" or "feature-28-compensated-horner".*;
+- Write **clean code** (explicit names, short functions, no deep nesting, ...);
+- Comment and document your code;
+- Do not hesitate to engage discussions with others;
+- Test your code;
+- Rebase your branch on main;
+- Propose pull request with detailed explanations of what's going on.
 
-Fixing bugs is one of the best ways to get onboarded on complex code bases. As every project on earth, aerobus warmly welcome bug fixes.
+### Fix bugs
 
-## Write tests
+Fixing bugs is one of the best ways to get onboarded on complex code bases. As every project on earth, `Aerobus` warmly welcome bug fixes.
+
+### Write tests
 
 Unit testing is a never-finished task. The more tests, the more stable the code. Tests are always welcomed.
 
-## Implement features
+### Implement features
 
-Before starting development, please engage discussion in an issue and open it if it's not already. 
-
+Before starting development, please engage discussion in an issue and open it if it's not already.
