### 2024-10-07 b3c40d4
implement #16 : compensated horner for polynomials
