+++ src/lib_tests.cpp
@@ -55,6 +55,21 @@ TEST(polynomials, eval) {
     EXPECT_EQ(vvvv, -1.0);
 }
 
+TEST(polynomials, compensated_eval) {
+    // 1 + 2x + 3x^2
+    using poly = make_int_polynomial_t<i32, 3, 2, 1>;
+    float vv = poly::compensated_eval(1.0f);
+    EXPECT_EQ(vv, 6.0f);
+
+    // 1/2 + 3x/2
+    using polyf = polynomial<q32>::val<q32::val<i32::val<3>, i32::val<2>>, q32::val<i32::val<1>, i32::val<2>>>;
+    float vvv = polyf::compensated_eval(1.0f);
+    EXPECT_EQ(vvv, 2.0f);
+
+    double vvvv = polyf::compensated_eval(-1.0);
+    EXPECT_EQ(vvvv, -1.0);
+}
+
 TEST(fraction_field, get) {
     using half = q32::val<i32::one, i32::val<2>>;
     constexpr float x = half::template get<float>();
