+++ src/benchmarks.cpp
@@ -77,9 +77,44 @@ static void BM_std_hermite(benchmark::State &state) {
     }
 }
 
+static void BM_compensated_horner_float(benchmark::State &state) {
+    using P = aerobus::make_int_polynomial_t<aerobus::i64, 1, -11, 55, -165, 330, -462, 462, -330, 165, -55, 11, -1>;
+
+    float *in = aerobus::aligned_malloc<float>(N, 64);
+    float *out = aerobus::aligned_malloc<float>(N, 64);
+    #pragma omp parallel for
+    for (size_t i = 0; i < N; ++i) {
+        in[i] = rand(0.9, 1.1);
+    }
+    for (auto _ : state) {
+        #pragma omp parallel for
+        for (size_t i = 0; i < N; ++i) {
+            out[i] = P::compensated_eval(in[i]);
+        }
+    }
+}
+static void BM_horner_double(benchmark::State &state) {
+    using P = aerobus::make_int_polynomial_t<aerobus::i64, 1, -11, 55, -165, 330, -462, 462, -330, 165, -55, 11, -1>;
+
+    double *in = aerobus::aligned_malloc<double>(N, 64);
+    double *out = aerobus::aligned_malloc<double>(N, 64);
+    #pragma omp parallel for
+    for (size_t i = 0; i < N; ++i) {
+        in[i] = rand(0.9, 1.1);
+    }
+    for (auto _ : state) {
+        #pragma omp parallel for
+        for (size_t i = 0; i < N; ++i) {
+            out[i] = P::eval(in[i]);
+        }
+    }
+}
+
 BENCHMARK(BM_std_expm1_12);
 BENCHMARK(BM_aero_expm1_12);
 BENCHMARK(BM_std_hermite);
 BENCHMARK(BM_aero_hermite);
+BENCHMARK(BM_horner_double);
+BENCHMARK(BM_compensated_horner_float);
 
 BENCHMARK_MAIN();
