### 2024-10-07 c700981
add tests for compensated float vs horner double -- epsilon visualization
