+++ examples/compensated_horner.cpp
@@ -16,12 +16,17 @@ DEVICE INLINED T eval(const T& x) {
 }
 
 template<typename T>
-DEVICE INLINED T exact(const T& x) {
+DEVICE INLINED T exact_large(const T& x) {
     return pow_scalar<T, 5>(0.75 - x) * pow_scalar<T, 11>(1 - x);
 }
 
+template<typename T>
+DEVICE INLINED T exact_small(const T& x) {
+    return pow_scalar<T, 3>(x - 1);
+}
+
 template<typename P, typename T, bool compensated>
-void run(T left, T right, const char *file_name) {
+void run(T left, T right, const char *file_name, T (*exact)(const T&)) {
     FILE *f = ::fopen(file_name, "w+");
     T step = (right - left) / NB_POINTS;
     T x = left;
@@ -33,23 +38,29 @@ void run(T left, T right, const char *file_name) {
 }
 
 int main() {
-    // (0.75 - x)^5 * (1 - x)^11
-    using P = mul_t<
-        pow_t<pq64, pq64::val<
-            typename q64::template inject_constant_t<-1>,
-            q64::val<i64::val<3>, i64::val<4>>>, 5>,
-        pow_t<pq64, pq64::val<typename q64::template inject_constant_t<-1>, typename q64::one>, 11>
-    >;
+    {
+        // (0.75 - x)^5 * (1 - x)^11
+        using P = mul_t<
+            pow_t<pq64, pq64::val<
+                typename q64::template inject_constant_t<-1>,
+                q64::val<i64::val<3>, i64::val<4>>>, 5>,
+            pow_t<pq64, pq64::val<typename q64::template inject_constant_t<-1>, typename q64::one>, 11>
+        >;
+        using FLOAT = double;
+        run<P, FLOAT, false>(0.68, 1.15, "plots/large_sample_horner.dat", &exact_large);
+        run<P, FLOAT, true>(0.68, 1.15, "plots/large_sample_comp_horner.dat", &exact_large);
 
-    ::printf("polynomial = %s\n", P::to_string().c_str());
+        run<P, FLOAT, false>(0.74995, 0.75005, "plots/first_root_horner.dat", &exact_large);
+        run<P, FLOAT, true>(0.74995, 0.75005, "plots/first_root_comp_horner.dat", &exact_large);
 
-    using FLOAT = double;
-    run<P, FLOAT, false>(0.68, 1.15, "plots/large_sample_horner.dat");
-    run<P, FLOAT, true>(0.68, 1.15, "plots/large_sample_comp_horner.dat");
+        run<P, FLOAT, false>(0.9935, 1.0065, "plots/second_root_horner.dat", &exact_large);
+        run<P, FLOAT, true>(0.9935, 1.0065, "plots/second_root_comp_horner.dat", &exact_large);
+    }
+    {
+        // (x - 1) ^ 3
+        using P = make_int_polynomial_t<i32, 1, -3, 3, -1>;
 
-    run<P, FLOAT, false>(0.74995, 0.75005, "plots/first_root_horner.dat");
-    run<P, FLOAT, true>(0.74995, 0.75005, "plots/first_root_comp_horner.dat");
-
-    run<P, FLOAT, false>(0.9935, 1.0065, "plots/second_root_horner.dat");
-    run<P, FLOAT, true>(0.9935, 1.0065, "plots/second_root_comp_horner.dat");
+        run<P, double, false>(1-0.00005, 1+0.00005, "plots/double.dat", &exact_small);
+        run<P, float, true>(1-0.00005, 1+0.00005, "plots/float_comp.dat", &exact_small);
+    }
 }
