### 2024-10-08 3df5c51
finish fp16 example
