### 2024-10-08 09f50bc
add double for CUDA sample
