### 2024-10-08 61834e4
fma trick in compensated eval
