### 2024-10-14 33ecf1e
partially rewrite and export benchmarks on figshare -- JOSS paper updated accordingly
