+++ paper.bib
@@ -114,3 +114,13 @@
   journal={Algebraic and Numerical Algorithms and},
   year={2006}
 }
+
+@misc{aerobus_benchmarks_2023,
+  author       = {Comua},
+  title        = {{Aerobus benchmarks}},
+  year         = {2024},
+  url          = {https://figshare.com/articles/dataset/Aerobus_benchmarks/27222777?file=49774845},
+  note         = {CSV dataset},
+  publisher    = {Figshare},
+  version      = {v1}
+}
