+++ paper.md
@@ -33,7 +33,7 @@ In addition, we expose a way to generate Taylor series of analytic functions, kn
 
 ## Statement of need
 
-By implementing general algebra concepts such as discrete rings, field of fractions and polynomials, `Aerobus` can serve multiple purposes, mainly polynomial arithmetic at compile time and efficient polynomial evaluation, regardless of the coefficients `ring`.
+By implementing general algebra concepts such as discrete rings, field of fractions and polynomials, `Aerobus` can serve multiple purposes, mainly polynomial arithmetic at compile time and efficient [@aerobus_benchmarks_2023] polynomial evaluation, regardless of the coefficients `ring`.
 
 The main application we want to express in this paper is the automatic (and configurable) generation or Taylor approximation of analytic functions such as `exp` or `sin`, by using polynomial arithmetic at compile time.
 
@@ -53,7 +53,7 @@ Some others can provide vectorized functions, such as [@wang2014intel] does. But
 
 `Aerobus` provides automatic generation of such functions, in a hardware-independent way, as tested on x86 and CUDA platforms. In addition, `Aerobus` provides a way to control the precision of the generated function by changing the degree of Taylor expansion, which can't be used in competing libraries without reimplementing the whole function or changing the array of coefficients.
 
-Please note that, `Aerobus` does not provide optimal approximation polynomials the way [@chevillard2010sollya] does. However, `Sollya` can be used beforehand to feed `aerobus` with appropriate coefficients. Similarly, `Aerobus` does not perform (yet) compensated Horner scheme like in [@graillat2006compensated] or floating point manipulations (domain normalization) to extend domain of approximation, like it is done in standard library.
+`Aerobus` does not provide optimal approximation polynomials the way [@chevillard2010sollya] does. However, `Sollya` could be used beforehand to feed `aerobus` with appropriate coefficients. `Aerobus` does not provide floating point manipulations (domain normalization) to extend domain of approximation, like it is done in standard library.
 
 ## Mathematic definitions
 
@@ -143,34 +143,15 @@ $$P(x) = a_0 + X (a_1 + X (a_2 + X(... + X(a_{n-1} + X an))))$$
 
 which is automaticcaly done by `aerobus` using the `polynomial::val::eval` function.
 
-The library also provides built-in integers and functions, such as:
+This evaluation function is `constexpr` and therefore will be completely computed at compile time when called on a constant.
 
-- `is_prime` : checks if an integer is prime;
-- `factorial_t` : factorial of an integer;
-- `pow_t` : power function ($a^n$);
-- `alternate_t` : $(-1)^p$;
-- `combination_t` : combination $\binom{n}{k}$;
-- `bernouilli_t` : Bernoulli numbers (used for some Taylor series).
+Polynomials also expose Compensated Horner scheme like in [@graillat2006compensated], to gain extra precision when evaluating polynomials close to its roots.
 
-And Taylor series for these functions:
+The library also provides built-in integers and functions, such as Bernouilli numbers, factorials or other utilities.
 
-- `exp`
-- `expm1` $e^x - 1$
-- `lnp1` $\ln(x+1)$
-- `geom` $\frac{1}{1-x}$
-- `sin`
-- `cos`
-- `tan`
-- `sh`
-- `cosh`
-- `tanh`
-- `asin`
-- `acos`
-- `acosh`
-- `asinh`
-- `atanh`
+Some well known Taylor series, such as `exp` or `acosh` come preimplemented.
 
-Additionally, the library comes with a type designed to help the users implement other Taylor series.
+The library comes with a type designed to help the users implement other Taylor series.
[…]
