+++ src/benchmarks.cpp
@@ -2,8 +2,6 @@
 #include <benchmark/benchmark.h>  // NOLINT
 #include <cmath>  // NOLINT
 
-static constexpr size_t N = 1 << 23;
-
 double rand(double min, double max) {
   double range = (max - min);
   double div = RAND_MAX / range;
@@ -17,104 +15,122 @@ INLINED double aero_expm1_12(const double x) {
 }
 
 static void BM_aero_expm1_12(benchmark::State &state) {
-    double *in = aerobus::aligned_malloc<double>(N, 64);
-    double *out = aerobus::aligned_malloc<double>(N, 64);
+    double *in = aerobus::aligned_malloc<double>(state.range(0), 64);
+    double *out = aerobus::aligned_malloc<double>(state.range(0), 64);
     #pragma omp parallel for
-    for (size_t i = 0; i < N; ++i) {
+    for (int64_t i = 0; i < state.range(0); ++i) {
         in[i] = rand(-0.01, 0.01);
     }
     for (auto _ : state) {
         #pragma omp parallel for
-        for (size_t i = 0; i < N; ++i) {
+        for (int64_t i = 0; i < state.range(0); ++i) {
             out[i] = aero_expm1_12(in[i]);
         }
     }
+
+    free(in);
+    free(out);
 }
 
 static void BM_std_expm1_12(benchmark::State &state) {
-    double *in = aerobus::aligned_malloc<double>(N, 64);
-    double *out = aerobus::aligned_malloc<double>(N, 64);
+    double *in = aerobus::aligned_malloc<double>(state.range(0), 64);
+    double *out = aerobus::aligned_malloc<double>(state.range(0), 64);
     #pragma omp parallel for
-    for (size_t i = 0; i < N; ++i) {
+    for (int64_t i = 0; i < state.range(0); ++i) {
         in[i] = rand(-0.01, 0.01);
     }
     for (auto _ : state) {
         #pragma omp parallel for
-        for (size_t i = 0; i < N; ++i) {
+        for (int64_t i = 0; i < state.range(0); ++i) {
             out[i] = ::expm1(::expm1(::expm1(::expm1(::expm1(::expm1(
                 ::expm1(::expm1(::expm1(::expm1(::expm1(::expm1(in[i]))))))))))));
         }
     }
+
+    free(in);
+    free(out);
 }
 
 static void BM_aero_hermite(benchmark::State &state) {
-    double *in = aerobus::aligned_malloc<double>(N, 64);
-    double *out = aerobus::aligned_malloc<double>(N, 64);
+    double *in = aerobus::aligned_malloc<double>(state.range(0), 64);
+    double *out = aerobus::aligned_malloc<double>(state.range(0), 64);
     #pragma omp parallel for
-    for (size_t i = 0; i < N; ++i) {
+    for (int64_t i = 0; i < state.range(0); ++i) {
         in[i] = rand(-0.01, 0.01);
     }
     for (auto _ : state) {
         #pragma omp parallel for
-        for (size_t i = 0; i < N; ++i) {
+        for (int64_t i = 0; i < state.range(0); ++i) {
             out[i] = aerobus::known_polynomials::hermite_phys<12>::eval<double>(in[i]);
         }
     }
+
+    free(in);
+    free(out);
 }
 
 static void BM_std_hermite(benchmark::State &state) {
-    double *in = aerobus::aligned_malloc<double>(N, 64);
-    double *out = aerobus::aligned_malloc<double>(N, 64);
+    double *in = aerobus::aligned_malloc<double>(state.range(0), 64);
[…]
