### 2024-10-14 #6233 @JeWaVe
I did what you suggested : remove benchmarks from paper and replace them by a reference to a figshare csv file
