### 2024-10-15 d61da01
fix cuda_fp16 (by modifying CUDA headers...) and improve sollya script
