### 2024-10-15 8cbaaa0
make compensated horner work with CUDA fp16
