### 2024-10-15 7ab653b
update README for CUDA
