### 2024-10-16 63dfbb6
cuda configuration in fp16.cu
