### 2024-10-18 a18bdc2
implementation of libm::sin and libm::cos, correct at epsilon + add benchmark
