+++ src/benchmarks.cpp
@@ -33,8 +33,8 @@ static void BM_aero_expm1_12(benchmark::State &state) {
 }
 
 static void BM_std_expm1_12(benchmark::State &state) {
-    double *in = aerobus::aligned_malloc<double>(state.range(0), 64);
-    double *out = aerobus::aligned_malloc<double>(state.range(0), 64);
+    float *in = aerobus::aligned_malloc<float>(state.range(0), 64);
+    float *out = aerobus::aligned_malloc<float>(state.range(0), 64);
     #pragma omp parallel for
     for (int64_t i = 0; i < state.range(0); ++i) {
         in[i] = rand(-0.01, 0.01);
@@ -44,6 +44,45 @@ static void BM_std_expm1_12(benchmark::State &state) {
         for (int64_t i = 0; i < state.range(0); ++i) {
             out[i] = ::expm1(::expm1(::expm1(::expm1(::expm1(::expm1(
                 ::expm1(::expm1(::expm1(::expm1(::expm1(::expm1(in[i]))))))))))));
+        }
+    }
+
+    free(in);
+    free(out);
+}
+
+
+static void BM_aero_sin_12(benchmark::State &state) {
+    float *in = aerobus::aligned_malloc<float>(state.range(0), 64);
+    float *out = aerobus::aligned_malloc<float>(state.range(0), 64);
+    #pragma omp parallel for
+    for (int64_t i = 0; i < state.range(0); ++i) {
+        in[i] = rand(-0.01, 0.01);
+    }
+    for (auto _ : state) {
+        #pragma omp parallel for
+        for (int64_t i = 0; i < state.range(0); ++i) {
+            out[i] = aerobus::libm::sin(aerobus::libm::sin(aerobus::libm::sin(
+                aerobus::libm::sin(aerobus::libm::sin(aerobus::libm::sin(
+                    aerobus::libm::sin(aerobus::libm::sin(aerobus::libm::sin(in[i])))))))));
+        }
+    }
+
+    free(in);
+    free(out);
+}
+
+static void BM_std_sin_12(benchmark::State &state) {
+    double *in = aerobus::aligned_malloc<double>(state.range(0), 64);
+    double *out = aerobus::aligned_malloc<double>(state.range(0), 64);
+    #pragma omp parallel for
+    for (int64_t i = 0; i < state.range(0); ++i) {
+        in[i] = rand(-0.01, 0.01);
+    }
+    for (auto _ : state) {
+        #pragma omp parallel for
+        for (int64_t i = 0; i < state.range(0); ++i) {
+            out[i] = ::sin(::sin(::sin(::sin(::sin(::sin(::sin(::sin(::sin(::sin(::sin(::sin(in[i]))))))))))));
         }
     }
 
@@ -126,6 +165,8 @@ static void BM_horner_double(benchmark::State &state) {
     free(out);
 }
 
+BENCHMARK(BM_std_sin_12)->Range(1 << 10, 1 << 24);
+BENCHMARK(BM_aero_sin_12)->Range(1 << 10, 1 << 24);
 BENCHMARK(BM_std_expm1_12)->Range(1 << 10, 1 << 24);
 BENCHMARK(BM_aero_expm1_12)->Range(1 << 10, 1 << 24);
 BENCHMARK(BM_std_hermite)->Range(1 << 10, 1 << 24);
