### 2024-10-18 4416a12
implementation of libm::sin and libm::cos, correct at epsilon + add benchmark
