+++ src/aerobus.h
@@ -12,6 +12,8 @@
 #include <string>
 #include <concepts> // NOLINT
 #include <array>
+#include <cmath>
+#include <limits>
 #ifdef WITH_CUDA_FP16
 #include <bit>
 #include <cuda_fp16.h>
@@ -283,6 +285,109 @@ namespace aerobus {
     }  // namespace internal
 }  // namespace aerobus
 
+namespace aerobus {
+    namespace internal {
+        struct double_double {};
+
+        template<typename T>
+        struct arithmetic_helpers;
+
+        template<>
+        struct arithmetic_helpers<double> {
+            using integers = int64_t;
+            using upper_type = double_double;
+            static constexpr double one = 1.0;
+            static constexpr double zero = 0.0;
+            static constexpr double pi = 0x1.921fb54442d18p1;
+            static constexpr double pi_2 = 0x1.921fb54442d18p0;
+            static constexpr double pi_4 = 0x1.921fb54442d18p-1;
+            static constexpr double two_pi = 0x1.921fb54442d18p2;
+            static constexpr double inv_two_pi = 0x1.45f306dc9c883p-3;
+            static constexpr double half = 0x1p-1;
+        };
+
+        template<>
+        struct arithmetic_helpers<float> {
+            using integers = int32_t;
+            using upper_type = double;
+            static constexpr double one = 1.0F;
+            static constexpr double zero = 0.0F;
+            static constexpr float pi = 0x1.921fb6p1f;
+            static constexpr double pi_2 = 0x1.921fb6p0f;
+            static constexpr double pi_4 = 0x1.921fb6p-1f;
+            static constexpr double two_pi = 0x1.921fb6p2f;
+            static constexpr double inv_two_pi = 0x1.45f306p-3f;
+            static constexpr double half = 0x1p-1f;
+        };
+
+        #ifdef WITH_CUDA_FP16
+        template<>
+        struct arithmetic_helpers<__half> {
+            using integers = int16_t;
+            using upper_type = float;
+            const __half one = CUDART_ONE_FP16;
+            const __half zero = CUDART_ZERO_FP16;
+            const __half pi = __half_raw(0x4248);
+            const __half pi_2 =   __half_raw(0b011111001001000);
+            const __half pi_4 =   __half_raw(0b011101001001000);
+            const __half two_pi = __half_raw(0b100011001001000);
+            const __half inv_two_pi = __half_raw(0b011000100011000);
+            const __half half = __half_raw(0b011100000000000);
+        };
+
+        template<>
+        struct arithmetic_helpers<__half2> {
+            using integers = int16_t;
+            using upper_type = float;  // TODO(JeWaVe) : check for float2
+            const __half2 one = __half2(CUDART_ONE_FP16, CUDART_ONE_FP16);
+            const __half2 zero = __half2(CUDART_ZERO_FP16, CUDART_ZERO_FP16);
+            const __half2 pi = __half2(__half_raw(0x4248), __half_raw(0x4248));
+            const __half2 pi_2 = __half2(__half_raw(0b011111001001000), __half_raw(0b011111001001000));
+            const __half2 pi_4 = __half2(__half_raw(0b011101001001000), __half_raw(0b011101001001000));
+            const __half2 two_pi = __half2(__half_raw(0b100011001001000), __half_raw(0b100011001001000));
+            const __half2 half = __half2(__half_raw(0b011100000000000), __half_raw(0b011100000000000));
+        };
+        #endif
+    }  // namespace internal
+
+    template<typename T>
+    struct meta_libm {
+        static INLINED DEVICE T floor(const T& f) {
+            return std::floor(f);
+        }
+        static INLINED DEVICE T fmod(const T& x, const T& d) {
[…]
