+++ src/aerobus.h
@@ -44,6 +44,8 @@
 
 //! \namespace aerobus::libm holds mathematical functions (such as cosine or sin), correct to epsilon
 
+//! \namespace aerobus::libm::internal internal implementation of libm, containins fpminimax polynomials given by Sollya
+
 // aligned allocation
 namespace aerobus {
     /**
@@ -285,6 +287,7 @@ namespace aerobus {
     }  // namespace internal
 }  // namespace aerobus
 
+// arithmetic helpers
 namespace aerobus {
     namespace internal {
         struct double_double {};
@@ -503,7 +506,7 @@ namespace aerobus {
     }  // namespace internal
 }  // namespace aerobus
 
-// utilities
+// type utilities
 namespace aerobus {
     namespace internal {
         template<template<typename...> typename TT, typename T>
@@ -748,6 +751,7 @@ namespace aerobus {
     struct Embed;
 }  // namespace aerobus
 
+// quotient rings
 namespace aerobus {
     /// @brief Quotient ring by the principal ideal generated by 'X'
     /// With i32 as Ring and i32::val<2> as X, Quotient is Z/2Z
@@ -4293,6 +4297,7 @@ namespace aerobus {
     }  // namespace known_polynomials
 }  // namespace aerobus
 
+// libm
 namespace aerobus {
     namespace libm {
         namespace internal {
@@ -4463,10 +4468,6 @@ namespace aerobus {
             #endif
         }  // namespace internal
 
-        /// @brief computes 2^x (NOT TESTED YET)
-        /// @tparam T arithmetic type (float only so far)
-        /// @param x input
-        /// @return 2^x
         template<typename T>
         static T exp2(const T&x) {
             using poly = internal::exp2_poly<T>::type;
@@ -4491,13 +4492,6 @@ namespace aerobus {
         template<typename T>
         static T cos(const T& x);
 
-        /// @brief computes sine function, following IEEE-754 recommendations
-        ///
-        /// exact values for +-0 and infinities, correct at epsilon for other values
-        ///
-        /// @tparam T arithmetic_type (float only so far)
-        /// @param x input value
-        /// @return sin(x)
         template<typename T>
         static T sin(const T& x) {
             using upper_type = aerobus::internal::arithmetic_helpers<T>::upper_type;
@@ -4530,14 +4524,6 @@ namespace aerobus {
             }
         }
 
-
-        /// @brief computes cosine function, following IEEE-754 recommendations
-        ///
-        /// exact values for +-0 and infinities, correct at epsilon for other values
-        ///
-        /// @tparam T arithmetic_type (float only so far)
-        /// @param x input value
-        /// @return cos(x)
         template<typename T>
         static T cos(const T& x) {
             using poly = internal::cos_poly<T>::type;
