+++ src/benchmarks.cpp
@@ -72,6 +72,27 @@ static void BM_aero_sin_12(benchmark::State &state) {
     free(out);
 }
 
+
+static void BM_aero_fast_sin_12(benchmark::State &state) {
+    float *in = aerobus::aligned_malloc<float>(state.range(0), 64);
+    float *out = aerobus::aligned_malloc<float>(state.range(0), 64);
+    #pragma omp parallel for
+    for (int64_t i = 0; i < state.range(0); ++i) {
+        in[i] = rand(-0.01, 0.01);
+    }
+    for (auto _ : state) {
+        #pragma omp parallel for
+        for (int64_t i = 0; i < state.range(0); ++i) {
+            out[i] = aerobus::libm::fast_sin(aerobus::libm::fast_sin(aerobus::libm::fast_sin(
+                aerobus::libm::fast_sin(aerobus::libm::fast_sin(aerobus::libm::fast_sin(
+                    aerobus::libm::fast_sin(aerobus::libm::fast_sin(aerobus::libm::fast_sin(in[i])))))))));
+        }
+    }
+
+    free(in);
+    free(out);
+}
+
 static void BM_std_sin_12(benchmark::State &state) {
     double *in = aerobus::aligned_malloc<double>(state.range(0), 64);
     double *out = aerobus::aligned_malloc<double>(state.range(0), 64);
@@ -208,12 +229,17 @@ static void BM_horner_double(benchmark::State &state) {
 
 BENCHMARK(BM_std_cos_12)->Range(1 << 10, 1 << 24);
 BENCHMARK(BM_aero_cos_12)->Range(1 << 10, 1 << 24);
+
 BENCHMARK(BM_std_sin_12)->Range(1 << 10, 1 << 24);
 BENCHMARK(BM_aero_sin_12)->Range(1 << 10, 1 << 24);
+BENCHMARK(BM_aero_fast_sin_12)->Range(1 << 10, 1 << 24);
+
 BENCHMARK(BM_std_expm1_12)->Range(1 << 10, 1 << 24);
 BENCHMARK(BM_aero_expm1_12)->Range(1 << 10, 1 << 24);
+
 BENCHMARK(BM_std_hermite)->Range(1 << 10, 1 << 24);
 BENCHMARK(BM_aero_hermite)->Range(1 << 10, 1 << 24);
+
 BENCHMARK(BM_horner_double)->Range(1 << 10, 1 << 24);
 BENCHMARK(BM_compensated_horner_float)->Range(1 << 10, 1 << 24);
 
