### 2024-11-02 #6233 @JeWaVe
Hi,

here is the archive for the release : 10.6084/m9.figshare.27517938

License is MIT, same as for the code

Latest release is v1.4 (github and figshare)
