### 2024-11-05 83b7f81
libm::sin runs in cuda fp16
