+++ examples/utilities.h
@@ -1,0 +1,45 @@
+
+double rand(double min, double max) {
+  double range = (max - min);
+  double div = RAND_MAX / range;
+  return min + (rand() / div);  // NOLINT
+}
+
+template<typename T>
+struct GetRandT;
+
+template<>
+struct GetRandT<double> {
+    static double func(double min, double max) {
+        return rand(min, max);
+    }
+};
+
+template<>
+struct GetRandT<float> {
+    static float func(double min, double max) {
+        return (float) rand(min, max);  // NOLINT
+    }
+};
+
+template<>
+struct GetRandT<__half2> {
+    static __half2 func(double min, double max) {
+        return __half2(__float2half((float)rand(min, max)), __float2half((float)rand(min, max)));  // NOLINT
+    }
+};
+
+template<>
+struct GetRandT<__half> {
+    static __half func(double min, double max) {
+        return __float2half((float)rand(min, max));  // NOLINT
+    }
+};
+
+#define cudaErrorCheck(ans) { gpuAssert((ans), __FILE__, __LINE__); }
+inline void gpuAssert(cudaError_t code, const char *file, int line, bool abort = true) {
+    if (code != cudaSuccess) {
+        fprintf(stderr, "GPUassert: %s %s %d\n", cudaGetErrorString(code), file, line);
+        if (abort) exit(code);
+    }
+}
