### 2024-11-05 b57d080
remove recursion in sinus range reduction -- should be faster on GPU
