+++ src/benchmarks.cpp
@@ -113,12 +113,12 @@ static void BM_std_sin_12(benchmark::State &state) {
 
 
 static void BM_aero_cos_12(benchmark::State &state) {
-    using constants = aerobus::internal::arithmetic_helpers<float>;
+    using constants = aerobus::arithmetic_helpers<float>;
     float *in = aerobus::aligned_malloc<float>(state.range(0), 64);
     float *out = aerobus::aligned_malloc<float>(state.range(0), 64);
     #pragma omp parallel for
     for (int64_t i = 0; i < state.range(0); ++i) {
-        in[i] = rand(constants::pi_2 - 0.01, constants::pi_2 + 0.01);
+        in[i] = rand(constants::pi_2() - 0.01, constants::pi_2() + 0.01);
     }
     for (auto _ : state) {
         #pragma omp parallel for
@@ -134,12 +134,12 @@ static void BM_aero_cos_12(benchmark::State &state) {
 }
 
 static void BM_std_cos_12(benchmark::State &state) {
-    using constants = aerobus::internal::arithmetic_helpers<float>;
+    using constants = aerobus::arithmetic_helpers<float>;
     double *in = aerobus::aligned_malloc<double>(state.range(0), 64);
     double *out = aerobus::aligned_malloc<double>(state.range(0), 64);
     #pragma omp parallel for
     for (int64_t i = 0; i < state.range(0); ++i) {
-        in[i] = rand(constants::pi_2 - 0.01, constants::pi_2 + 0.01);
+        in[i] = rand(constants::pi_2() - 0.01, constants::pi_2() + 0.01);
     }
     for (auto _ : state) {
         #pragma omp parallel for
