### 2024-11-28 #6233 @JeWaVe
Sure.

Algebra : https://doi.org/10.1007/978-1-4613-0041-0 (https://link.springer.com/book/10.1007/978-1-4613-0041-0)

Sollya : 0.1007/978-3-642-15582-6_5￿. (https://inria.hal.science/hal-00761644/document)
