### 2024-12-10 3636e44
amdgpu example
