### 2025-01-06 03673e1
Update paper.bib
add missing DOIs
