+++ paper.bib
@@ -38,14 +38,16 @@
   author={Lang, Serge},
   volume={211},
   year={2012},
-  publisher={Springer Science \& Business Media}
+  publisher={Springer Science \& Business Media},
+  doi="10.1007/978-1-4613-0041-0",
 }
 
 @book{bourbaki2013algebra,
   title={Algebra II: Chapters 4-7},
   author={Bourbaki, Nicolas},
   year={2013},
-  publisher={Springer Science \& Business Media}
+  publisher={Springer Science \& Business Media},
+  doi="10.1007/978-3-642-61698-3"
 }
 
 @article{evariste1846memoire,
@@ -62,7 +64,8 @@
   title={Handbook of computational group theory},
   author={Holt, Derek F and Eick, Bettina and O'Brien, Eamonn A},
   year={2005},
-  publisher={CRC Press}
+  publisher={CRC Press},
+  doi="10.1201/9781420035216"
 }
 
 @book{heath1956thirteen,
@@ -86,7 +89,8 @@
   title={Art of computer programming, volume 2: Seminumerical algorithms},
   author={Knuth, Donald E},
   year={2014},
-  publisher={Addison-Wesley Professional}
+  publisher={Addison-Wesley Professional},
+  doi="10.1137/1012065"
 }
 
 @article{karney2013algorithms,
@@ -96,8 +100,10 @@
   volume={87},
   pages={43--55},
   year={2013},
-  publisher={Springer}
+  publisher={Springer},
+  doi="10.1007/s00190-012-0578-z"
 }
+
 @INPROCEEDINGS{ChevillardJoldesLauter2010,
   author = {S.~Chevillard and M.~Jolde\c{s} and C.~Lauter},
   title = {Sollya: An Environment for the Development of Numerical Codes},
@@ -116,7 +122,8 @@
               floating-point software and hardware design. Its features
               include on-the-fly faithful rounding, specialized approximation algorithms 
               and extensive support for floating-point code generation.},
-  keywords = {numerical software, faithful rounding, computer algebra, development tool, function approximation}
+  keywords = {numerical software, faithful rounding, computer algebra, development tool, function approximation},
+  doi="10.1007/978-3-642-15582-6_5"
 }
 
 @article{graillat2006compensated,
