### 2025-01-06 #6233 @JeWaVe
I fixed the DOIs in the bibliography

Latest release : 1.5
DOI : https://doi.org/10.6084/m9.figshare.28143233.v1  (MIT)
