### 2025-01-07 8300c2f
Update paper.bib
fix sollya reference (copied from springer)
