+++ paper.bib
@@ -105,24 +105,16 @@
 }
 
 @INPROCEEDINGS{ChevillardJoldesLauter2010,
-  author = {S.~Chevillard and M.~Jolde\c{s} and C.~Lauter},
-  title = {Sollya: An Environment for the Development of Numerical Codes},
-  booktitle = {Mathematical Software - ICMS 2010},
-  volume = {6327},
-  series = {Lecture Notes in Computer Science},
-  year = {2010},
-  editor = {K.~Fukuda and J.~van~der~Hoeven and M.~Joswig and N.~Takayama},
-  pages = {28--31},
-  address = {Heidelberg, Germany},
-  month = {September},
-  publisher = {Springer},
-  abstract = {Sollya has become a mature tool for the development of
-              numerical software. With about $175$ built-in algorithms and a broad
-              extensibility, it offers a complete tool-chain for fixed- and
-              floating-point software and hardware design. Its features
-              include on-the-fly faithful rounding, specialized approximation algorithms 
-              and extensive support for floating-point code generation.},
-  keywords = {numerical software, faithful rounding, computer algebra, development tool, function approximation},
+  author="Chevillard, Sylvain and Jolde{\c{s}}, Mioara and Lauter, Christoph",
+  editor="Fukuda, Komei and Hoeven, Joris van der and Joswig, Michael and Takayama, Nobuki",
+  title="Sollya: An Environment for the Development of Numerical Codes",
+  booktitle="Mathematical Software -- ICMS 2010",
+  year="2010",
+  publisher="Springer Berlin Heidelberg",
+  address="Berlin, Heidelberg",
+  pages="28--31",
+  abstract="Sollya has become a mature tool for the development of numerical software. With about 175 built-in algorithms and a broad extensibility, it offers a complete tool-chain for fixed- and floating-point software and hardware design. Its features include on-the-fly faithful rounding, specialized approximation algorithms and extensive support for floating-point code generation.",
+  isbn="978-3-642-15582-6",
   doi="10.1007/978-3-642-15582-6_5"
 }
 
