### 2025-01-13 #6233 @JeWaVe
I understand, It's the reference in the paper not in the bibliography.

Fixed in version 1.5.2 : https://doi.org/10.6084/m9.figshare.28193474.v1
