### 2025-01-16 1054cc2
fix CI by removing wrongly commited UT for libm::sin
