### 2025-03-10 #6233 @gkthiruvathukal
@pitsianis We are on the same page. I was definitely not saying that you have to check any boxes that you do not agree are completed. Furthermore, you have clarified that there are remaining blockers for acceptance.  I apologize for not having seen your comments. There is some very weird user interface behavior with the issue thread collapsing in the middle and it also caused me not to see one of the checklists.

@JeWaVe Please let me know when you have addressed these comments or continue to engage with the reviewer(s) until all input has been addressed satisfactorily. We cannot proceed without considering all input, whether it be via the review checklist, within the comment thread, or reported issues using the issue tracker.

So please don't assume we are proceeding without you, @pitsianis. Thank you for helping us!!
