### 2025-03-11 #6233 @JeWaVe
> [@gkthiruvathukal](https://github.com/gkthiruvathukal) the issue is not about checking the remaining boxes but whether I agree to check them. 😅
>
> [I expressed my opinion on some issues that have not been resolved (adequately or at all). ](https://github.com/openjournals/joss-reviews/issues/6233#issuecomment-2386417045)
>
> The strongest of all is the contradiction of _discrete Euclidean domains_ in the title and the main application to be
>
> > the automatic (and configurable) generation or Taylor approximation of analytic functions such as `exp` or `sin`, by using polynomial arithmetic at compile time.
>
> Discrete Euclidean domains have nothing to do with differentiation and Taylor series approximation.
>
> See my comments above. If you disagree with me or find that the issues I raised have been addressed, feel free to ignore me and proceed without me. I have no bad feelings.

Hi

I understand your comment. In fact it has to do (even if mathematically speaking it does not), because being able to manipulate discrete euclidean domains allow me to generate those taylor series (as long as coefficients are rationals).
Should I rephrase to state that better ?
