### 2025-03-11 #6233 @pitsianis
> I understand your comment. In fact it has to do (even if mathematically speaking it does not), because being able to manipulate discrete euclidean domains allow me to generate those taylor series (as long as coefficients are rationals).
> Should I rephrase to state that better ?

Yes, please provide a paragraph—-here initially—and try to answer [each of my comments](https://github.com/openjournals/joss-reviews/issues/6233#issuecomment-2386417045), especially this seeming contradiction regarding differentiation and discrete Euclidean domains. Perhaps the library's inner workings are the "glue" that connects the Galois Field multiplication tables with the continued fractions, the family of orthogonal polynomial bases, and the approximations of transcendental functions. So far it is not clear to me.

Every time I re-read your paper, I get more questions! What do you mean when you say that

> Laguerre, Hermite, Bernstein, Bessel polynomials have integers coefficients by default, but can be defined (specialized) with coefficients in any Ring (such as ℤ/2ℤ).

What do you get evaluating the Laguerre polynomial:

$L_n(x) = \frac{e^x}{n!} \frac{d^n}{dx^n} (x^n e^{-x})$

with values from a ℤ/2ℤ ring? Are there applications that somehow get meaningful answers from such evaluations? I am going on a limb and throwing in the air perhaps combinatorics, coding theory, error-correcting codes, or some super-duper quantum states.  If so, please cite them; if not, is there the possibility of such a use?
