### 2025-03-12 #6233 @JeWaVe
Thanks @pitsianis for your valuable comments.

I'll try my best to answer them

> By implementing general algebra concepts such as discrete rings, field of fractions and
polynomials, Aerobus can serve multiple purposes, mainly polynomial arithmetic at compile
time and efficient (Comua, 2024) polynomial evaluation, regardless of the coefficients ring.
> The main application we want to express in this paper is the automatic (and configurable)
generation or Taylor approximation of analytic functions such as exp or sin, by using polynomial
arithmetic at compile time.

I propose to replace by

> The core of the library is an implementation of general algebra concepts such as discrete rings, field of fractions, quotient rings, finite fields and polynomials. All these concepts are as abstracted as possible in the C++ type arithmetic, making possible to manipulate "high-level" concepts such as polynomials or continued fractions without knowing the actual implementation of the underlying ring of coefficients.
> Cryptography or combinatorics make extensive use of polynomials over finite fields (Mullen 2013). On the other side, Taylor series of most of the usual transcendental functions (used in numerical simulations) are polynomials with rationals or integers coefficients. Aerobus can operate in the two worlds at compile time and evaluate efficiently at runtime (Comua, 2024), precisely because it has this agnostic representation of the ring of coefficients.

```@book{mullen2013handbook,
  title={Handbook of finite fields},
  author={Mullen, Gary L and Panario, Daniel},
  volume={17},
  year={2013},
  publisher={CRC press Boca Raton}
}
```

> Back to the first line. What does it mean to "implement the concept of discrete ring"? Given by the user an addition and a multiplication operator, does it verify the ring or the field axioms like associativity, distribution of the multiplication over addition, existence of zero (and one) identity element(s), and inverses? What would happen if the user attempted to use operators that do not satisfy the corresponding axioms, for instance, using the semiring operators min(int32,int32) and plus(int32,int32) as the addition and multiplication?

I stated explicitly elsewhere (answering your comment) that rings must be associative and commutative (I thought it was implicit, since integer arithmetic is exact in IEEE).

> Some of the claims might also be incorrect. Why precomputing shrinks binary size? For example if I implement Horner's rule as a loop, and compare it against the code with the same loop unrolled, which code will have the smaller binary for a polynomial of degree 20?

This is indeed wrong and removed from the latest version of the paper.

> Why is it advantageous to precompute at compile-time the addition and multiplication tables of a Galois field using macros instead of "normal" run-time codes? Why is it advantageous to have a macro to form continued fractions instead of a function at run time?

Well I see multiple ways to answer that good question. One would be that using compile time representation of polynomials, arithmetic over them is exact, as is arithmetic over integers in IEEE. Therefore, adding (1 + x^2) and (1/10^n + x^2) will result in (exactly) ((10^n+1)/10^n + 2x^2) instead of 1 + 2x^2 if n is large enough when using floating point coefficients. Similarly, (1+x^20) / (1+x^20) will lead to one at compile time (and therefore a constant evaluation function) with aerobus while a runtime library can't skip the long division algorithm before concluding the same way.
For the continued fractions the thing is to have compile time constants at the desired precision. Pi (or any other) in float16 is not the same as in float64 and this can be abstracted away by aerobus.

Hoping my answers are satisfactory,

Regis Portalez
