### 2025-03-31 #6233 @pitsianis
I will try to follow up by Friday.. In the meantime, please review past messages for questions or observations that have not been answered.

> On Mar 31, 2025, at 3:42 PM, Portalez Régis ***@***.***> wrote:
>
>
> JeWaVe
>  left a comment
> (openjournals/joss-reviews#6233)
> ping
>
