### 2025-04-05 #6233 @pitsianis
1. I believe we need a better title that reflects the contents of the library. To me, it is a contradiction to only state *discrete Euclidean domains* in the title and the main application to be
> the automatic (and configurable) generation or Taylor approximation of analytic functions such as exp or sin, by using polynomial arithmetic at compile time.

Discrete Euclidean domains have nothing to do with differentiation and Taylor series approximation. Your answer:

> I understand your comment. In fact it has to do (even if mathematically speaking it does not), because being able to manipulate discrete euclidean domains allow me to generate those taylor series (as long as coefficients are rationals).

confused me even more. Please try again. This is the heart of the issue.

2. What do you mean when you say that

> Laguerre, Hermite, Bernstein, Bessel polynomials have integers coefficients by default, but can be defined (specialized) with coefficients in any Ring (such as ℤ/2ℤ).

What do you get by evaluating the Laguerre polynomial:

$L_n(x) = \frac{e^x}{n!} \frac{d^n}{dx^n} (x^n e^{-x})$
with values from a ℤ/2ℤ ring?

3. What does it mean for a library to "implement the concept of discrete ring"? What exactly does this provide to the user?

4. Given by the user an addition and a multiplication operator, does the implementation verify the ring or the field axioms like associativity, distribution of the multiplication over addition, existence of zero (and one) identity element(s), and inverses?
What do you mean by "If some type meets the `IsRing' concept requirement"?

5. What would happen if a user attempted to use operators that do not satisfy the corresponding ring or field axioms, for instance, using the semiring operators `min(int32,int32)` and `plus(int32,int32)` as the addition and multiplication?

6. I do not understand your answer to my question

>> Why is it advantageous to precompute at compile-time the addition and multiplication tables of a Galois field using macros instead of "normal" run-time codes? Why is it advantageous to have a macro to form continued fractions instead of a function at run time?
>
> Well I see multiple ways to answer that good question. One would be that using compile time representation of polynomials, arithmetic over them is exact, as is arithmetic over integers in IEEE. Therefore, adding (1 + x^2) and (1/10^n + x^2) will result in (exactly) ((10^n+1)/10^n + 2x^2) instead of 1 + 2x^2 if n is large enough when using floating point coefficients. Similarly, (1+x^20) / (1+x^20) will lead to one at compile time (and therefore a constant evaluation function) with aerobus while a runtime library can't skip the long division algorithm before concluding the same way. For the continued fractions the thing is to have compile time constants at the desired precision. Pi (or any other) in float16 is not the same as in float64 and this can be abstracted away by aerobus.

  a. I do not understand your first example.

  b. In your second example, you may be overstating the capabilities of the library. Does Aerobus include symbolic algebra simplification capabilities? I do understand that an operation may be computed at compile time and *may* result in a simplification, but only in trivial cases.

  c. Isn't possible that the opposite may happen as well, i.e. the resulting generated unrolled code may be too large and expensive?

Also you did not answer the original questions:

7. Why is it advantageous to precompute at compile-time the addition and multiplication tables of a Galois field using macros instead of "normal" run-time codes?

8. Why is it advantageous to have a macro to form continued fractions at compile-time instead of a function at run time?

9. Typos and grammar errors! ***PLEASE USE an EDITORIAL SERVICE*** they are now free, like Grammarly, or ChatGPT. Do not use your reviewers for this. Here are just a few
 - Ln 1: polynomial
 - Ln 80: "division ***or*** greatest"?
 - Ln 73,74,187: Euclidean is not always capitalized
 - Ln 125: "development"? I do not understand the use of this word here...
 - Ln 135: Bernouilli or Bernoulli ?
 - Ln 174: Reference***s***
 - Ln 152: "existe"
 - Ln 153: Aerobus provide***s***
 - Ln 154: Rephrase, do not use ellipses (...)
