### 2025-04-28 #6233 @JeWaVe
Hello.

With all due respect, I'm going to discontinue this publication.
It's been 16 months now and I'm fed up. Believe me, I'm sorry, but I'm going to publish on ArXiv and give up publishing with reviewers, because it's such a pain every time.

I know my ramanujan and I understand the importance of the publication process: **checking the content and its intelligibility** (and nothing more).

Unfortunately it seems that this process has gone completely astray and consists in satisfying the ego and nervous whims of certain people eager to show that they are in a position of power, and obviously smarter than the author.

I understand why most of my friends have left academia, and I'm delighted to have done the same. This environment is toxic and intensely depressive.

Your own journal states, “We built this journal because we believe that after you've done the hard work of writing great software, it shouldn't take weeks and months to write a paper about your work.”

Apparently this is not the case. While some reviewers have been quite pleasant, producing useful and sensible comments, others have been particularly noxious.

I warmly thank the former and congratulate the latter on their genius and brilliant careers (no doubt built on the burnouts of their PhD students).

Best regards

Regis Portalez
