### 2025-04-28 #6233 @gkthiruvathukal
Dear @JeWaVe, we are sorry that we could not get to a resolution in this case. As an author, you always have the ability to withdraw your work from further consideration. We also hope that this process was helpful to you as you work on your final ArXiv submission.

I also wish to take a moment to thank the reviewers @lucaferranti and @pitsianis for their efforts here.
